Home

LightQuery.jl

LightQuery.Nameless

LightQuery.based_on

LightQuery.columns

LightQuery.flatten_columns

LightQuery.group_by

LightQuery.inner_join

LightQuery.order_by

LightQuery.pretty

LightQuery.rows

LightQuery.select

LightQuery.transform

LightQuery.where

LightQuery.@>

LightQuery.@_

LightQuery.Nameless — Type.

A container for a function and the expression that generated it

source

LightQuery.based_on — Method.

based_on(data; kwargs...)

julia> using LightQuery

julia> based_on((a = 1, b = 2), c = @_ _.a + _.b)
(c = 3,)

source

LightQuery.columns — Function.

columns(data)

julia> using LightQuery

julia> columns([(a = 1, b = 2), (a = 2, b = 1)])
(a = [1, 2], b = [2, 1])

source

LightQuery.flatten_columns — Method.

flatten_columns(iterator)

Flatten the columns of an iterator of named tuples.

julia> using LightQuery

julia> flatten_columns([(a = [1, 2], b = [4, 3]), (a = [3, 4], b = [2, 1])])
(a = [1, 2, 3, 4], b = [4, 3, 2, 1])

source

LightQuery.group_by — Method.

group_by(data, s::Symbol)

julia> using LightQuery

julia> group_by((a = [1, 1, 2, 2], b = [1, 2, 3, 4]), :a)
2-element Array{NamedTuple{(:a, :b),Tuple{Array{Int64,1},Array{Int64,1}}},1}:
 (a = [1, 1], b = [1, 2])
 (a = [2, 2], b = [3, 4])

source

LightQuery.inner_join — Method.

inner_join(data1, data2)

julia> using LightQuery

julia> inner_join(
            [(a = 1, b = 1), (a = 2, b = 2)],
            [(a = 2, c = 2), (a = 3, c = 3)],
            :a
        )
1-element Array{NamedTuple{(:a, :b, :c),Tuple{Int64,Int64,Int64}},1}:
 (a = 2, b = 2, c = 2)

source

LightQuery.order_by — Method.

order_by(data, s::Symbol)

julia> using LightQuery

julia> order_by((a = [1, 2], b = [2, 1]), :b)
(a = [2, 1], b = [1, 2])

source

LightQuery.pretty — Method.

pretty(data)

julia> using LightQuery

julia> pretty([(a = 1, b = 2), (a = 2, b = 1)]);

source

LightQuery.rows — Method.

rows(data)

julia> using LightQuery

julia> rows((a = [1, 2], b = [2, 1]))
2-element Array{NamedTuple{(:a, :b),Tuple{Int64,Int64}},1}:
 (a = 1, b = 2)
 (a = 2, b = 1)

source

LightQuery.select — Method.

select(data, args...)

julia> using LightQuery

julia> select((a = 1, b = 2, c = 3), :a, :c)
(a = 1, c = 3)

source

LightQuery.transform — Method.

transform(data; kwargs...)

julia> using LightQuery

julia> transform((a = 1, b = 2), c = @_ _.a + _.b)
(a = 1, b = 2, c = 3)

source

LightQuery.where — Method.

where(data, f)

julia> using LightQuery

julia> where((a = [1, 2], b = [2, 1]), @_ _.b .> 1)
(a = [1], b = [2])

source

LightQuery.@> — Macro.

macro >(body)

If body is in the form body_ |> tail_, call @_ on tail, and recur on body.

julia> using LightQuery

julia> @> 0 |> _ + 1 |> _ - 1
0

source

LightQuery.@_ — Macro.

macro _(body::Expr)

Create an Nameless object. The arguments are inside the body; the first arguments is _, the second argument is __, etc. Also stores a quoted version of the function.

julia> using LightQuery

julia> 1 |> @_(_ + 1)
2

julia> map(@_(__ - _), (1, 2), (2, 1))
(1, -1)

julia> @_(_ + 1).expression
:(_ + 1)

source