Interface

LightQuery.Apply

LightQuery.Backwards

LightQuery.By

LightQuery.Enumerate

LightQuery.Group

LightQuery.Length

LightQuery.Name

LightQuery.Peek

LightQuery.Rows

LightQuery.gather

LightQuery.index

LightQuery.key

LightQuery.make_columns

LightQuery.mix

LightQuery.mix

LightQuery.mix

LightQuery.mix

LightQuery.named_tuple

LightQuery.order

LightQuery.over

LightQuery.reduce_rows

LightQuery.remove

LightQuery.rename

LightQuery.row_info

LightQuery.spread

LightQuery.transform

LightQuery.unname

LightQuery.value

LightQuery.when

LightQuery.@>

LightQuery.@_

LightQuery.@if_known

LightQuery.@name

LightQuery.@name_str

LightQuery.Apply — Type

struct Apply{Names}

Apply Names to unnamed values.

julia> using LightQuery

julia> using Test: @inferred

julia> @name @inferred Apply((name"a", name"b", name"c", name"d", name"e", name"f"))((1, 1.0, 1, 1.0, 1, 1.0))
((name"a", 1), (name"b", 1.0), (name"c", 1), (name"d", 1.0), (name"e", 1), (name"f", 1.0))

LightQuery.Backwards — Type

Backwards(something)

Reverse sorting order.

julia> using LightQuery

julia> using Test: @inferred

julia> collect(@inferred order([1, -2], Backwards))
2-element Array{Int64,1}:
  1
 -2

LightQuery.By — Type

By(sorted, key_function)

Mark that sorted has been pre-sorted by key_function. Use with Group, or mix.

julia> using LightQuery

julia> using Test: @inferred

julia> @inferred By([1, -2], abs)
By{Array{Int64,1},typeof(abs)}([1, -2], abs)

LightQuery.Enumerate — Type

Enumerate{Unenumerated}

"Sees through" most iterators into their parent.

julia> using LightQuery

julia> using Test: @inferred

julia> @inferred collect(Enumerate(when([4, 3, 2, 1], iseven)))
2-element Array{Tuple{Int64,Int64},1}:
 (1, 4)
 (3, 2)

LightQuery.Group — Method

Group(ungrouped::By)

Group consecutive keys in ungrouped.

julia> using LightQuery

julia> using Test: @inferred

julia> @inferred collect(Group(By([1, -1, -2, 2, 3, -3], abs)))
3-element Array{Tuple{Int64,SubArray{Int64,1,Array{Int64,1},Tuple{UnitRange{Int64}},true}},1}:
 (1, [1, -1])
 (2, [-2, 2])
 (3, [3, -3])

julia> (@inferred collect(Group(By(Int[], abs)))) == []
true

Requires a presorted object (see By); order first if not.

julia> using LightQuery

julia> using Test: @inferred

julia> first_group =
        @> [2, 1, 2, 1] |>
        order(_, identity) |>
        Group(By(_, identity)) |>
        first;

julia> key(first_group)
1

julia> collect(value(first_group))
2-element Array{Int64,1}:
 1
 1

LightQuery.Length — Type

Length(iterator, new_length)

Allow optimizations based on length.

julia> using LightQuery

julia> collect(Length(when(1:4, iseven), 2))
2-element Array{Int64,1}:
 2
 4

LightQuery.Name — Method

Name(name)

Create a typed Name. Inverse of unname. See also @name_str.

julia> using LightQuery

julia> Name(:a)
name"a"

LightQuery.Peek — Method

Peek(rows, maximum_length = 4)

Peek an iterator which returns named tuples. Will show no more than maximum_length rows.

julia> using LightQuery

julia> @name Peek(Rows((a = 1:5, b = 5:-1:1)))
Showing 4 of 5 rows
| name"a" | name"b" |
| -------:| -------:|
|       1 |       5 |
|       2 |       4 |
|       3 |       3 |
|       4 |       2 |

LightQuery.Rows — Method

Rows(named_columns)

Iterator over rows of a table. Always lazy. Use Peek to view.

julia> using LightQuery

julia> using Test: @inferred

julia> lazy = @name @inferred Rows((a = [1, 2], b = [1.0, 2.0]))
2-element Rows{Tuple{Tuple{Name{:a},Int64},Tuple{Name{:b},Float64}},1,Tuple{Array{Int64,1},Array{Float64,1}},Tuple{Name{:a},Name{:b}}}:
 ((name"a", 1), (name"b", 1.0))
 ((name"a", 2), (name"b", 2.0))

julia> @inferred collect(lazy)
2-element Array{Tuple{Tuple{Name{:a},Int64},Tuple{Name{:b},Float64}},1}:
 ((name"a", 1), (name"b", 1.0))
 ((name"a", 2), (name"b", 2.0))

julia> @name @inferred Rows((a = [1, 2],))
2-element Rows{Tuple{Tuple{Name{:a},Int64}},1,Tuple{Array{Int64,1}},Tuple{Name{:a}}}:
 ((name"a", 1),)
 ((name"a", 2),)

All arguments to Rows must have the same axes. Use @inbounds to override the check.

julia> result = @name Rows((a = 1:2, b = 1:3))
ERROR: DimensionMismatch("All arguments to `Rows` must have the same axes")
[...]

LightQuery.gather — Method

gather(data, new_name_old_names...)

For each new_name, old_names pair in new_name_old_names, gather the old_names into a single new_name. Inverse of spread.

julia> using LightQuery

julia> using Test: @inferred

julia> data = @name (a = 1, b = 1.0, c = 1, d = 1.0, e = 1, f = 1.0);

julia> @name @inferred gather(data, g = (name"b", name"e"), h = (name"c", name"f"))
((name"a", 1), (name"d", 1.0), (name"g", ((name"b", 1.0), (name"e", 1))), (name"h", ((name"c", 1), (name"f", 1.0))))

LightQuery.index — Method

index(unindexed, key_function)

Index unindexed by the results of key_function. Results of key_function must be unique.

julia> using LightQuery

julia> using Test: @inferred

julia> result = @inferred index([-2, 1], abs)
LightQuery.Indexed{Int64,Int64,Array{Int64,1},Dict{Int64,Int64}} with 2 entries:
  2 => -2
  1 => 1

julia> @inferred result[2]
-2

LightQuery.key — Method

key(pair)

The key in a key => value pair.

julia> using LightQuery

julia> using Test: @inferred

julia> @inferred key(:a => 1)
:a

julia> @inferred key((:a, 1))
:a

LightQuery.make_columns — Method

make_columns(rows)

Collect into columns.

julia> using LightQuery

julia> using Test: @inferred

julia> stable(x) = @name (a = x, b = x + 0.0, c = x, d = x + 0.0, e = x, f = x + 0.0);

julia> @inferred make_columns(over(1:4, stable))
((name"a", [1, 2, 3, 4]), (name"b", [1.0, 2.0, 3.0, 4.0]), (name"c", [1, 2, 3, 4]), (name"d", [1.0, 2.0, 3.0, 4.0]), (name"e", [1, 2, 3, 4]), (name"f", [1.0, 2.0, 3.0, 4.0]))

julia> unstable(x) =
            @name if x <= 2
                (a = missing, b = string(x), d = string(x))
            else
                (a = x, b = missing, c = x)
            end;

julia> make_columns(over(1:4, unstable))
((name"d", Union{Missing, String}["1", "2", missing, missing]), (name"a", Union{Missing, Int64}[missing, missing, 3, 4]), (name"b", Union{Missing, String}["1", "2", missing, missing]), (name"c", Union{Missing, Int64}[missing, missing, 3, 4]))

julia> make_columns(when(over(1:4, unstable), row -> true))
((name"d", Union{Missing, String}["1", "2", missing, missing]), (name"a", Union{Missing, Int64}[missing, missing, 3, 4]), (name"b", Union{Missing, String}["1", "2", missing, missing]), (name"c", Union{Missing, Int64}[missing, missing, 3, 4]))

LightQuery.mix — Method

mix(::Name{:inner},  left::By, right::By)

Find all pairs where isequal(left.key_function(left.sorted), right.key_function(right.sorted)). Assumes left and right are both strictly sorted (no repeats). If there are repeats, Group first. See By.

julia> using LightQuery

julia> using Test: @inferred

julia> @name @inferred collect(mix(name"inner", By([1, -2, 5, -6], abs), By([-1, 3, -4, 6], abs)))
2-element Array{Tuple{Int64,Int64},1}:
 (1, -1)
 (-6, 6)

julia> (@name @inferred collect(mix(name"inner", By(Int[], abs), By(Int[], abs)))) == []
true

julia> (@name @inferred collect(mix(name"inner", By([1], abs), By(Int[], abs)))) == []
true

julia> (@name @inferred collect(mix(name"inner", By(Int[], abs), By([1], abs)))) == []
true

LightQuery.mix — Method

mix(::Name{:left}, left::By, right::By)

Find all pairs where isequal(left.key_function(left.sorted), right.key_function(right.sorted)), using missing when there is no right match. Assumes left and right are both strictly sorted (no repeats). If there are repeats, Group first. See By.

julia> using LightQuery

julia> @name collect(mix(name"left", By([1, -2, 5, -6], abs), By([-1, 3, -4, 6], abs)))
4-element Array{Tuple{Int64,Union{Missing, Int64}},1}:
 (1, -1)
 (-2, missing)
 (5, missing)
 (-6, 6)

julia> (@name collect(mix(name"left", By(Int[], abs), By(Int[], abs)))) == []
true

julia> @name collect(mix(name"left", By([1], abs), By(Int[], abs)))
1-element Array{Tuple{Int64,Union{Missing, Int64}},1}:
 (1, missing)

julia> (@name collect(mix(name"left", By(Int[], abs), By([1], abs)))) == []
true

LightQuery.mix — Method

mix(::Name{:outer}, left::By, right::By)

Find all pairs where isequal(left.key_function(left.sorted), right.key_function(right.sorted)), using missing when there is no left or right match. Assumes left and right are both strictly sorted (no repeats). If there are repeats, Group first. See By.

julia> using LightQuery

julia> @name collect(mix(name"outer", By([1, -2, 5, -6], abs), By([-1, 3, -4, 6], abs)))
6-element Array{Tuple{Union{Missing, Int64},Union{Missing, Int64}},1}:
 (1, -1)
 (-2, missing)
 (missing, 3)
 (missing, -4)
 (5, missing)
 (-6, 6)

julia> (@name collect(mix(name"outer", By(Int[], abs), By(Int[], abs)))) == []
true

julia> @name collect(mix(name"outer", By([1], abs), By(Int[], abs)))
1-element Array{Tuple{Union{Missing, Int64},Union{Missing, Int64}},1}:
 (1, missing)

julia> @name collect(mix(name"outer", By(Int[], abs), By([1], abs)))
1-element Array{Tuple{Union{Missing, Int64},Union{Missing, Int64}},1}:
 (missing, 1)

LightQuery.mix — Method

mix(::Name{:right}, left::By, right::By)

Find all pairs where isequal(left.key_function(left.sorted), right.key_function(right.sorted)), using missing when there is no left match. Assumes left and right are both strictly sorted (no repeats). If there are repeats, Group first. See By.

julia> using LightQuery

julia> @name collect(mix(name"right", By([1, -2, 5, -6], abs), By([-1, 3, -4, 6], abs)))
4-element Array{Tuple{Union{Missing, Int64},Int64},1}:
 (1, -1)
 (missing, 3)
 (missing, -4)
 (-6, 6)

julia> (@name collect(mix(name"right", By(Int[], abs), By(Int[], abs)))) == []
true

julia> (@name collect(mix(name"right", By([1], abs), By(Int[], abs)))) == []
true

julia> @name collect(mix(name"right", By(Int[], abs), By([1], abs)))
1-element Array{Tuple{Union{Missing, Int64},Int64},1}:
 (missing, 1)

LightQuery.named_tuple — Method

named_tuple(data)

Convert data to a named_tuple (see @name).

julia> using LightQuery

julia> using Test: @inferred

julia> @inferred named_tuple((a = 1, b = 1.0, c = 1, d = 1.0, e = 1, f = 1.0))
((name"a", 1), (name"b", 1.0), (name"c", 1), (name"d", 1.0), (name"e", 1), (name"f", 1.0))

For stability working with arbitrary structs, propertynames must constant propagate.

julia> struct MyType
            a::Int
            b::Float64
            c::Int
            d::Float64
            e::Int
            f::Float64
        end

julia> import Base: propertynames

julia> @inline propertynames(::MyType) = (:a, :b, :c, :d, :e, :f);

julia> @inferred named_tuple(MyType(1, 1.0, 1, 1.0, 1, 1.0))
((name"a", 1), (name"b", 1.0), (name"c", 1), (name"d", 1.0), (name"e", 1), (name"f", 1.0))

LightQuery.order — Method

order(unordered, key_function; keywords...)

Generalized sort. keywords will be passed to sort!; see the documentation there for options. Use By to mark that an object has been sorted. If the results of key_function are type unstable, consider using hash ∘ key_function instead.

julia> using LightQuery

julia> using Test: @inferred

julia> collect(@inferred order([-2, 1], abs))
2-element Array{Int64,1}:
  1
 -2

LightQuery.over — Method

over(iterator, call)

Lazy map with the reverse argument order.

julia> using LightQuery

julia> using Test: @inferred

julia> @inferred collect(over([1, -2, -3, 4], abs))
4-element Array{Int64,1}:
 1
 2
 3
 4

LightQuery.reduce_rows — Method

function reduce_rows(rows, a_function, columns...)

Reduce a function over each of columns in rows.

julia> using LightQuery

julia> using Test: @inferred

julia> @name @inferred reduce_rows(Rows((a = [1, 1], b = [1.0, 1.0])), +, name"a", name"b")
((name"a", 2), (name"b", 2.0))

LightQuery.remove — Method

remove(data, old_names...)

Remove old_names from data.

julia> using LightQuery

julia> using Test: @inferred

julia> data = @name (a = 1, b = 1.0, c = 1, d = 1.0, e = 1, f = 1.0);

julia> @name @inferred remove(data, name"c", name"f")
((name"a", 1), (name"b", 1.0), (name"d", 1.0), (name"e", 1))

LightQuery.rename — Method

rename(data, new_name_old_names...)

Rename data.

julia> using LightQuery

julia> using Test: @inferred

julia> data = @name (a = 1, b = 1.0, c = 1, d = 1.0, e = 1, f = 1.0);

julia> @name @inferred rename(data, c2 = name"c", f2 = name"f")
((name"a", 1), (name"b", 1.0), (name"d", 1.0), (name"e", 1), (name"c2", 1), (name"f2", 1.0))

LightQuery.row_info — Method

row_info(::CSV.File)

Get row info for the CSV file. Can be used as a type stable selector function.

julia> using LightQuery

julia> using Test: @inferred

julia> using CSV: File

julia> test = File("test.csv");

julia> template = row_info(test)
(LightQuery.InRow{:a,CSV.Column{Int64,Int64}}([1]), LightQuery.InRow{:b,CSV.Column{Float64,Float64}}([1.0]), LightQuery.InRow{:c,CSV.Column{Int64,Int64}}([1]), LightQuery.InRow{:d,CSV.Column{Float64,Float64}}([1.0]), LightQuery.InRow{:e,CSV.Column{Int64,Int64}}([1]), LightQuery.InRow{:f,CSV.Column{Float64,Float64}}([1.0]))

julia> @inferred template(first(test))
((name"a", 1), (name"b", 1.0), (name"c", 1), (name"d", 1.0), (name"e", 1), (name"f", 1.0))

LightQuery.spread — Method

spread(data, some_names...)

Unnest nested named tuples. Inverse of gather.

julia> using LightQuery

julia> using Test: @inferred

julia> gathered = @name (a = 1, d = 1.0, g = (b = 1.0, e = 1), h = (c = 1, f = 1.0));

julia> @name @inferred spread(gathered, name"g", name"h")
((name"a", 1), (name"d", 1.0), (name"b", 1.0), (name"e", 1), (name"c", 1), (name"f", 1.0))

LightQuery.transform — Method

transform(data, assignments...)

Merge assignments into data, overwriting old values.

julia> using LightQuery

julia> using Test: @inferred

julia> data = @name (a = 1, b = 1.0, c = 1, d = 1.0, e = 1, f = 1.0);

julia> @name @inferred transform(data, c = 2.0, f = 2, g = 1, h = 1.0)
((name"a", 1), (name"b", 1.0), (name"d", 1.0), (name"e", 1), (name"c", 2.0), (name"f", 2), (name"g", 1), (name"h", 1.0))

LightQuery.unname — Method

unname(::Name{name}) where name

Inverse of Name.

julia> using LightQuery

julia> using Test: @inferred

julia> @inferred unname(Name(:a))
:a

LightQuery.value — Method

value(pair)

The value in a key => value pair.

julia> using LightQuery

julia> using Test: @inferred

julia> @inferred value(:a => 1)
1

julia> @inferred value((:a, 1))
1

LightQuery.when — Method

when(iterator, call)

Lazy filter with the reverse argument order.

julia> using LightQuery

julia> using Test: @inferred

julia> @inferred collect(when(1:4, iseven))
2-element Array{Int64,1}:
 2
 4

LightQuery.@> — Macro

macro >(body)

If body is in the form object_ |> call_, call @_ on call, and recur on object.

julia> using LightQuery

julia> @> 0 |> _ - 1 |> abs
1

You can nest chains:

julia> @> 1 |> (@> _ + 1 |> _ + 1)
3

Handles interpolations seamlessly:

julia> @> 1 |> :(_ + $_)
:(_ + 1)

LightQuery.@_ — Macro

macro _(body)

Terser function syntax. The arguments are inside the body; the first argument is _, the second argument is __, etc. Will @inline.

julia> using LightQuery

julia> using Test: @inferred

julia> @inferred (@_ _ + 1)(1)
2

julia> @inferred map((@_ __ - _), (1, 2), (2, 1))
(1, -1)

If there are no _ arguments, read as is.

julia> (@_ x -> x + 1)(1)
2

LightQuery.@if_known — Macro

@if_known(something)

If something is missing, return missing, otherwise, something.

julia> using LightQuery

julia> using Test: @inferred

julia> function test(x)
            first(@if_known(x))
        end;

julia> @inferred test((1, 2))
1

julia> @inferred test(missing)
missing

LightQuery.@name — Macro

macro name(code)

Switch to named_tuples.

julia> using LightQuery

julia> using Test: @inferred

julia> data = @name (a = 1, b = 1.0, c = 1, d = 1.0, e = 1, f = 1.0)
((name"a", 1), (name"b", 1.0), (name"c", 1), (name"d", 1.0), (name"e", 1), (name"f", 1.0))

based on typed Names.

julia> name"a"
name"a"

Names can be used as properties

julia> @name @inferred data.c
1

julia> @name data.g
ERROR: BoundsError: attempt to access ((name"a", 1), (name"b", 1.0), (name"c", 1), (name"d", 1.0), (name"e", 1), (name"f", 1.0))
  at index [name"g"]
[...]

and selector functions.

julia> @name @inferred name"c"(data)
1

Multiple names can be used as selector functions

julia> @name @inferred (name"c", name"f")(data)
((name"c", 1), (name"f", 1.0))

You can also convert back to NamedTuples.

julia> @inferred NamedTuple(data)
(a = 1, b = 1.0, c = 1, d = 1.0, e = 1, f = 1.0)

LightQuery.@name_str — Macro

name_str(name)

Create a typed Name.

julia> using LightQuery

julia> name"a"
name"a"