Home

LightQuery.jl

LightQuery.@_

LightQuery.@query

LightQuery.@_ — Macro.

macro _(body::Expr)

Another syntax for anonymous functions. The arguments are inside the body; the first arguments is _, the second argument is __, etc.

julia> using LightQuery

julia> 1 |> (@_ _ + 1)
2

julia> map((@_ __ - _), (1, 2), (2, 1))
(1, -1)

source

LightQuery.@query — Macro.

macro query(body::Expr)

Query your code. If body is a chain head_ |> tail_, recur on head. If tail is a function call, and the function ends with a number (the parity), anonymize and quote arguments past that parity. Either way, anonymize the whole tail, then call it on head.

julia> using LightQuery

julia> call(source1, source2, anonymous, quoted) = anonymous(source1, source2);

julia> @query 1 |> (_ - 2) |> abs(_) |> call2(_, 2, _ + __)
3

julia> @query 1 |> call2(_)
ERROR: LoadError: Expecting at least 2 argument(s)
[...]

source