Tutorial

Tutorial

I'm going to use the flights data from the dplyr tutorial. This data is in the test folder of this package. I re-export CSV for input-output.

julia> using LightQuery

julia> airports_file = CSV.File("airports.csv", allowmissing = :auto)
CSV.File("airports.csv", rows=1458):
Tables.Schema:
 :faa    String
 :name   String
 :lat    Float64
 :lon    Float64
 :alt    Int64
 :tz     Int64
 :dst    String
 :tzone  String

allowmissing = :auto tells CSV to guess whether columns might contain misisng data.

Let's take a look at the first row. Use named_tuple to coerce a CSV.Row to a NamedTuple. I'm going to make heavy use of the chaining macro @> and lazy call macro @_.

julia> airport =
        @> airports_file |>
        first |>
        named_tuple
(faa = "04G", name = "Lansdowne Airport", lat = 41.1304722, lon = -80.6195833, alt = 1044, tz = -5, dst = "A", tzone = "America/New_York")

As a start, I want to rename so that I understand what the columns mean. When you rename, names need to be wrapped with Name.

julia> airport =
        @> airport |>
        rename(_,
            airport_code = Name(:faa),
            latitude = Name(:lat),
            longitude = Name(:lon),
            altitude = Name(:alt),
            time_zone_offset = Name(:tz),
            daylight_savings = Name(:dst),
            time_zone = Name(:tzone)
        )
(name = "Lansdowne Airport", airport_code = "04G", latitude = 41.1304722, longitude = -80.6195833, altitude = 1044, time_zone_offset = -5, daylight_savings = "A", time_zone = "America/New_York")

Let's create a proper TimeZone, and add units to our variables.

julia> using TimeZones: VariableTimeZone, TimeZone

julia> using Unitful: °, ft

julia> airport =
        @> airport |>
        transform(_,
            time_zone = TimeZone(_.time_zone),
            latitude = _.latitude * °,
            longitude = _.longitude * °,
            altitude = _.altitude * ft
        )
(name = "Lansdowne Airport", airport_code = "04G", latitude = 41.1304722°, longitude = -80.6195833°, altitude = 1044 ft, time_zone_offset = -5, daylight_savings = "A", time_zone = America/New_York (UTC-5/UTC-4))

Now that we have a true timezone, we can remove all data that is contingent on timezone.

julia> airport =
        @> airport |>
        remove(_,
            :time_zone_offset,
            :daylight_savings
        )
(name = "Lansdowne Airport", airport_code = "04G", latitude = 41.1304722°, longitude = -80.6195833°, altitude = 1044 ft, time_zone = America/New_York (UTC-5/UTC-4))

Notice that we know that there will be one entry for each airport code. This signals that this data might be best stored as a Dict. Let's put everything together:

julia> const airports = Dict(
            airport.airport_code => remove(airport, :airport_code)
        )
Dict{String,NamedTuple{(:name, :latitude, :longitude, :altitude, :time_zone),Tuple{String,Unitful.Quantity{Float64,NoDims,Unitful.FreeUnits{(°,),NoDims,nothing}},Unitful.Quantity{Float64,NoDims,Unitful.FreeUnits{(°,),NoDims,nothing}},Unitful.Quantity{Int64,𝐋,Unitful.FreeUnits{(ft,),𝐋,nothing}},VariableTimeZone}}} with 1 entry:
  "04G" => (name = "Lansdowne Airport", latitude = 41.1305°, longitude = -80.61…

julia> function process_airport(airport_row)
            airport =
                @> named_tuple(airport_row) |>
                rename(_,
                    airport_code = Name(:faa),
                    latitude = Name(:lat),
                    longitude = Name(:lon),
                    altitude = Name(:alt),
                    time_zone_offset = Name(:tz),
                    daylight_savings = Name(:dst),
                    time_zone = Name(:tzone)
                ) |>
                transform(_,
                    time_zone = TimeZone(_.time_zone),
                    latitude = _.latitude * °,
                    longitude = _.longitude * °,
                    altitude = _.altitude * ft
                ) |>
                remove(_,
                    :time_zone_offset,
                    :daylight_savings
                )
            airports[airport.airport_code] = remove(airport, :airport_code)
        end
process_airport (generic function with 1 method)

julia> foreach(process_airport, airports_file)
ERROR: ArgumentError: Unknown time zone "Asia/Chongqing"

Uh oh. A bit of googling shows that "Asia/Chongqing" is an alias for "Asia/Shanghai". Try again.

julia> function process_airport(airport_row)
            airport =
                @> named_tuple(airport_row) |>
                rename(_,
                    airport_code = Name(:faa),
                    latitude = Name(:lat),
                    longitude = Name(:lon),
                    altitude = Name(:alt),
                    time_zone_offset = Name(:tz),
                    daylight_savings = Name(:dst),
                    time_zone = Name(:tzone)
                ) |>
                transform(_,
                    time_zone = TimeZone(
                        if _.time_zone == "Asia/Chongqing"
                            "Asia/Shanghai"
                        else
                            _.time_zone
                        end
                    ),
                    latitude = _.latitude * °,
                    longitude = _.longitude * °,
                    altitude = _.altitude * ft
                ) |>
                remove(_,
                    :time_zone_offset,
                    :daylight_savings
                )
            airports[airport.airport_code] = remove(airport, :airport_code)
            nothing
        end
process_airport (generic function with 1 method)

julia> foreach(process_airport, airports_file)
ERROR: ArgumentError: Unknown time zone "\N"

Hadley, you're killing me. "\N" is definitely not a timezone. I'm not in the mood for playing games; let's just ignore those airports. I use when to lazily filter aiports.

julia> @> airports_file |>
        when(_, @_ _.tzone != "\\N") |>
        foreach(process_airport, _)

We did it! That was just the warm-up. Now let's get started working on the flights data.

julia> flights_file = CSV.File("flights.csv", allowmissing = :auto)
CSV.File("flights.csv", rows=336776):
Tables.Schema:
 :year            Int64
 :month           Int64
 :day             Int64
 :dep_time        Union{Missing, Int64}
 :sched_dep_time  Int64
 :dep_delay       Union{Missing, Int64}
 :arr_time        Union{Missing, Int64}
 :sched_arr_time  Int64
 :arr_delay       Union{Missing, Int64}
 :carrier         String
 :flight          Int64
 :tailnum         Union{Missing, String}
 :origin          String
 :dest            String
 :air_time        Union{Missing, Int64}
 :distance        Int64
 :hour            Int64
 :minute          Int64
 :time_hour       String

julia> flight =
        @> flights_file |>
        first |>
        named_tuple(_)
(year = 2013, month = 1, day = 1, dep_time = 517, sched_dep_time = 515, dep_delay = 2, arr_time = 830, sched_arr_time = 819, arr_delay = 11, carrier = "UA", flight = 1545, tailnum = "N14228", origin = "EWR", dest = "IAH", air_time = 227, distance = 1400, hour = 5, minute = 15, time_hour = "2013-01-01 05:00:00")

Again, some renaming:

julia> flight =
        @> flight |>
        rename(_,
            departure_time = Name(:dep_time),
            scheduled_departure_time = Name(:sched_dep_time),
            departure_delay = Name(:dep_delay),
            arrival_time = Name(:arr_time),
            scheduled_arrival_time = Name(:sched_arr_time),
            arrival_delay = Name(:arr_delay),
            tail_number = Name(:tailnum),
            destination = Name(:dest)
        )
(year = 2013, month = 1, day = 1, carrier = "UA", flight = 1545, origin = "EWR", air_time = 227, distance = 1400, hour = 5, minute = 15, time_hour = "2013-01-01 05:00:00", departure_time = 517, scheduled_departure_time = 515, departure_delay = 2, arrival_time = 830, scheduled_arrival_time = 819, arrival_delay = 11, tail_number = "N14228", destination = "IAH")

We can use our airports data to make datetimes with timezones.

julia> using Dates: DateTime

julia> using TimeZones: ZonedDateTime

julia> scheduled_departure_time = ZonedDateTime(
            DateTime(flight.year, flight.month, flight.day, flight.hour, flight.minute),
            airports[flight.origin].time_zone
        )
2013-01-01T05:15:00-05:00

Note the scheduled arrival time is 818. This means 8:18. We can use divrem(_, 100) to split it up.

julia> scheduled_arrival_time = ZonedDateTime(
            DateTime(flight.year, flight.month, flight.day, divrem(flight.scheduled_arrival_time, 100)...),
            airports[flight.destination].time_zone
        )
2013-01-01T08:19:00-06:00

What if it was an over-night flight? We can add a day to the arrival time if it wasn't later than the departure time.

julia> using Dates: Day

julia> if !(scheduled_arrival_time > scheduled_departure_time)
            scheduled_arrival_time = scheduled_arrival_time + Day(1)
        end

Now let's add the data back into our flight, and remove redundant columns.

julia> flight =
        @> flight |>
        transform(_,
            scheduled_departure_time = scheduled_departure_time,
            scheduled_arrival_time = scheduled_arrival_time
        ) |>
        remove(_, :year, :month, :day, :hour, :minute, :time_hour,
            :departure_time, :arrival_time)
(carrier = "UA", flight = 1545, origin = "EWR", air_time = 227, distance = 1400, scheduled_departure_time = 2013-01-01T05:15:00-05:00, departure_delay = 2, scheduled_arrival_time = 2013-01-01T08:19:00-06:00, arrival_delay = 11, tail_number = "N14228", destination = "IAH")

Now let's add in some units:

julia> using Dates: Minute

julia> using Unitful: mi

julia> flight =
        @> flight |>
        transform(_,
            air_time = Minute(_.air_time),
            distance = _.distance * mi,
            departure_delay = Minute(_.departure_delay),
            arrival_delay = Minute(_.arrival_delay)
        )
(carrier = "UA", flight = 1545, origin = "EWR", air_time = Minute(227), distance = 1400 mi, scheduled_departure_time = 2013-01-01T05:15:00-05:00, departure_delay = Minute(2), scheduled_arrival_time = 2013-01-01T08:19:00-06:00, arrival_delay = Minute(11), tail_number = "N14228", destination = "IAH")

Put it all together. I'm conducting some mild type piracy to get Minute to work with missing data. When it comes time to collect, I'm calling make_columns then rows. It makes sense to store this data column-wise. This is because there are multiple columns that might contain missing data. I use over to lazily map. Note that I'm only considering flights with corresponding airport data.

julia> import Dates: Minute

julia> Minute(::Missing) = missing
Minute

julia> function process_flight(row)
            flight =
                @> named_tuple(row) |>
                rename(_,
                    departure_time = Name(:dep_time),
                    scheduled_departure_time = Name(:sched_dep_time),
                    departure_delay = Name(:dep_delay),
                    arrival_time = Name(:arr_time),
                    scheduled_arrival_time = Name(:sched_arr_time),
                    arrival_delay = Name(:arr_delay),
                    tail_number = Name(:tailnum),
                    destination = Name(:dest)
                )
            scheduled_departure_time = ZonedDateTime(
                DateTime(flight.year, flight.month, flight.day, flight.hour, flight.minute),
                airports[flight.origin].time_zone
            )
            scheduled_arrival_time = ZonedDateTime(
                DateTime(flight.year, flight.month, flight.day, divrem(flight.scheduled_arrival_time, 100)...),
                airports[flight.destination].time_zone
            )
            if !(scheduled_arrival_time > scheduled_departure_time)
                scheduled_arrival_time = scheduled_arrival_time + Day(1)
            end
            @> flight |>
                transform(_,
                    scheduled_departure_time = scheduled_departure_time,
                    scheduled_arrival_time = scheduled_arrival_time,
                    air_time = Minute(_.air_time),
                    distance = _.distance * mi,
                    departure_delay = Minute(_.departure_delay),
                    arrival_delay = Minute(_.arrival_delay)
                ) |>
                remove(_, :year, :month, :day, :hour, :minute, :time_hour,
                    :departure_time, :arrival_time)
        end
process_flight (generic function with 1 method)

julia> flights =
        @> flights_file |>
        when(_, @_ haskey(airports, _.origin) && haskey(airports, _.dest)) |>
        over(_, process_flight) |>
        make_columns |>
        rows;

julia> Peek(flights)
Showing 4 of 329174 rows
| :carrier | :flight | :origin |   :air_time | :distance | :scheduled_departure_time | :departure_delay |   :scheduled_arrival_time | :arrival_delay | :tail_number | :destination |
| --------:| -------:| -------:| -----------:| ---------:| -------------------------:| ----------------:| -------------------------:| --------------:| ------------:| ------------:|
|       UA |    1545 |     EWR | 227 minutes |   1400 mi | 2013-01-01T05:15:00-05:00 |        2 minutes | 2013-01-01T08:19:00-06:00 |     11 minutes |       N14228 |          IAH |
|       UA |    1714 |     LGA | 227 minutes |   1416 mi | 2013-01-01T05:29:00-05:00 |        4 minutes | 2013-01-01T08:30:00-06:00 |     20 minutes |       N24211 |          IAH |
|       AA |    1141 |     JFK | 160 minutes |   1089 mi | 2013-01-01T05:40:00-05:00 |        2 minutes | 2013-01-01T08:50:00-05:00 |     33 minutes |       N619AA |          MIA |
|       DL |     461 |     LGA | 116 minutes |    762 mi | 2013-01-01T06:00:00-05:00 |       -6 minutes | 2013-01-01T08:37:00-05:00 |    -25 minutes |       N668DN |          ATL |

You might notice that if the origin and the destination are the same, then the distance is also the same. We can see this by ordering the data. Note that Names can be used as a function to select columns.

julia> by_path =
        @> flights |>
        order(_, Names(:origin, :destination));

julia> Peek(by_path)
Showing 4 of 329174 rows
| :carrier | :flight | :origin |  :air_time | :distance | :scheduled_departure_time | :departure_delay |   :scheduled_arrival_time | :arrival_delay | :tail_number | :destination |
| --------:| -------:| -------:| ----------:| ---------:| -------------------------:| ----------------:| -------------------------:| --------------:| ------------:| ------------:|
|       EV |    4112 |     EWR | 33 minutes |    143 mi | 2013-01-01T13:17:00-05:00 |       -2 minutes | 2013-01-01T14:23:00-05:00 |    -10 minutes |       N13538 |          ALB |
|       EV |    3260 |     EWR | 36 minutes |    143 mi | 2013-01-01T16:21:00-05:00 |       34 minutes | 2013-01-01T17:24:00-05:00 |     40 minutes |       N19554 |          ALB |
|       EV |    4170 |     EWR | 31 minutes |    143 mi | 2013-01-01T20:04:00-05:00 |       52 minutes | 2013-01-01T21:12:00-05:00 |     44 minutes |       N12540 |          ALB |
|       EV |    4316 |     EWR | 33 minutes |    143 mi | 2013-01-02T13:27:00-05:00 |        5 minutes | 2013-01-02T14:33:00-05:00 |    -14 minutes |       N14153 |          ALB |

How can we remove this redundant information from our dataset? Let's make a distances dataset.

julia> const distances = Dict(
            Names(:origin, :destination)(flight) => flight.distance
        )
Dict{NamedTuple{(:origin, :destination),Tuple{String,String}},Unitful.Quantity{Int64,𝐋,Unitful.FreeUnits{(mi,),𝐋,nothing}}} with 1 entry:
  (origin = "EWR", destination = "IAH") => 1400 mi

Now we can Group our flights data By the path. Each group will be a key (the origin and destination) mapped to a value (a sub-table). We can group the already sorted data.

julia> @> by_path |>
        Group(By(_, Names(:origin, :destination))) |>
        foreach((@_ distances[key(_)] = first(value(_)).distance), _)

We can drop the distance variable from flights now. To do this efficiently, we use columns, which is always lazy.

julia> flights =
        @> flights |>
        columns |>
        remove(_, :distance) |>
        rows;

Let's take a look at all of our beautiful data!

julia> Peek(airports)
Showing 4 of 1455 rows
| :first |                                                                                                                                             :second |
| ------:| ---------------------------------------------------------------------------------------------------------------------------------------------------:|
|    JES |  (name = "Jesup-Wayne County Airport", latitude = 31.553889°, longitude = -81.8825°, altitude = 107 ft, time_zone = America/New_York (UTC-5/UTC-4)) |
|    PPV | (name = "Port Protection Seaplane Base", latitude = 56.328889°, longitude = -133.61°, altitude = 0 ft, time_zone = America/Anchorage (UTC-9/UTC-8)) |
|    DTA | (name = "Delta Municipal Airport", latitude = 39.3806386°, longitude = -112.5077147°, altitude = 4759 ft, time_zone = America/Denver (UTC-7/UTC-6)) |
|    X21 |         (name = "Arthur Dunn Airpark", latitude = 28.622552°, longitude = -80.83541°, altitude = 30 ft, time_zone = America/New_York (UTC-5/UTC-4)) |

julia> Peek(flights)
Showing 4 of 329174 rows
| :carrier | :flight | :origin |   :air_time | :scheduled_departure_time | :departure_delay |   :scheduled_arrival_time | :arrival_delay | :tail_number | :destination |
| --------:| -------:| -------:| -----------:| -------------------------:| ----------------:| -------------------------:| --------------:| ------------:| ------------:|
|       UA |    1545 |     EWR | 227 minutes | 2013-01-01T05:15:00-05:00 |        2 minutes | 2013-01-01T08:19:00-06:00 |     11 minutes |       N14228 |          IAH |
|       UA |    1714 |     LGA | 227 minutes | 2013-01-01T05:29:00-05:00 |        4 minutes | 2013-01-01T08:30:00-06:00 |     20 minutes |       N24211 |          IAH |
|       AA |    1141 |     JFK | 160 minutes | 2013-01-01T05:40:00-05:00 |        2 minutes | 2013-01-01T08:50:00-05:00 |     33 minutes |       N619AA |          MIA |
|       DL |     461 |     LGA | 116 minutes | 2013-01-01T06:00:00-05:00 |       -6 minutes | 2013-01-01T08:37:00-05:00 |    -25 minutes |       N668DN |          ATL |

julia> Peek(distances)
Showing 4 of 217 rows
|                                :first | :second |
| -------------------------------------:| -------:|
| (origin = "LGA", destination = "PHL") |   96 mi |
| (origin = "JFK", destination = "SAN") | 2446 mi |
| (origin = "EWR", destination = "MSY") | 1167 mi |
| (origin = "JFK", destination = "MSY") | 1182 mi |

Interface

Macros

LightQuery.@_ — Macro.

macro _(body)

Terser function syntax. The arguments are inside the body; the first argument is _, the second argument is __, etc. Will always @inline.

julia> using LightQuery

julia> (@_ _ + 1)(1)
2

julia> map((@_ __ - _), (1, 2), (2, 1))
(1, -1)

source

LightQuery.@> — Macro.

macro >(body)

If body is in the form body_ |> tail_, call @_ on tail, and recur on body.

julia> using LightQuery

julia> @> 0 |> _ - 1 |> abs
1

source

Columns

LightQuery.named_tuple — Function.

named_tuple(it)

Coerce to a named_tuple. For performance with working with arbitrary structs, define and @inline propertynames.

julia> using LightQuery

julia> @inline Base.propertynames(p::Pair) = (:first, :second)

julia> named_tuple(:a => 1)
(first = :a, second = 1)

source

LightQuery.Name — Type.

Name(name)

Create a typed name. Can be used as a function to getproperty, with a default to missing. For multiple names, see Names.

julia> using LightQuery

julia> (a = 1,) |>
        Name(:a)
1

julia> (a = 1,) |>
        Name(:b)
missing

julia> missing |>
        Name(:a)
missing

source

LightQuery.Names — Type.

Names(the_names...)

Create typed names. Can be used to as a function to assign or select names, with a default to missing. For just one name, see Name.

julia> using LightQuery

julia> (1, 1.0) |>
        Names(:a, :b)
(a = 1, b = 1.0)

julia> (a = 1, b = 1.0) |>
        Names(:a)
(a = 1,)

julia> (a = 1,) |>
        Names(:a, :b)
(a = 1, b = missing)

julia> missing |>
        Names(:a)
(a = missing,)

source

LightQuery.rename — Function.

rename(it; renames...)

Rename it. Because constants do not constant propagate through key-word arguments, wrap with Name.

julia> using LightQuery

julia> @> (a = 1, b = 1.0) |>
        rename(_, c = Name(:a))
(b = 1.0, c = 1)

source

LightQuery.transform — Function.

transform(it; assignments...)

Merge assignments into it. Inverse of remove.

julia> using LightQuery

julia> @> (a = 1,) |>
        transform(_, b = 1.0)
(a = 1, b = 1.0)

source

LightQuery.remove — Function.

remove(it, the_names...)

Remove the_names. Inverse of transform.

julia> using LightQuery

julia> @> (a = 1, b = 1.0) |>
        remove(_, :b)
(a = 1,)

source

LightQuery.gather — Function.

gather(it; assignments...)

For each key => value pair in assignments, gather the Names in value into a single key. Inverse of spread.

julia> using LightQuery

julia> @> (a = 1, b = 1.0, c = 1//1) |>
        gather(_, d = Names(:a, :c))
(b = 1.0, d = (a = 1, c = 1//1))

source

LightQuery.spread — Function.

spread(it::NamedTuple, the_names...)

Unnest nested it in name. Inverse of gather.

julia> using LightQuery

julia> @> (b = 1.0, d = (a = 1, c = 1//1)) |>
        spread(_, :d)
(b = 1.0, a = 1, c = 1//1)

source

Rows

LightQuery.unzip — Function.

unzip(it, n)

Unzip an iterator it which returns tuples of length n. Use Val(n) to guarantee type stability.

julia> using LightQuery

julia> unzip([(1, 1.0), (2, 2.0)], 2)
([1, 2], [1.0, 2.0])

julia> unzip([(1, 1.0), (2, 2.0)], Val(2))
([1, 2], [1.0, 2.0])

source

LightQuery.over — Function.

over(it, call)

Lazy map with argument order reversed.

source

LightQuery.when — Function.

when(it, call)

Lazy filter with argument order reversed.

source

LightQuery.order — Function.

order(it, call; keywords...)

Generalized sort. keywords will be passed to sort!; see the documentation there for options. See By for a way to explicitly mark that an object has been sorted.

julia> using LightQuery

julia> order([2, 1], identity)
2-element view(::Array{Int64,1}, [2, 1]) with eltype Int64:
 1
 2

source

    order(it, call, condition; keywords...)

If `call` is not type stable, consider adding a `condition` to filter.

jldoctest order julia> using LightQuery

julia> order([2, 1, missing], identity, !ismissing) 2-element view(::Array{Union{Missing, Int64},1}, [2, 1]) with eltype Union{Missing, Int64}: 1 2 ```

source

LightQuery.By — Type.

By(it, call)

Mark that it has been pre-sorted by call. For use with Group or Join.

julia> using LightQuery

julia> By([1, 2], identity)
By{Array{Int64,1},typeof(identity)}([1, 2], identity)

source

LightQuery.Group — Type.

Group(it::By)

Group consecutive keys in it. Requires a presorted object (see By). Relies on the fact that iteration states can be converted to indices; thus, you might have to define LightQuery.state_to_index for unrecognized types.

julia> using LightQuery

julia> Group(By([1, 3, 2, 4], iseven)) |> collect
2-element Array{Pair{Bool,SubArray{Int64,1,Array{Int64,1},Tuple{UnitRange{Int64}},true}},1}:
 0 => [1, 3]
 1 => [2, 4]

source

LightQuery.key — Function.

key(it)

The key in a key => value pair.

source

LightQuery.value — Function.

value(it)

The value in a key => value pair.

source

LightQuery.Join — Type.

Join(left::By, right::By)

Find all pairs where isequal(left.call(left.it), right.call(right.it)).

julia> using LightQuery

julia> Join(
            By([1, 2, 5, 6], identity),
            By([1, 3, 4, 6], identity)
        ) |>
        collect
6-element Array{Pair{Union{Missing, Int64},Union{Missing, Int64}},1}:
       1 => 1
       2 => missing
 missing => 3
 missing => 4
       5 => missing
       6 => 6

Assumes left and right are both strictly sorted (no repeats). If there are repeats, Group first.

julia> @> [1, 1, 2, 2] |>
        Group(By(_, identity)) |>
        By(_, first) |>
        Join(_, By([1, 2], identity)) |>
        collect
2-element Array{Pair{Pair{Int64,SubArray{Int64,1,Array{Int64,1},Tuple{UnitRange{Int64}},true}},Int64},1}:
 (1 => [1, 1]) => 1
 (2 => [2, 2]) => 2

For other join flavors, combine with when. Make sure to annotate with Length if you know it.

julia> @> Join(
            By([1, 2, 5, 6], identity),
            By([1, 3, 4, 6], identity)
        ) |>
        when(_, @_ !ismissing(_.first)) |>
        Length(_, 4) |>
        collect
4-element Array{Pair{Union{Missing, Int64},Union{Missing, Int64}},1}:
 1 => 1
 2 => missing
 5 => missing
 6 => 6

source

LightQuery.Length — Type.

Length(it, length)

Allow optimizations based on length. Especially useful after Join and before make_columns.

julia> using LightQuery

julia> @> Filter(iseven, 1:4) |>
        Length(_, 2) |>
        collect
2-element Array{Int64,1}:
 2
 4

source

Pivot

LightQuery.item_names — Function.

item_names(it)

Find names of items in it. Used in Peek and make_columns.

julia> using LightQuery

julia> [(a = 1, b = 1.0), (a = 2, b = 2.0)] |>
        item_names
(:a, :b)

If inference cannot detect names, it will use propertynames of the first item. Map Names over it to override this behavior.

julia> [(a = 1,), (a = 2, b = 2.0)] |>
        Peek
|  :a |
| ---:|
|   1 |
|   2 |

julia> @> [(a = 1,), (a = 2, b = 2.0)] |>
        over(_, Names(:a, :b)) |>
        Peek
|  :a |      :b |
| ---:| -------:|
|   1 | missing |
|   2 |     2.0 |

source

LightQuery.rows — Function.

rows(it)

Iterator over rows of a NamedTuple of arrays. Always lazy. Inverse of columns. See Peek for a way to view.

julia> using LightQuery

julia> (a = [1, 2], b = [1.0, 2.0]) |>
        rows |>
        collect
2-element Array{NamedTuple{(:a, :b),Tuple{Int64,Float64}},1}:
 (a = 1, b = 1.0)
 (a = 2, b = 2.0)

source

LightQuery.Peek — Type.

Peek(it; max_rows = 4)

Get a peek of an iterator which returns named tuples. Will show no more than max_rows. Relies on item_names.

julia> using LightQuery

julia> (a = 1:5, b = 5:-1:1) |>
        rows |>
        Peek
Showing 4 of 5 rows
|  :a |  :b |
| ---:| ---:|
|   1 |   5 |
|   2 |   4 |
|   3 |   3 |
|   4 |   2 |

source

LightQuery.columns — Function.

columns(it)

Inverse of rows. Always lazy, see make_columns for eager version.

julia> using LightQuery

julia> (a = [1], b = [1.0]) |>
        rows |>
        columns
(a = [1], b = [1.0])

source

LightQuery.make_columns — Function.

make_columns(it)

Collect into columns. Always eager, see columns for lazy version. Relies on item_names.

julia> using LightQuery

julia> [(a = 1, b = 1.0), (a = 2, b = 2.0)] |>
        make_columns
(a = [1, 2], b = [1.0, 2.0])

source