LightQuery.jl

LightQuery.jl

LightQuery.By

LightQuery.FullJoin

LightQuery.Group

LightQuery.Length

LightQuery.Name

LightQuery.Names

LightQuery.Peek

LightQuery.columns

LightQuery.gather

LightQuery.in_common

LightQuery.make_columns

LightQuery.named_tuple

LightQuery.order

LightQuery.remove

LightQuery.rename

LightQuery.row_type

LightQuery.rows

LightQuery.spread

LightQuery.transform

LightQuery.unzip

LightQuery.@>

LightQuery.@_

LightQuery.By — Type.

By(it, call)

Marks that it has been pre-sorted by the key call. For use with Group or FullJoin.

source

LightQuery.FullJoin — Type.

FullJoin(left::By, right::By)

Find all pairs where isequal(left.call(left.it), right.call(right.it)). Assumes left and right are both strictly sorted (no repeats). If there are repeats, Group first. For other join flavors, combine with Filter.

julia> using LightQuery

julia> FullJoin(
            By([1, 2, 5, 6], identity),
            By([1, 3, 4, 6], identity)
        ) |>
        collect
6-element Array{Pair{Union{Missing, Int64},Union{Missing, Int64}},1}:
       1 => 1
       2 => missing
 missing => 3
 missing => 4
       5 => missing
       6 => 6

julia> @> [1, 1, 2, 2] |>
        Group(By(_, identity)) |>
        By(_, first) |>
        FullJoin(_, By([1, 2], identity)) |>
        collect
2-element Array{Pair{Pair{Int64,SubArray{Int64,1,Array{Int64,1},Tuple{UnitRange{Int64}},true}},Int64},1}:
 (1=>[1, 1]) => 1
 (2=>[2, 2]) => 2

julia> @> FullJoin(
            By([1, 2, 5, 6], identity),
            By([1, 3, 4, 6], identity)
        ) |>
        Filter((@_ !ismissing(_.first)), _) |>
        collect
4-element Array{Pair{Union{Missing, Int64},Union{Missing, Int64}},1}:
 1 => 1
 2 => missing
 5 => missing
 6 => 6

source

LightQuery.Group — Method.

Group(it::By)

Group consecutive keys in it. Requires a presorted object (see By). Relies on the fact that iteration states can be converted to indices; thus, you might have to define LightQuery.state_to_index for unrecognized types.

julia> using LightQuery

julia> Group(By([1, 3, 2, 4], iseven)) |> collect
2-element Array{Pair{Bool,SubArray{Int64,1,Array{Int64,1},Tuple{UnitRange{Int64}},true}},1}:
 0 => [1, 3]
 1 => [2, 4]

source

LightQuery.Length — Type.

Length(it, length)

Allow optimizations based on length. Especially useful before make_columns.

julia> using LightQuery

julia> @> Filter(iseven, 1:4) |>
        Length(_, 2) |>
        collect
2-element Array{Int64,1}:
 2
 4

source

LightQuery.Name — Method.

Name(name)

Force into the type domain. Can also be used as a function to getproperty

julia> using LightQuery

julia> @> (a = 1,) |>
        Name(:a)(_)
1

source

LightQuery.Names — Method.

Names(the_names...)

Force into the type domain. Can be used to as a function to select columns.

julia> using LightQuery

julia> @> (a = 1, b = 1.0) |>
        Names(:a)(_)
(a = 1,)

source

LightQuery.Peek — Method.

Peek(it; max_columns = 7, max_rows = 4)

Get a peek of an iterator which returns named tuples. If inference cannot detect names, it will use the names of the first item.

julia> using LightQuery

julia> [(a = 1, b = 2, c = 3, d = 4, e = 5, f = 6, g = 7, h = 8)] |>
        Peek
Showing 7 of 8 columns
|  :a |  :b |  :c |  :d |  :e |  :f |  :g |
| ---:| ---:| ---:| ---:| ---:| ---:| ---:|
|   1 |   2 |   3 |   4 |   5 |   6 |   7 |

julia> (a = 1:6,) |>
        rows |>
        Peek
Showing 4 of 6 rows
|  :a |
| ---:|
|   1 |
|   2 |
|   3 |
|   4 |

julia> @> (a = 1:2,) |>
        rows |>
        Filter((@_ _.a > 1), _) |>
        Peek
Showing at most 4 rows
|  :a |
| ---:|
|   2 |

source

LightQuery.columns — Method.

columns(it)

Inverse of rows.

julia> using LightQuery

julia> (a = [1], b = [1.0]) |>
        rows |>
        columns
(a = [1], b = [1.0])

source

LightQuery.gather — Method.

gather(it; assignments...)

For each key => value pair in assignments, gather the Names in value into a single key. Inverse of spread.

julia> using LightQuery

julia> @> (a = 1, b = 1.0, c = 1//1) |>
        gather(_, d = Names(:a, :c))
(b = 1.0, d = (a = 1, c = 1//1))

source

LightQuery.in_common — Method.

in_common(it1, it2)

Find the the_names in common between it1 and it2.

julia> using LightQuery

julia> in_common((a = 1, b = 1.0), (a = 2, c = 2//2))
(:a,)

source

LightQuery.make_columns — Method.

make_columns(it)

Collect into columns. See also columns. If inference cannot detect names, it will use the names of the first item. Much more efficient if it has a known length/size.

julia> using LightQuery

julia> [(a = 1, b = 1.0), (a = 2, b = 2.0)] |>
        make_columns
(a = [1, 2], b = [1.0, 2.0])

source

LightQuery.named_tuple — Method.

named_tuple(it)

Coerce to a named_tuple. For performance with working with arbitrary structs, requires propertynames to constant propagate.

julia> using LightQuery

julia> @inline Base.propertynames(p::Pair) = (:first, :second);

julia> named_tuple(:a => 1)
(first = :a, second = 1)

source

LightQuery.order — Method.

order(it, call; keywords...)
order(it, call, condition; keywords...)

Generalized sort. keywords will be passed to sort!; see the documentation there for options. See By for a way to explicitly mark that an object has been sorted. Most performant if call is type stable, if not, consider using a condition to filter.

julia> using LightQuery

julia> order([2, 1], identity)
2-element view(::Array{Int64,1}, [2, 1]) with eltype Int64:
 1
 2

julia> order([1, 2, missing], identity, !ismissing)
2-element view(::Array{Union{Missing, Int64},1}, [1, 2]) with eltype Union{Missing, Int64}:
 1
 2

source

LightQuery.remove — Method.

remove(it, the_names...)

Remove the_names. Inverse of transform.

julia> using LightQuery

julia> @> (a = 1, b = 1.0) |>
        remove(_, :b)
(a = 1,)

source

LightQuery.rename — Method.

rename(it; renames::Name...)

Rename it. Use Name for type stability; constants don't propagate through keyword arguments :(

julia> using LightQuery

julia> @> (a = 1, b = 1.0) |>
        rename(_, c = Name(:a))
(b = 1.0, c = 1)

source

LightQuery.row_type — Method.

row_type(f::File)

Find the type of a row of a CSV.File if it was converted to a named_tuple.

``` julia> using LightQuery

julia> @> "test.csv" |> CSV.File(, allowmissing = :auto) |> rowtype NamedTuple{(:a, :b),T} where T<:Tuple{Int64,Float64}

source

LightQuery.rows — Method.

rows(it)

Iterator over rows of a NamedTuple of arrays. Inverse of columns. See Peek for a way to view rows.

julia> using LightQuery

julia> (a = [1, 2], b = [1.0, 2.0]) |>
        rows |>
        collect
2-element Array{NamedTuple{(:a, :b),Tuple{Int64,Float64}},1}:
 (a = 1, b = 1.0)
 (a = 2, b = 2.0)

source

LightQuery.spread — Method.

spread(it::NamedTuple, the_names...)

Unnest nested it in name. Inverse of gather.

julia> using LightQuery

julia> @> (b = 1.0, d = (a = 1, c = 1//1)) |>
        spread(_, :d)
(b = 1.0, a = 1, c = 1//1)

source

LightQuery.transform — Method.

transform(it; assignments...)

Merge assignments into it.

julia> using LightQuery

julia> transform((a = 1,), b = 1.0)
(a = 1, b = 1.0)

source

LightQuery.unzip — Method.

unzip(it, n)

Unzip an iterator it which returns tuples of length n.

julia> using LightQuery

julia> unzip([(1, 1.0), (2, 2.0)], 2)
([1, 2], [1.0, 2.0])

source

LightQuery.@> — Macro.

macro >(body)

If body is in the form body_ |> tail_, call @_ on tail, and recur on body.

julia> using LightQuery

julia> @> 0 |> _ - 1 |> abs
1

source

LightQuery.@_ — Macro.

macro _(body)

Terser function syntax. The arguments are inside the body; the first argument is _, the second argument is __, etc.

julia> using LightQuery

julia> (@_ _ + 1)(1)
2

julia> map((@_ __ - _), (1, 2), (2, 1))
(1, -1)

source