LightQuery.jl

LightQuery.jl

LightQuery.By

LightQuery.Group

LightQuery.LeftJoin

LightQuery.Name

LightQuery.Names

LightQuery.Peek

LightQuery.columns

LightQuery.gather

LightQuery.in_common

LightQuery.make_columns

LightQuery.named_tuple

LightQuery.order

LightQuery.over

LightQuery.remove

LightQuery.rename

LightQuery.rows

LightQuery.spread

LightQuery.transform

LightQuery.unzip

LightQuery.when

LightQuery.@>

LightQuery.@_

LightQuery.By — Type.

By(it, call)

Marks that it has been pre-sorted by the key call. For use with Group or LeftJoin.

source

LightQuery.Group — Method.

Group(it::By)

Group consecutive keys in it. Requires a presorted object (see By). Relies on the fact that iteration states can be converted to indices; thus, you might have to define LightQuery.state_to_index for unrecognized types.

julia> using LightQuery

julia> Group(By([1, 3, 2, 4], iseven)) |> collect
2-element Array{Pair{Bool,SubArray{Int64,1,Array{Int64,1},Tuple{UnitRange{Int64}},true}},1}:
 0 => [1, 3]
 1 => [2, 4]

julia> Group(By([1], iseven)) |> collect
1-element Array{Pair{Bool,SubArray{Int64,1,Array{Int64,1},Tuple{UnitRange{Int64}},true}},1}:
 0 => [1]

source

LightQuery.LeftJoin — Type.

LeftJoin(left::By, right::By)

For each value in left, look for a value with the same key in right. Requires both to be presorted (see By).

julia> using LightQuery

julia> joined = LeftJoin(
            By([1, 2, 5, 6], identity),
            By([1, 3, 4, 6], identity)
       );

julia> size(joined)
(4,)

julia> length(joined)
4

julia> collect(joined)
4-element Array{Pair{Int64,Union{Missing, Int64}},1}:
 1 => 1
 2 => missing
 5 => missing
 6 => 6

source

LightQuery.Name — Type.

Name(name)

Force into the type domain. Can also be used as a function to getproperty

julia> using LightQuery

julia> Name(:a)((a = 1,))
1

source

LightQuery.Names — Method.

Names(names...)

Force into the type domain. Can be used to as a function to select columns.

julia> using LightQuery

julia> using Test: @inferred

julia> test(x) = Names(:a)(x);

julia> @inferred test((a = 1, b = 1.0))
(a = 1,)

source

LightQuery.Peek — Method.

Peek(it; max_columns = 7, max_rows = 4)

Get a peek of an iterator which returns named tuples. If inference cannot detect names, it will use the names of the first item. Map a Names object over it to help inference.

julia> using LightQuery

julia> [(a = 1, b = 2, c = 3, d = 4, e = 5, f = 6, g = 7, h = 8)] |>
        Peek
Showing 7 of 8 columns
|  :a |  :b |  :c |  :d |  :e |  :f |  :g |
| ---:| ---:| ---:| ---:| ---:| ---:| ---:|
|   1 |   2 |   3 |   4 |   5 |   6 |   7 |

julia> (a = 1:6,) |>
        rows |>
        Peek
Showing 4 of 6 rows
|  :a |
| ---:|
|   1 |
|   2 |
|   3 |
|   4 |

julia> @> (a = 1:2,) |>
        rows |>
        when(_, @_ _.a > 1) |>
        Peek
Showing at most 4 rows
|  :a |
| ---:|
|   2 |

source

LightQuery.columns — Method.

columns(it)

Inverse of rows.

julia> using LightQuery

julia> (a = [1], b = [1.0]) |> rows |> columns
(a = [1], b = [1.0])

source

LightQuery.gather — Method.

gather(it, name, names...)

Gather all the it in names into a single name. Inverse of spread.

julia> using LightQuery

julia> using Test: @inferred

julia> test(x) = gather(x, :d, :a, :c);

julia> @inferred test((a = 1, b = 1.0, c = 1//1))
(b = 1.0, d = (a = 1, c = 1//1))

source

LightQuery.in_common — Method.

in_common(it1, it2)

Find the names in common between it1 and it2.

julia> using LightQuery

julia> in_common((a = 1, b = 1.0), (a = 2, c = 2//2))
(:a,)

source

LightQuery.make_columns — Method.

make_columns(it)

Collect into columns. See also columns. In same cases, will error if inference cannot detect the names. In this case, use the names of the first row.

julia> using LightQuery

julia> using Test: @inferred

julia> [(a = 1, b = 1.0), (a = 2, b = 2.0)] |>
        make_columns
(a = [1, 2], b = [1.0, 2.0])

julia> @> 1:2 |>
        over(_, x -> error()) |>
        make_columns
ERROR: Can't infer names due to inner function error

source

LightQuery.named_tuple — Method.

named_tuple(it)

Coerce to a named_tuple. For performance with working with arbitrary structs, explicitly define inlined propertynames.

julia> using LightQuery

julia> using Test: @inferred

julia> @inline Base.propertynames(p::Pair) = (:first, :second);

julia> @inferred named_tuple(:a => 1)
(first = :a, second = 1)

source

LightQuery.order — Method.

order(it, call; keywords...)
order(it, call, condition; keywords...)

Generalized sort. keywords will be passed to sort!; see the documentation there for options. See By for a way to explicitly mark that an object has been sorted. Most performant if call is type stable, if not, consider using a condition to filter.

julia> using LightQuery

julia> order([2, 1], identity)
2-element view(::Array{Int64,1}, [2, 1]) with eltype Int64:
 1
 2

julia> order([1, 2, missing], identity, !ismissing)
2-element view(::Array{Union{Missing, Int64},1}, [1, 2]) with eltype Union{Missing, Int64}:
 1
 2

source

LightQuery.over — Method.

over(it, call)

Lazy version of map, with the reverse argument order.

julia> using LightQuery

julia> over([1, 2], x -> x + 0.0) |> collect
2-element Array{Float64,1}:
 1.0
 2.0

source

LightQuery.remove — Method.

remove(it, names...)

Remove names. Inverse of transform.

julia> using LightQuery

julia> using Test: @inferred

julia> test(x) = remove(x, :b);

julia> @inferred test((a = 1, b = 1.0))
(a = 1,)

source

LightQuery.rename — Method.

rename(it; renames::Name...)

Rename it. Use Name for type stability; constants don't propagate through keyword arguments :(

julia> using LightQuery

julia> using Test: @inferred

julia> test(x) = rename(x, c = Name(:a));

julia> @inferred test((a = 1, b = 1.0))
(b = 1.0, c = 1)

source

LightQuery.rows — Method.

rows(it)

Iterator over rows of a NamedTuple of arrays. Inverse of columns. See Peek for a way to view rows.

julia> using LightQuery

julia> (a = [1, 2], b = [1.0, 2.0]) |> rows |> Peek
|  :a |  :b |
| ---:| ---:|
|   1 | 1.0 |
|   2 | 2.0 |

source

LightQuery.spread — Method.

spread(it::NamedTuple, name)

Unnest nested it in name. Inverse of gather.

julia> using LightQuery

julia> using Test: @inferred

julia> test(x) = spread(x, :d);

julia> @inferred test((b = 1.0, d = (a = 1, c = 1//1)))
(b = 1.0, a = 1, c = 1//1)

source

LightQuery.transform — Method.

transform(it; assignments...)

Merge assignments into it.

julia> using LightQuery

julia> using Test: @inferred

julia> @inferred transform((a = 1,), b = 1.0)
(a = 1, b = 1.0)

source

LightQuery.unzip — Method.

unzip(it, n)

Unzip an iterator it which returns tuples of length n.

julia> using LightQuery

julia> using Test: @inferred

julia> f(x) = (x, x + 0.0);

julia> test(x) = unzip(x, 2);

julia> test(over([1, missing], f))
(Union{Missing, Int64}[1, missing], Union{Missing, Float64}[1.0, missing])

julia> @inferred test(zip([1], [1.0]))
([1], [1.0])

julia> @inferred test([(1, 1.0)])
([1], [1.0])

julia> test(over(when([1, missing], x -> true), f))
(Union{Missing, Int64}[1, missing], Union{Missing, Float64}[1.0, missing])

julia> zip([1], [1, 2])
ERROR: ArgumentError: All arrays passed to zip must have the same size

source

LightQuery.when — Method.

when(it, call)

Lazy version of filter, with the reverse argument order.

julia> using LightQuery

julia> when([1, 2], x -> x > 1) |> collect
1-element Array{Int64,1}:
 2

source

LightQuery.@> — Macro.

macro >(body)

If body is in the form body_ |> tail_, call @_ on tail, and recur on body.

julia> using LightQuery

julia> @> 0 |> _ - 1 |> abs
1

source

LightQuery.@_ — Macro.

macro _(body)

Terser function syntax. The arguments are inside the body; the first argument is _, the second argument is __, etc.

julia> using LightQuery

julia> (@_ _ + 1)(1)
2

julia> map((@_ __ - _), (1, 2), (2, 1))
(1, -1)

source

Tutorial

I started following the tutorial here here, but got side-tracked by a data cleaning task. You get to see the results, hooray. I've included the flights data in test folder of this package.

I've reexported CSV from the CSV package for convenient IO. We can process the data as we are reading in the file itself! This is because CSV.File returns an iterator.

I do a lot of steps here so let's break them down:

First, I bring in some dates functions, and make them work with missing data.

In the CSV.File step, I mark that missing is denoted by "NA" in the file. I also specify that I want strings to come in as categorical data.

Then I transform each row. First, I convert each row to a named tuple. Then, I combine the scheduled arrival and departure times into a single date-time. The bit at the end about divrem basically means that in a number like 530, hours are in the hundreds place and minutes are in the rest (e.g. 5:30). Then I rename to get rid of abbreviations (you'll figure out if you haven't already that I'm a big fan of whole words). Then, I select just the columns we want. I've taken the liberty of ignoring calculated data, that is, data that we can calculate later on. Also, mapping a Names object with over helps Julia out. Julia can't infer the names of the rows without this step. This is the because we have two or more type-unstable columns (e.g. departure delay and arrival delay both might be missing).

julia> using LightQuery

julia> using Dates: DateTime

julia> import Dates: Minute

julia> Minute(::Missing) = missing;

julia> using Unitful: mi

julia> flight_columns =
            @> CSV.File("flights.csv", missingstring = "NA", categorical = true) |>
            over(_, @_ @> named_tuple(_) |>
                transform(_,
                    departure_time = DateTime(_.year, _.month, _.day,
                        divrem(_.sched_dep_time, 100)...),
                    departure_delay = Minute(_.dep_delay),
                    arrival_time = DateTime(_.year, _.month, _.day,
                        divrem(_.sched_arr_time, 100)...),
                    arrival_delay = Minute(_.arr_delay),
                    distance = _.distance * mi
                ) |>
                rename(_,
                    tail_number = Name(:tailnum),
                    destination = Name(:dest)
                )
            ) |>
            over(_, Names(:departure_time, :departure_delay, :arrival_time,
                :arrival_delay, :carrier, :flight, :tail_number, :origin,
                :destination, :distance)
            ) |>
            make_columns;

Now that we have our data into Julia, let's have a look see. We'll start by converting it back to columns, and then taking a Peek. I've taken the liberty of increasing the number of visible columns to 10. The default max Peek size is 7 columns by 4 rows.

julia> flights = rows(flight_columns);

julia> Peek(flights, max_columns = 10)
Showing 4 of 336776 rows
|     :departure_time | :departure_delay |       :arrival_time | :arrival_delay | :carrier | :flight | :tail_number | :origin | :destination | :distance |
| -------------------:| ----------------:| -------------------:| --------------:| --------:| -------:| ------------:| -------:| ------------:| ---------:|
| 2013-01-01T05:15:00 |        2 minutes | 2013-01-01T08:19:00 |     11 minutes |       UA |    1545 |       N14228 |     EWR |          IAH |   1400 mi |
| 2013-01-01T05:29:00 |        4 minutes | 2013-01-01T08:30:00 |     20 minutes |       UA |    1714 |       N24211 |     LGA |          IAH |   1416 mi |
| 2013-01-01T05:40:00 |        2 minutes | 2013-01-01T08:50:00 |     33 minutes |       AA |    1141 |       N619AA |     JFK |          MIA |   1089 mi |
| 2013-01-01T05:45:00 |        -1 minute | 2013-01-01T10:22:00 |    -18 minutes |       B6 |     725 |       N804JB |     JFK |          BQN |   1576 mi |

There's one more cleaning step we can make. Note that distance is a calculated field. That is, the distance between two locations is always going to be the same. How can we get a clean dataset which only contains the distances between two airports?

Let's start out by grouping our data by path. Before you group, you MUST sort. Otherwise, you will get incorrect results. Of course, if we have pre-sorted data, no need.

julia> by_path =
            @> flights |>
            order(_, Names(:origin, :destination)) |>
            Group(By(_, Names(:origin, :destination))) |>
            collect;

I encourage you to collect after Grouping. This will not use much additional data, it will only store the keys locations of the groups. Grouping is a little different from dplyr; each group is a pair from key to sub-data-frame:

julia> first_group = first(by_path);

julia> first_group.first
(origin = CategoricalArrays.CategoricalString{UInt32} "EWR", destination = CategoricalArrays.CategoricalString{UInt32} "ALB")

julia> Peek(first_group.second, max_columns = 10)
Showing 4 of 439 rows
|     :departure_time | :departure_delay |       :arrival_time | :arrival_delay | :carrier | :flight | :tail_number | :origin | :destination | :distance |
| -------------------:| ----------------:| -------------------:| --------------:| --------:| -------:| ------------:| -------:| ------------:| ---------:|
| 2013-01-01T13:17:00 |       -2 minutes | 2013-01-01T14:23:00 |    -10 minutes |       EV |    4112 |       N13538 |     EWR |          ALB |    143 mi |
| 2013-01-01T16:21:00 |       34 minutes | 2013-01-01T17:24:00 |     40 minutes |       EV |    3260 |       N19554 |     EWR |          ALB |    143 mi |
| 2013-01-01T20:04:00 |       52 minutes | 2013-01-01T21:12:00 |     44 minutes |       EV |    4170 |       N12540 |     EWR |          ALB |    143 mi |
| 2013-01-02T13:27:00 |        5 minutes | 2013-01-02T14:33:00 |    -14 minutes |       EV |    4316 |       N14153 |     EWR |          ALB |    143 mi |

So there are 439 flights between Newark (EWR lol) and Albany. The distances are the same, so we really only need the first one. So here's what we do: Calling make_columns and then rows will collect out data (the smart way).

julia> paths =
        @> by_path |>
        over(_, @_ first(_.second)) |>
        over(_, Names(:origin, :destination, :distance)) |>
        make_columns |>
        rows;

julia> Peek(paths)
Showing 4 of 224 rows
| :origin | :destination | :distance |
| -------:| ------------:| ---------:|
|     EWR |          ALB |    143 mi |
|     EWR |          ANC |   3370 mi |
|     EWR |          ATL |    746 mi |
|     EWR |          AUS |   1504 mi |

Ok, now let's do the whole damn thing in reverse, just for fun. How? We need to join back into the original data. Our data that was grouped by path is sorted by the first item, and our path data is sorted by :origin and :destination. Note there are no repeats: we've pregrouped our data. This is important for a left join.

julia> joined = LeftJoin(
            By(by_path, first),
            By(paths, Names(:origin, :destination))
        );

A join will a row on the left and a row on the right. And the row on the left is a group, so it's also got a key and a value. Oy.

julia> pair = first(joined);

julia> pair.first.first
(origin = CategoricalArrays.CategoricalString{UInt32} "EWR", destination = CategoricalArrays.CategoricalString{UInt32} "ALB")

julia> Peek(pair.first.second, max_columns = 10)
Showing 4 of 439 rows
|     :departure_time | :departure_delay |       :arrival_time | :arrival_delay | :carrier | :flight | :tail_number | :origin | :destination | :distance |
| -------------------:| ----------------:| -------------------:| --------------:| --------:| -------:| ------------:| -------:| ------------:| ---------:|
| 2013-01-01T13:17:00 |       -2 minutes | 2013-01-01T14:23:00 |    -10 minutes |       EV |    4112 |       N13538 |     EWR |          ALB |    143 mi |
| 2013-01-01T16:21:00 |       34 minutes | 2013-01-01T17:24:00 |     40 minutes |       EV |    3260 |       N19554 |     EWR |          ALB |    143 mi |
| 2013-01-01T20:04:00 |       52 minutes | 2013-01-01T21:12:00 |     44 minutes |       EV |    4170 |       N12540 |     EWR |          ALB |    143 mi |
| 2013-01-02T13:27:00 |        5 minutes | 2013-01-02T14:33:00 |    -14 minutes |       EV |    4316 |       N14153 |     EWR |          ALB |    143 mi |

julia> pair.second
(origin = CategoricalArrays.CategoricalString{UInt32} "EWR", destination = CategoricalArrays.CategoricalString{UInt32} "ALB", distance = 143 mi)

How are we gonna get this all back into a flat data-frame? We need to make use of flatten (which is reexported from Base).

julia> @> joined |>
        over(_, pair -> over(pair.first.second, @_ transform(_, distance_again = pair.second.distance))) |>
        flatten |>
        Peek(_, max_columns = 11)

Showing at most 4 rows
|     :departure_time | :departure_delay |       :arrival_time | :arrival_delay | :carrier | :flight | :tail_number | :origin | :destination | :distance | :distance_again |
| -------------------:| ----------------:| -------------------:| --------------:| --------:| -------:| ------------:| -------:| ------------:| ---------:| ---------------:|
| 2013-01-01T13:17:00 |       -2 minutes | 2013-01-01T14:23:00 |    -10 minutes |       EV |    4112 |       N13538 |     EWR |          ALB |    143 mi |          143 mi |
| 2013-01-01T16:21:00 |       34 minutes | 2013-01-01T17:24:00 |     40 minutes |       EV |    3260 |       N19554 |     EWR |          ALB |    143 mi |          143 mi |
| 2013-01-01T20:04:00 |       52 minutes | 2013-01-01T21:12:00 |     44 minutes |       EV |    4170 |       N12540 |     EWR |          ALB |    143 mi |          143 mi |
| 2013-01-02T13:27:00 |        5 minutes | 2013-01-02T14:33:00 |    -14 minutes |       EV |    4316 |       N14153 |     EWR |          ALB |    143 mi |          143 mi |

Look! The distances match. Hooray!

Are you exhaused? I'll admit, I am. Hope you learned something.