LightQuery.jl

LightQuery.jl

For performance with working with arbitrary structs, explicitly define public propertynames.

For an example of how to use this package, see the demo below, which follows the tutorial here.

julia> using LightQuery

julia> import CSV

julia> flights =
          @> CSV.read("flights.csv", missingstring = "NA") |>
          named_tuple |>
          remove(_, Symbol(""));

julia> @> flights |>
          rows |>
          when(_, @_ _.month == 1 && _.day == 1);

julia> @> flights |>
          rows |>
          when(_, @_ _.month == 1 || _.month == 2);

julia> @> flights |>
          rows |>
          _[1:10];

julia> @> flights |>
          rows |>
          order_by(_, select(:arr_delay));

julia> @> flights |>
          select(_, :year, :month, :day);

julia> @> flights |>
          remove(_, :year, :month, :day);

julia> @> flights |>
          rename(_, tail_num = :tailnum);

julia> @> flights |>
          transform(_,
            gain = _.arr_delay .- _.dep_delay,
            speed = _.distance ./ _.air_time .* 60
          ) |>
          transform(_,
            gain_per_hour = _.gain ./ (_.air_time / 60)
          );

julia> using Statistics: mean;

julia> mean(skipmissing(flights.dep_delay));

julia> @> flights |>
          rows |>
          order_by(_, select(:tailnum)) |>
          Group |>
          over(_,  @_ transform(_.first,
                    count = length(_.second),
                    distance = column(_.second, :distance) |> skipmissing |> mean,
                    delay = column(_.second, :arr_delay) |> skipmissing |> mean
          )) |>
          columns(_, :tailnum, :count, :distance, :delay);

julia> per_day =
          @> flights |>
          rows |>
          order_by(_, select(:year, :month, :day)) |>
          Group |>
          over(_, @_ transform(_.first, flights = length(_.second))) |>
          columns(_, :year, :month, :day, :flights);

julia> per_month =
          @> per_day|>
          rows |>
          By(_, select(:year, :month)) |>
          Group |>
          over(_, @_ transform(_.first, flights = sum(column(_.second, :flights)))) |>
          columns(_, :year, :month, :flights);

julia> per_year =
          @> per_month |>
          rows |>
          By(_, select(:year)) |>
          Group |>
          over(_, @_ transform(_.first, flights = sum(column(_.second, :flights)))) |>
          columns(_, :year, :flights);

julia> @> flights |>
          select(_, :year, :month, :day, :arr_delay, :dep_delay) |>
          rows |>
          order_by(_, select(:year, :month, :day)) |>
          Group |>
          over(_, @_ transform(_.first,
                    arr = column(_.second, :arr_delay) |> skipmissing |> mean,
                    dep = column(_.second, :dep_delay) |> skipmissing |> mean,
          )) |>
          when(_, @_ _.arr > 30 || _.dep > 30) |>
          columns(_, :year, :month, :day, :arr, :dep);

LightQuery.By

LightQuery.Group

LightQuery.LeftJoin

LightQuery.Name

LightQuery.column

LightQuery.columns

LightQuery.gather

LightQuery.named_tuple

LightQuery.order_by

LightQuery.over

LightQuery.pretty

LightQuery.remove

LightQuery.rename

LightQuery.rows

LightQuery.select

LightQuery.spread

LightQuery.transform

LightQuery.unzip

LightQuery.when

LightQuery.@>

LightQuery.@_

LightQuery.By — Type.

By(it, f)

Marks that it has been pre-sorted by the key f. For use with Group or LeftJoin.

source

LightQuery.Group — Method.

group(b::By)

Group consecutive keys in b. Requires a presorted object (see By).

julia> using LightQuery

julia> Group(By([1, 3, 2, 4], iseven)) |> collect
2-element Array{Pair{Bool,SubArray{Int64,1,Array{Int64,1},Tuple{UnitRange{Int64}},true}},1}:
 0 => [1, 3]
 1 => [2, 4]

source

LightQuery.LeftJoin — Type.

LeftJoin(left::By, right::By)

For each value in left, look for a value with the same key in right. Requires both to be presorted (see By).

julia> using LightQuery

julia> LeftJoin(
            By([1, 2, 5, 6], identity),
            By([1, 3, 4, 6], identity)
       ) |> collect
4-element Array{Pair{Int64,Union{Missing, Int64}},1}:
 1 => 1
 2 => missing
 5 => missing
 6 => 6

source

LightQuery.Name — Type.

Name(x)

Force into the type domain. Can also be used as a function.

julia> using LightQuery

julia> Name(:a)((a = 1,))
1

source

LightQuery.column — Method.

column(it, name)

Access just one column.

julia> using LightQuery

julia> using Test: @inferred

julia> test(x) = column(x, :a);

julia> @inferred test([(a = 1, b = 1.0)]) |> collect
1-element Array{Int64,1}:
 1

julia> @inferred test(rows((a = [1, 2], b = [2, 1])))
2-element Array{Int64,1}:
 1
 2

source

LightQuery.columns — Method.

columns(it, names...)

Collect into columns. Inverse of rows.

julia> using LightQuery

julia> using Test: @inferred

julia> test(x) = columns(x, :b, :a);

julia> @inferred test([(a = 1, b = 1.0)])
(b = [1.0], a = [1])

julia> @inferred test(rows((a = [1, 2], b = [2, 1])))
(b = [2, 1], a = [1, 2])

source

LightQuery.gather — Method.

gather(data, name, names...)

Gather all the data in names into a single name. Inverse of spread.

julia> using LightQuery

julia> using Test: @inferred

julia> test(x) = gather(x, :d, :a, :c);

julia> @inferred test((a = 1, b = 2.0, c = "c"))
(b = 2.0, d = (a = 1, c = "c"))

source

LightQuery.named_tuple — Method.

named_tuple(x)

Coerce to a named_tuple.

julia> using LightQuery

julia> named_tuple(:a => 1)
(first = :a, second = 1)

source

LightQuery.order_by — Method.

order_by(it, f)

Generalized sort. Will return a By object.

julia> using LightQuery

julia> order_by(["b", "a"], identity).it
2-element view(::Array{String,1}, [2, 1]) with eltype String:
 "a"
 "b"

source

LightQuery.over — Method.

over(it, f)

Hackable reversed version of Base.Generator.

julia> using LightQuery

julia> over([1, 2], x -> x + 1) |> collect
2-element Array{Int64,1}:
 2
 3

source

LightQuery.pretty — Method.

pretty(x)

Pretty display.

julia> using LightQuery

julia> pretty((a = [1, 2], b = [1.0, 2.0]))
2×2 DataFrames.DataFrame
│ Row │ a     │ b       │
│     │ Int64 │ Float64 │
├─────┼───────┼─────────┤
│ 1   │ 1     │ 1.0     │
│ 2   │ 2     │ 2.0     │

source

LightQuery.remove — Method.

remove(data, names...)

Remove names. Inverse of transform.

julia> using LightQuery

julia> using Test: @inferred

julia> test(x) = remove(x, :b);

julia> @inferred test((a = 1, b = 2.0))
(a = 1,)

source

LightQuery.rename — Method.

rename(data; renames...)

Rename data. Currently unstable without Name

julia> using LightQuery

julia> using Test: @inferred

julia> test(x) = rename(x, c = Name(:a));

julia> @inferred test((a = 1, b = 2.0))
(b = 2.0, c = 1)

source

LightQuery.rows — Method.

rows(n::NamedTuple)

Iterator over rows of a NamedTuple of names. Inverse of columns.

julia> using LightQuery

julia> using Test: @inferred

julia> @inferred first(rows((a = [1, 2], b = [2, 1])))
(a = 1, b = 2)

source

LightQuery.select — Method.

select(data::NamedTuple, names...)

Select names.

select(ss::Symbol...)

Curried form.

julia> using LightQuery

julia> using Test: @inferred

julia> test(x) = select(x, :a);

julia> @inferred test((a = 1, b = 2.0))
(a = 1,)

julia> test(x) = select(:a)(x);

julia> @inferred test((a = 1, b = 2.0))
(a = 1,)

source

LightQuery.spread — Method.

spread(data::NamedTuple, name)

Unnest nested data in name. Inverse of gather.

julia> using LightQuery

julia> using Test: @inferred

julia> test(x) = spread(x, :d);

julia> @inferred test((b = 2.0, d = (a = 1, c = "c")))
(b = 2.0, a = 1, c = "c")

source

LightQuery.transform — Method.

transform(data; assignments...)

Apply the functions in assignments to data, assign to the corresponding keys, and merge back in the original.

julia> using LightQuery

julia> using Test: @inferred

julia> @inferred transform((a = 1, b = 2.0), c = "3")
(a = 1, b = 2.0, c = "3")

source

LightQuery.unzip — Method.

unzip(it, n)

Unzip an iterator it which returns tuples of length n.

julia> using LightQuery

julia> f(x) = (x, x + 1.0);

julia> unzip(over([1], f), 2)
([1], [2.0])

julia> unzip(over([1, missing], f), 2)
(Union{Missing, Int64}[1, missing], Union{Missing, Float64}[2.0, missing])

julia> unzip(zip([1], [1.0]), 2)
([1], [1.0])

julia> unzip([(1, 1.0)], 2)
([1], [1.0])

julia> unzip(over(when([1, missing, 2], x -> ismissing(x) || x > 1), f), 2)
(Union{Missing, Int64}[missing, 2], Union{Missing, Float64}[missing, 3.0])

source

LightQuery.when — Method.

when(it, f)

Hackable reversed version of Base.Iterators.Filter.

julia> using LightQuery

julia> when([1, 2], x -> x > 1) |> collect
1-element Array{Int64,1}:
 2

source

LightQuery.@> — Macro.

macro >(body)

If body is in the form body_ |> tail_, call @_ on tail, and recur on body.

julia> using LightQuery

julia> @> 0 |> _ - 1 |> abs
1

source

LightQuery.@_ — Macro.

macro _(body)

Terser function syntax. The arguments are inside the body; the first argument is _, the second argument is __, etc.

julia> using LightQuery

julia> 1 |> @_(_ + 1)
2

julia> map(@_(__ - _), (1, 2), (2, 1))
(1, -1)

source