LightQuery.jl

LightQuery.jl

Introduction to LightQuery

Follows the tutorial here. I'll make heavy use of the chaining macro @> and lazy calling macro @_ included in this package. The syntax here is in most cases more verbose but also more flexible than dplyr. Fortunately, programming with LightQuery is much easier, so define shortcuts.

Data

Use Peek to view iterators of NamedTuples. The data is stored column-wise; rows is a lazy view.

julia> using LightQuery

julia> flights =
        @> "flights.csv" |>
        CSV.read(_, allowmissing = :auto) |>
        named_tuple |>
        rows;

julia> Peek(flights)
Showing 4 of 336776 rows
| :year | :month | :day | :dep_time | :sched_dep_time | :dep_delay | :arr_time | :sched_arr_time | :arr_delay | :carrier | :flight | :tailnum | :origin | :dest | :air_time | :distance | :hour | :minute |          :time_hour |
| -----:| ------:| ----:| ---------:| ---------------:| ----------:| ---------:| ---------------:| ----------:| --------:| -------:| --------:| -------:| -----:| ---------:| ---------:| -----:| -------:| -------------------:|
|  2013 |      1 |    1 |       517 |             515 |          2 |       830 |             819 |         11 |       UA |    1545 |   N14228 |     EWR |   IAH |       227 |      1400 |     5 |      15 | 2013-01-01 05:00:00 |
|  2013 |      1 |    1 |       533 |             529 |          4 |       850 |             830 |         20 |       UA |    1714 |   N24211 |     LGA |   IAH |       227 |      1416 |     5 |      29 | 2013-01-01 05:00:00 |
|  2013 |      1 |    1 |       542 |             540 |          2 |       923 |             850 |         33 |       AA |    1141 |   N619AA |     JFK |   MIA |       160 |      1089 |     5 |      40 | 2013-01-01 05:00:00 |
|  2013 |      1 |    1 |       544 |             545 |         -1 |      1004 |            1022 |        -18 |       B6 |     725 |   N804JB |     JFK |   BQN |       183 |      1576 |     5 |      45 | 2013-01-01 05:00:00 |

Single table verbs

Filter rows

julia> @> flights |>
        when(_, @_ _.month == 1 && _.day == 1) |>
        Peek
Showing at most 4 rows
| :year | :month | :day | :dep_time | :sched_dep_time | :dep_delay | :arr_time | :sched_arr_time | :arr_delay | :carrier | :flight | :tailnum | :origin | :dest | :air_time | :distance | :hour | :minute |          :time_hour |
| -----:| ------:| ----:| ---------:| ---------------:| ----------:| ---------:| ---------------:| ----------:| --------:| -------:| --------:| -------:| -----:| ---------:| ---------:| -----:| -------:| -------------------:|
|  2013 |      1 |    1 |       517 |             515 |          2 |       830 |             819 |         11 |       UA |    1545 |   N14228 |     EWR |   IAH |       227 |      1400 |     5 |      15 | 2013-01-01 05:00:00 |
|  2013 |      1 |    1 |       533 |             529 |          4 |       850 |             830 |         20 |       UA |    1714 |   N24211 |     LGA |   IAH |       227 |      1416 |     5 |      29 | 2013-01-01 05:00:00 |
|  2013 |      1 |    1 |       542 |             540 |          2 |       923 |             850 |         33 |       AA |    1141 |   N619AA |     JFK |   MIA |       160 |      1089 |     5 |      40 | 2013-01-01 05:00:00 |
|  2013 |      1 |    1 |       544 |             545 |         -1 |      1004 |            1022 |        -18 |       B6 |     725 |   N804JB |     JFK |   BQN |       183 |      1576 |     5 |      45 | 2013-01-01 05:00:00 |

In some cases, a column-wise filter might be more efficient.

julia> @> flights |>
        columns |>
        (_.month .== 1 .& _.day .== 1) |>
        view(flights, _) |>
        Peek
Showing 4 of 13936 rows
| :year | :month | :day | :dep_time | :sched_dep_time | :dep_delay | :arr_time | :sched_arr_time | :arr_delay | :carrier | :flight | :tailnum | :origin | :dest | :air_time | :distance | :hour | :minute |          :time_hour |
| -----:| ------:| ----:| ---------:| ---------------:| ----------:| ---------:| ---------------:| ----------:| --------:| -------:| --------:| -------:| -----:| ---------:| ---------:| -----:| -------:| -------------------:|
|  2013 |      1 |    1 |       517 |             515 |          2 |       830 |             819 |         11 |       UA |    1545 |   N14228 |     EWR |   IAH |       227 |      1400 |     5 |      15 | 2013-01-01 05:00:00 |
|  2013 |      1 |    1 |       533 |             529 |          4 |       850 |             830 |         20 |       UA |    1714 |   N24211 |     LGA |   IAH |       227 |      1416 |     5 |      29 | 2013-01-01 05:00:00 |
|  2013 |      1 |    1 |       542 |             540 |          2 |       923 |             850 |         33 |       AA |    1141 |   N619AA |     JFK |   MIA |       160 |      1089 |     5 |      40 | 2013-01-01 05:00:00 |
|  2013 |      1 |    1 |       544 |             545 |         -1 |      1004 |            1022 |        -18 |       B6 |     725 |   N804JB |     JFK |   BQN |       183 |      1576 |     5 |      45 | 2013-01-01 05:00:00 |

Order rows

julia> @> flights |>
        order(_, Names(:year, :month, :day)) |>
        Peek
Showing 4 of 336776 rows
| :year | :month | :day | :dep_time | :sched_dep_time | :dep_delay | :arr_time | :sched_arr_time | :arr_delay | :carrier | :flight | :tailnum | :origin | :dest | :air_time | :distance | :hour | :minute |          :time_hour |
| -----:| ------:| ----:| ---------:| ---------------:| ----------:| ---------:| ---------------:| ----------:| --------:| -------:| --------:| -------:| -----:| ---------:| ---------:| -----:| -------:| -------------------:|
|  2013 |      1 |    1 |       517 |             515 |          2 |       830 |             819 |         11 |       UA |    1545 |   N14228 |     EWR |   IAH |       227 |      1400 |     5 |      15 | 2013-01-01 05:00:00 |
|  2013 |      1 |    1 |       533 |             529 |          4 |       850 |             830 |         20 |       UA |    1714 |   N24211 |     LGA |   IAH |       227 |      1416 |     5 |      29 | 2013-01-01 05:00:00 |
|  2013 |      1 |    1 |       542 |             540 |          2 |       923 |             850 |         33 |       AA |    1141 |   N619AA |     JFK |   MIA |       160 |      1089 |     5 |      40 | 2013-01-01 05:00:00 |
|  2013 |      1 |    1 |       544 |             545 |         -1 |      1004 |            1022 |        -18 |       B6 |     725 |   N804JB |     JFK |   BQN |       183 |      1576 |     5 |      45 | 2013-01-01 05:00:00 |

For performance, filter out instabilities.

julia> @> flights |>
        order(_, Names(:arr_delay), (@_ !ismissing(_.arr_delay)), rev = true) |>
        Peek
Showing 4 of 327346 rows
| :year | :month | :day | :dep_time | :sched_dep_time | :dep_delay | :arr_time | :sched_arr_time | :arr_delay | :carrier | :flight | :tailnum | :origin | :dest | :air_time | :distance | :hour | :minute |          :time_hour |
| -----:| ------:| ----:| ---------:| ---------------:| ----------:| ---------:| ---------------:| ----------:| --------:| -------:| --------:| -------:| -----:| ---------:| ---------:| -----:| -------:| -------------------:|
|  2013 |      1 |    9 |       641 |             900 |       1301 |      1242 |            1530 |       1272 |       HA |      51 |   N384HA |     JFK |   HNL |       640 |      4983 |     9 |       0 | 2013-01-09 09:00:00 |
|  2013 |      6 |   15 |      1432 |            1935 |       1137 |      1607 |            2120 |       1127 |       MQ |    3535 |   N504MQ |     JFK |   CMH |        74 |       483 |    19 |      35 | 2013-06-15 19:00:00 |
|  2013 |      1 |   10 |      1121 |            1635 |       1126 |      1239 |            1810 |       1109 |       MQ |    3695 |   N517MQ |     EWR |   ORD |       111 |       719 |    16 |      35 | 2013-01-10 16:00:00 |
|  2013 |      9 |   20 |      1139 |            1845 |       1014 |      1457 |            2210 |       1007 |       AA |     177 |   N338AA |     JFK |   SFO |       354 |      2586 |    18 |      45 | 2013-09-20 18:00:00 |

Select columns

Lazily convert to columns then back to rows for columnwise operations.

julia> @> flights |>
        columns |>
        Names(:year, :month, :day)(_) |>
        rows |>
        Peek
Showing 4 of 336776 rows
| :year | :month | :day |
| -----:| ------:| ----:|
|  2013 |      1 |    1 |
|  2013 |      1 |    1 |
|  2013 |      1 |    1 |
|  2013 |      1 |    1 |

julia> @> flights |>
        columns |>
        remove(_, :year, :month, :day) |>
        rows |>
        Peek
Showing 4 of 336776 rows
| :dep_time | :sched_dep_time | :dep_delay | :arr_time | :sched_arr_time | :arr_delay | :carrier | :flight | :tailnum | :origin | :dest | :air_time | :distance | :hour | :minute |          :time_hour |
| ---------:| ---------------:| ----------:| ---------:| ---------------:| ----------:| --------:| -------:| --------:| -------:| -----:| ---------:| ---------:| -----:| -------:| -------------------:|
|       517 |             515 |          2 |       830 |             819 |         11 |       UA |    1545 |   N14228 |     EWR |   IAH |       227 |      1400 |     5 |      15 | 2013-01-01 05:00:00 |
|       533 |             529 |          4 |       850 |             830 |         20 |       UA |    1714 |   N24211 |     LGA |   IAH |       227 |      1416 |     5 |      29 | 2013-01-01 05:00:00 |
|       542 |             540 |          2 |       923 |             850 |         33 |       AA |    1141 |   N619AA |     JFK |   MIA |       160 |      1089 |     5 |      40 | 2013-01-01 05:00:00 |
|       544 |             545 |         -1 |      1004 |            1022 |        -18 |       B6 |     725 |   N804JB |     JFK |   BQN |       183 |      1576 |     5 |      45 | 2013-01-01 05:00:00 |

julia> @> flights |>
        columns |>
        rename(_, tail_num = Name(:tailnum)) |>
        rows |>
        Peek
Showing 4 of 336776 rows
| :year | :month | :day | :dep_time | :sched_dep_time | :dep_delay | :arr_time | :sched_arr_time | :arr_delay | :carrier | :flight | :origin | :dest | :air_time | :distance | :hour | :minute |          :time_hour | :tail_num |
| -----:| ------:| ----:| ---------:| ---------------:| ----------:| ---------:| ---------------:| ----------:| --------:| -------:| -------:| -----:| ---------:| ---------:| -----:| -------:| -------------------:| ---------:|
|  2013 |      1 |    1 |       517 |             515 |          2 |       830 |             819 |         11 |       UA |    1545 |     EWR |   IAH |       227 |      1400 |     5 |      15 | 2013-01-01 05:00:00 |    N14228 |
|  2013 |      1 |    1 |       533 |             529 |          4 |       850 |             830 |         20 |       UA |    1714 |     LGA |   IAH |       227 |      1416 |     5 |      29 | 2013-01-01 05:00:00 |    N24211 |
|  2013 |      1 |    1 |       542 |             540 |          2 |       923 |             850 |         33 |       AA |    1141 |     JFK |   MIA |       160 |      1089 |     5 |      40 | 2013-01-01 05:00:00 |    N619AA |
|  2013 |      1 |    1 |       544 |             545 |         -1 |      1004 |            1022 |        -18 |       B6 |     725 |     JFK |   BQN |       183 |      1576 |     5 |      45 | 2013-01-01 05:00:00 |    N804JB |

Add new columns

julia> @> flights |>
        columns |>
        transform(_,
            gain = _.arr_delay .- _.dep_delay,
            speed = _.distance ./ _.air_time * 60
        ) |>
        rows |>
        Peek
Showing 4 of 336776 rows
| :year | :month | :day | :dep_time | :sched_dep_time | :dep_delay | :arr_time | :sched_arr_time | :arr_delay | :carrier | :flight | :tailnum | :origin | :dest | :air_time | :distance | :hour | :minute |          :time_hour | :gain |             :speed |
| -----:| ------:| ----:| ---------:| ---------------:| ----------:| ---------:| ---------------:| ----------:| --------:| -------:| --------:| -------:| -----:| ---------:| ---------:| -----:| -------:| -------------------:| -----:| ------------------:|
|  2013 |      1 |    1 |       517 |             515 |          2 |       830 |             819 |         11 |       UA |    1545 |   N14228 |     EWR |   IAH |       227 |      1400 |     5 |      15 | 2013-01-01 05:00:00 |     9 | 370.04405286343615 |
|  2013 |      1 |    1 |       533 |             529 |          4 |       850 |             830 |         20 |       UA |    1714 |   N24211 |     LGA |   IAH |       227 |      1416 |     5 |      29 | 2013-01-01 05:00:00 |    16 |  374.2731277533039 |
|  2013 |      1 |    1 |       542 |             540 |          2 |       923 |             850 |         33 |       AA |    1141 |   N619AA |     JFK |   MIA |       160 |      1089 |     5 |      40 | 2013-01-01 05:00:00 |    31 |            408.375 |
|  2013 |      1 |    1 |       544 |             545 |         -1 |      1004 |            1022 |        -18 |       B6 |     725 |   N804JB |     JFK |   BQN |       183 |      1576 |     5 |      45 | 2013-01-01 05:00:00 |   -17 |  516.7213114754098 |

Grouped operations

You can only group sorted data. Each group is a pair, from key to value (sub-table).

julia> using Statistics: mean

julia> @> flights |>
        order(_, Names(:tailnum), (@_ !ismissing(_.tailnum))) |>
        Group(By(_, Names(:tailnum))) |>
        over(_, @_ transform(key(_),
            count = length(value(_)),
            dist = (@> columns(value(_)).distance |> skipmissing |> mean),
            delay = (@> columns(value(_)).arr_delay |> skipmissing |> mean)
        )) |>
        Peek
Showing at most 4 rows
| :tailnum | :count |             :dist |             :delay |
| --------:| ------:| -----------------:| ------------------:|
|   D942DN |      4 |             854.5 |               31.5 |
|   N0EGMQ |    371 |  676.188679245283 |  9.982954545454545 |
|   N10156 |    153 | 757.9477124183006 | 12.717241379310344 |
|   N102UW |     48 |           535.875 |             2.9375 |

Two table verbs

Follows the tutorial here.

You can only Join presorted data. A join will return pairs from an item on the left (first) to an item on the right (second). Group data with repeats. flatten is reexported from Base.

julia> flights2 =
        @> flights |>
        columns |>
        Names(:year, :month, :day, :hour, :origin, :dest, :tailnum, :carrier)(_) |>
        rows;

julia> airlines =
        @> "airlines.csv" |>
        CSV.read(_, allowmissing = :auto,) |>
        named_tuple |>
        rows;

julia> Peek(airlines)
Showing 4 of 16 rows
| :carrier |                  :name |
| --------:| ----------------------:|
|       9E |      Endeavor Air Inc. |
|       AA | American Airlines Inc. |
|       AS |   Alaska Airlines Inc. |
|       B6 |        JetBlue Airways |

julia> @> flights2 |>
        order(_, Names(:carrier)) |>
        Group(By(_, Names(:carrier))) |>
        By(_, key) |>
        Join(_, By(airlines, Names(:carrier))) |>
        over(_, pair -> over(value(pair.first), @_ merge(_, pair.second))) |>
        flatten |>
        Peek
Showing at most 4 rows
| :year | :month | :day | :hour | :origin | :dest | :tailnum | :carrier |             :name |
| -----:| ------:| ----:| -----:| -------:| -----:| --------:| --------:| -----------------:|
|  2013 |      1 |    1 |     8 |     JFK |   MSP |   N915XJ |       9E | Endeavor Air Inc. |
|  2013 |      1 |    1 |    15 |     JFK |   IAD |   N8444F |       9E | Endeavor Air Inc. |
|  2013 |      1 |    1 |    14 |     JFK |   BUF |   N920XJ |       9E | Endeavor Air Inc. |
|  2013 |      1 |    1 |    15 |     JFK |   SYR |   N8409N |       9E | Endeavor Air Inc. |

Join is full by default; use when to mimic other kinds of joins.

julia> weather =
        @> "weather.csv" |>
        CSV.read(_, allowmissing = :auto) |>
        named_tuple |>
        rows;

julia> Peek(weather)
Showing 4 of 26115 rows
| :origin | :year | :month | :day | :hour | :temp | :dewp | :humid | :wind_dir | :wind_speed | :wind_gust | :precip | :pressure | :visib |          :time_hour |
| -------:| -----:| ------:| ----:| -----:| -----:| -----:| ------:| ---------:| -----------:| ----------:| -------:| ---------:| ------:| -------------------:|
|     EWR |  2013 |      1 |    1 |     1 | 39.02 | 26.06 |  59.37 |       270 |    10.35702 |    missing |     0.0 |    1012.0 |   10.0 | 2013-01-01 01:00:00 |
|     EWR |  2013 |      1 |    1 |     2 | 39.02 | 26.96 |  61.63 |       250 |     8.05546 |    missing |     0.0 |    1012.3 |   10.0 | 2013-01-01 02:00:00 |
|     EWR |  2013 |      1 |    1 |     3 | 39.02 | 28.04 |  64.43 |       240 |     11.5078 |    missing |     0.0 |    1012.5 |   10.0 | 2013-01-01 03:00:00 |
|     EWR |  2013 |      1 |    1 |     4 | 39.92 | 28.04 |  62.21 |       250 |    12.65858 |    missing |     0.0 |    1012.2 |   10.0 | 2013-01-01 04:00:00 |

julia> const selector = Names(:origin, :year, :month, :day, :hour);

julia> @> flights2 |>
        order(_, selector) |>
        Group(By(_, selector)) |>
        By(_, key) |>
        Join(_, By(weather, selector)) |>
        when(_, @_ !ismissing(_.first)) |>
        over(_, pair -> over(value(pair.first), @_ merge(_, pair.second))) |>
        flatten |>
        Peek
Showing at most 4 rows
| :year | :month | :day | :hour | :origin | :dest | :tailnum | :carrier | :temp | :dewp | :humid | :wind_dir | :wind_speed | :wind_gust | :precip | :pressure | :visib |          :time_hour |
| -----:| ------:| ----:| -----:| -------:| -----:| --------:| --------:| -----:| -----:| ------:| ---------:| -----------:| ----------:| -------:| ---------:| ------:| -------------------:|
|  2013 |      1 |    1 |     5 |     EWR |   IAH |   N14228 |       UA | 39.02 | 28.04 |  64.43 |       260 |    12.65858 |    missing |     0.0 |    1011.9 |   10.0 | 2013-01-01 05:00:00 |
|  2013 |      1 |    1 |     5 |     EWR |   ORD |   N39463 |       UA | 39.02 | 28.04 |  64.43 |       260 |    12.65858 |    missing |     0.0 |    1011.9 |   10.0 | 2013-01-01 05:00:00 |
|  2013 |      1 |    1 |     6 |     EWR |   FLL |   N516JB |       B6 | 37.94 | 28.04 |  67.21 |       240 |     11.5078 |    missing |     0.0 |    1012.4 |   10.0 | 2013-01-01 06:00:00 |
|  2013 |      1 |    1 |     6 |     EWR |   SFO |   N53441 |       UA | 37.94 | 28.04 |  67.21 |       240 |     11.5078 |    missing |     0.0 |    1012.4 |   10.0 | 2013-01-01 06:00:00 |

julia> planes =
        @> "planes.csv" |>
        CSV.read(_, allowmissing = :auto) |>
        named_tuple |>
        rename(_, model_year = Name(:year)) |>
        rows;

julia> Peek(planes)
Showing 4 of 3322 rows
| :tailnum |                   :type |    :manufacturer |    :model | :engines | :seats |  :speed |   :engine | :model_year |
| --------:| -----------------------:| ----------------:| ---------:| --------:| ------:| -------:| ---------:| -----------:|
|   N10156 | Fixed wing multi engine |          EMBRAER | EMB-145XR |        2 |     55 | missing | Turbo-fan |        2004 |
|   N102UW | Fixed wing multi engine | AIRBUS INDUSTRIE |  A320-214 |        2 |    182 | missing | Turbo-fan |        1998 |
|   N103US | Fixed wing multi engine | AIRBUS INDUSTRIE |  A320-214 |        2 |    182 | missing | Turbo-fan |        1999 |
|   N104UW | Fixed wing multi engine | AIRBUS INDUSTRIE |  A320-214 |        2 |    182 | missing | Turbo-fan |        1999 |

julia> @> flights2 |>
            order(_, Names(:tailnum), (@_ !ismissing(_.tailnum))) |>
            Group(By(_, Names(:tailnum))) |>
            By(_, key) |>
            Join(_, By(planes, Names(:tailnum))) |>
            when(_, @_ ismissing(_.second)) |>
            over(_, @_ transform(key(_.first),
                n = length(value(_.first))
            )) |>
            make_columns |>
            rows |>
            order(_, Names(:n), rev = true) |>
            Peek
Showing 4 of 721 rows
| :tailnum |  :n |
| --------:| ---:|
|   N725MQ | 575 |
|   N722MQ | 513 |
|   N723MQ | 507 |
|   N713MQ | 483 |

Window functions

Follows the tutorial here.

julia> batting =
        @> "batting.csv" |>
        CSV.read(_, allowmissing = :auto) |>
        named_tuple |>
        rows;

julia> Peek(batting)
Showing 4 of 19404 rows
| :playerID | :yearID | :teamID |  :G | :AB |  :R |  :H |
| ---------:| -------:| -------:| ---:| ---:| ---:| ---:|
| aaronha01 |    1954 |     ML1 | 122 | 468 |  58 | 131 |
| aaronha01 |    1955 |     ML1 | 153 | 602 | 105 | 189 |
| aaronha01 |    1956 |     ML1 | 153 | 609 | 106 | 200 |
| aaronha01 |    1957 |     ML1 | 151 | 615 | 118 | 198 |

julia> players =
        @> batting |>
        Group(By(_, Names(:playerID)));

julia> @> players |>
        over(_, @_ @> order(value(_), Names(:H)) |> view(_, 1:2)) |>
        flatten |>
        when(_, @_ _.H > 0) |>
        Peek
Showing at most 4 rows
| :playerID | :yearID | :teamID |  :G | :AB |  :R |  :H |
| ---------:| -------:| -------:| ---:| ---:| ---:| ---:|
| aaronha01 |    1976 |     ML4 |  85 | 271 |  22 |  62 |
| aaronha01 |    1974 |     ATL | 112 | 340 |  47 |  91 |
| abreubo01 |    1996 |     HOU |  15 |  22 |   1 |   5 |
| abreubo01 |    2012 |     LAA |   8 |  24 |   1 |   5 |

julia> using StatsBase: ordinalrank

julia> @> players |>
        over(_, @_ @> columns(value(_)) |>
            transform(_, G_rank = ordinalrank(_.G)) |>
            rows
        ) |>
        flatten |>
        Peek
Showing at most 4 rows
| :playerID | :yearID | :teamID |  :G | :AB |  :R |  :H | :G_rank |
| ---------:| -------:| -------:| ---:| ---:| ---:| ---:| -------:|
| aaronha01 |    1954 |     ML1 | 122 | 468 |  58 | 131 |       4 |
| aaronha01 |    1955 |     ML1 | 153 | 602 | 105 | 189 |      13 |
| aaronha01 |    1956 |     ML1 | 153 | 609 | 106 | 200 |      14 |
| aaronha01 |    1957 |     ML1 | 151 | 615 | 118 | 198 |      12 |

Programming with LightQuery

Follows the tutorial here.

Be sure to @inline if you rely on constant propagation.

julia> df = (
            g1 = [1, 1, 2, 2, 2],
            g2 = [1, 2, 1, 2, 1],
            a = 1:5,
            b = 1:5
        ) |>
        rows;

julia> Peek(df)
Showing 4 of 5 rows
| :g1 | :g2 |  :a |  :b |
| ---:| ---:| ---:| ---:|
|   1 |   1 |   1 |   1 |
|   1 |   2 |   2 |   2 |
|   2 |   1 |   3 |   3 |
|   2 |   2 |   4 |   4 |

julia> my_summarize_(df, group_var) =
        @> df |>
        order(_, group_var) |>
        Group(By(_, group_var)) |>
        over(_, @_ transform(key(_),
            a = sum(columns(value(_)).a)
        )) |>
        make_columns |>
        rows;

julia> Peek(my_summarize_(df, Names(:g1)))
| :g1 |  :a |
| ---:| ---:|
|   1 |   3 |
|   2 |  12 |

julia> Peek(my_summarize_(df, Names(:g2)))
| :g2 |  :a |
| ---:| ---:|
|   1 |   9 |
|   2 |   6 |

julia> @inline my_summarize(df, group_vars...) =
        my_summarize_(df, Names(group_vars...));

julia> Peek(my_summarize(df, :g1))
| :g1 |  :a |
| ---:| ---:|
|   1 |   3 |
|   2 |  12 |

Index

LightQuery.By

LightQuery.Group

LightQuery.Join

LightQuery.Length

LightQuery.Name

LightQuery.Names

LightQuery.Peek

LightQuery.columns

LightQuery.gather

LightQuery.key

LightQuery.make_columns

LightQuery.named_tuple

LightQuery.order

LightQuery.over

LightQuery.remove

LightQuery.rename

LightQuery.row_type

LightQuery.rows

LightQuery.spread

LightQuery.transform

LightQuery.unzip

LightQuery.value

LightQuery.when

LightQuery.@>

LightQuery.@_

Autodocs

LightQuery.By — Type.

By(it, call)

Mark that it has been pre-sorted by call. For use with Group or Join.

julia> using LightQuery

julia> By([1, 2], identity)
By{Array{Int64,1},typeof(identity)}([1, 2], identity)

source

LightQuery.Group — Method.

Group(it::By)

Group consecutive keys in it. Requires a presorted object (see By). Relies on the fact that iteration states can be converted to indices; thus, you might have to define LightQuery.state_to_index for unrecognized types.

julia> using LightQuery

julia> Group(By([1, 3, 2, 4], iseven)) |> collect
2-element Array{Pair{Bool,SubArray{Int64,1,Array{Int64,1},Tuple{UnitRange{Int64}},true}},1}:
 0 => [1, 3]
 1 => [2, 4]

source

LightQuery.Join — Type.

Join(left::By, right::By)

Find all pairs where isequal(left.call(left.it), right.call(right.it)).

julia> using LightQuery

julia> Join(
            By([1, 2, 5, 6], identity),
            By([1, 3, 4, 6], identity)
        ) |>
        collect
6-element Array{Pair{Union{Missing, Int64},Union{Missing, Int64}},1}:
       1 => 1
       2 => missing
 missing => 3
 missing => 4
       5 => missing
       6 => 6

Assumes left and right are both strictly sorted (no repeats). If there are repeats, Group first.

julia> @> [1, 1, 2, 2] |>
        Group(By(_, identity)) |>
        By(_, first) |>
        Join(_, By([1, 2], identity)) |>
        collect
2-element Array{Pair{Pair{Int64,SubArray{Int64,1,Array{Int64,1},Tuple{UnitRange{Int64}},true}},Int64},1}:
 (1=>[1, 1]) => 1
 (2=>[2, 2]) => 2

For other join flavors, combine with when.

julia> @> Join(
            By([1, 2, 5, 6], identity),
            By([1, 3, 4, 6], identity)
        ) |>
        when(_, @_ !ismissing(_.first)) |>
        collect
4-element Array{Pair{Union{Missing, Int64},Union{Missing, Int64}},1}:
 1 => 1
 2 => missing
 5 => missing
 6 => 6

source

LightQuery.Length — Type.

Length(it, length)

Allow optimizations based on length. Especially useful before make_columns.

julia> using LightQuery

julia> @> Filter(iseven, 1:4) |>
        Length(_, 2) |>
        collect
2-element Array{Int64,1}:
 2
 4

source

LightQuery.Name — Method.

Name(name)

Create a typed name. Can be used as a function to getproperty, with a default to missing. For multiple names, see Names.

julia> using LightQuery

julia> (a = 1,) |>
        Name(:a)
1

julia> (a = 1,) |>
        Name(:b)
missing

julia> missing |>
        Name(:a)
missing

source

LightQuery.Names — Method.

Names(the_names...)

Create typed names. Can be used to as a function to assign or select names, with a default to missing. For just one name, see Name.

julia> using LightQuery

julia> (1, 1.0) |>
        Names(:a, :b)
(a = 1, b = 1.0)

julia> (a = 1, b = 1.0) |>
        Names(:a)
(a = 1,)

julia> (a = 1,) |>
        Names(:a, :b)
(a = 1, b = missing)

julia> missing |>
        Names(:a)
(a = missing,)

source

LightQuery.Peek — Method.

Peek(it; max_rows = 4)

Get a peek of an iterator which returns named tuples. Will show no more than max_rows.

julia> using LightQuery

julia> (a = 1:5, b = 5:-1:1) |>
        rows |>
        Peek
Showing 4 of 5 rows
|  :a |  :b |
| ---:| ---:|
|   1 |   5 |
|   2 |   4 |
|   3 |   3 |
|   4 |   2 |

If inference cannot detect names, it will use the names of the first item. Map Names over it to override this behavior.

julia> [(a = 1,), (a = 2, b = 2.0)] |>
        Peek
|  :a |
| ---:|
|   1 |
|   2 |

julia> @> [(a = 1,), (a = 2, b = 2.0)] |>
        over(_, Names(:a, :b)) |>
        Peek
|  :a |      :b |
| ---:| -------:|
|   1 | missing |
|   2 |     2.0 |

source

LightQuery.columns — Method.

columns(it)

Inverse of rows. Always lazy, see make_columns for eager version.

julia> using LightQuery

julia> (a = [1], b = [1.0]) |>
        rows |>
        columns
(a = [1], b = [1.0])

source

LightQuery.gather — Method.

gather(it; assignments...)

For each key => value pair in assignments, gather the Names in value into a single key. Inverse of spread.

julia> using LightQuery

julia> @> (a = 1, b = 1.0, c = 1//1) |>
        gather(_, d = Names(:a, :c))
(b = 1.0, d = (a = 1, c = 1//1))

source

LightQuery.key — Method.

key(it)

The key in a key => value pair.

source

LightQuery.make_columns — Method.

make_columns(it)

Collect into columns. Always eager, see columns lazy version.

julia> using LightQuery

julia> [(a = 1, b = 1.0), (a = 2, b = 2.0)] |>
        make_columns
(a = [1, 2], b = [1.0, 2.0])

If inference cannot detect names, it will use the names of the first item. Map Names over it to override this behavior.

julia> [(a = 1,), (a = 2, b = 2.0)] |>
        Peek
|  :a |
| ---:|
|   1 |
|   2 |

julia> @> [(a = 1,), (a = 2, b = 2.0)] |>
        over(_, Names(:a, :b)) |>
        Peek
|  :a |      :b |
| ---:| -------:|
|   1 | missing |
|   2 |     2.0 |

source

LightQuery.named_tuple — Method.

named_tuple(it)

Coerce to a named_tuple. For performance with working with arbitrary structs, define and @inline propertynames.

julia> using LightQuery

julia> @inline Base.propertynames(p::Pair) = (:first, :second);

julia> named_tuple(:a => 1)
(first = :a, second = 1)

source

LightQuery.order — Method.

order(it, call; keywords...)
order(it, call, condition; keywords...)

Generalized sort. keywords will be passed to sort!; see the documentation there for options. See By for a way to explicitly mark that an object has been sorted. Most performant if call is type stable, if not, consider using a condition to filter.

julia> using LightQuery

julia> order([2, 1], identity)
2-element view(::Array{Int64,1}, [2, 1]) with eltype Int64:
 1
 2

julia> order([2, 1, missing], identity, !ismissing)
2-element view(::Array{Union{Missing, Int64},1}, [2, 1]) with eltype Union{Missing, Int64}:
 1
 2

source

LightQuery.over — Method.

over(it, call)

Lazy map with argument order reversed.

source

LightQuery.remove — Method.

remove(it, the_names...)

Remove the_names. Inverse of transform.

julia> using LightQuery

julia> @> (a = 1, b = 1.0) |>
        remove(_, :b)
(a = 1,)

source

LightQuery.rename — Method.

rename(it; renames...)

Rename it. Because constants do not constant propagate through key-word arguments, wrap with Name.

julia> using LightQuery

julia> @> (a = 1, b = 1.0) |>
        rename(_, c = Name(:a))
(b = 1.0, c = 1)

source

LightQuery.row_type — Method.

row_type(file::CSV.File)

Find the type of a row of a CSV.File if it was converted to a named_tuple. Useful to make type annotations to ensure stability.

julia> using LightQuery

julia> @> "test.csv" |>
        CSV.File(_, allowmissing = :auto) |>
        row_type
NamedTuple{(:a, :b),T} where T<:Tuple{Int64,Float64}

source

LightQuery.rows — Method.

rows(it)

Iterator over rows of a NamedTuple of arrays. Inverse of columns. See Peek for a way to view rows.

julia> using LightQuery

julia> (a = [1, 2], b = [1.0, 2.0]) |>
        rows |>
        collect
2-element Array{NamedTuple{(:a, :b),Tuple{Int64,Float64}},1}:
 (a = 1, b = 1.0)
 (a = 2, b = 2.0)

source

LightQuery.spread — Method.

spread(it::NamedTuple, the_names...)

Unnest nested it in name. Inverse of gather.

julia> using LightQuery

julia> @> (b = 1.0, d = (a = 1, c = 1//1)) |>
        spread(_, :d)
(b = 1.0, a = 1, c = 1//1)

source

LightQuery.transform — Method.

transform(it; assignments...)

Merge assignments into it. Inverse of remove.

julia> using LightQuery

julia> @> (a = 1,) |>
        transform(_, b = 1.0)
(a = 1, b = 1.0)

source

LightQuery.unzip — Method.

unzip(it, n)

Unzip an iterator it which returns tuples of length n. Use Val(n) to guarantee type stability.

julia> using LightQuery

julia> unzip([(1, 1.0), (2, 2.0)], 2)
([1, 2], [1.0, 2.0])

julia> unzip([(1, 1.0), (2, 2.0)], Val(2))
([1, 2], [1.0, 2.0])

source

LightQuery.value — Method.

value(it)

The value in a key => value pair.

source

LightQuery.when — Method.

when(it, call)

Lazy filter with argument order reversed.

source

LightQuery.@> — Macro.

macro >(body)

If body is in the form body_ |> tail_, call @_ on tail, and recur on body.

julia> using LightQuery

julia> @> 0 |> _ - 1 |> abs
1

source

LightQuery.@_ — Macro.

macro _(body)

Terser function syntax. The arguments are inside the body; the first argument is _, the second argument is __, etc. Will always @inline.

julia> using LightQuery

julia> (@_ _ + 1)(1)
2

julia> map((@_ __ - _), (1, 2), (2, 1))
(1, -1)

source