Tutorial

Tutorial

I'm going to use the flights data from the dplyr tutorial. This data is in the test folder of this package; I created it with the following R code:

library(nycflights13)
setwd("C:/Users/hp/.julia/dev/LightQuery/test")
write.csv(airports, "airports.csv", na = "", row.names = FALSE)
write.csv(flights, "flights.csv", na = "", row.names = FALSE)

Let's import the tools we need. I'm pulling in from tools from Dates and TimeZones and modifying them to work with missing data.

julia> using LightQuery

julia> using Dates: DateTime, Day, Minute

julia> Minute(::Missing) = missing;

julia> using Unitful: mi, °, ft

julia> using TimeZones: TimeZone, VariableTimeZone, ZonedDateTime

julia> TimeZone_or_missing(time_zone) =
            try
                TimeZone(time_zone)
            catch an_error
                if isa(an_error, ArgumentError)
                    missing
                else
                    rethrow(an_error)
                end
            end;

julia> ZonedDateTime(::DateTime, ::Missing) = missing;

I re-export CSV for input-output. See the documentation there for information about CSV.File.

julia> airports_file = CSV.File("airports.csv",
            allowmissing = :auto
        )
CSV.File("airports.csv", rows=1458):
Tables.Schema:
 :faa    String
 :name   String
 :lat    Float64
 :lon    Float64
 :alt    Int64
 :tz     Int64
 :dst    String
 :tzone  String

Let's take a look at the first row. Use named_tuple to coerce a CSV.Row to a NamedTuple.

julia> airport =
        airports_file |>
        first |>
        named_tuple
(faa = "04G", name = "Lansdowne Airport", lat = 41.1304722, lon = -80.6195833, alt = 1044, tz = -5, dst = "A", tzone = "America/New_York")

As a start, I want to rename so that I understand what the columns mean. When you rename, names need to be wrapped with Name. Here, I use the chaining macro @> to chain several calls together.

julia> airport =
        @> airport |>
        rename(_,
            airport_code = Name(:faa),
            latitude = Name(:lat),
            longitude = Name(:lon),
            altitude = Name(:alt),
            time_zone_offset = Name(:tz),
            daylight_savings = Name(:dst),
            time_zone = Name(:tzone)
        )
(name = "Lansdowne Airport", airport_code = "04G", latitude = 41.1304722, longitude = -80.6195833, altitude = 1044, time_zone_offset = -5, daylight_savings = "A", time_zone = "America/New_York")

Let's create a proper TimeZone. Note that it could be the case that Julia can't recognize the time-zone string. Thus, I make a TimeZone_or_missing function to return missing in that case.

julia> airport =
        @> airport |>
        transform(_,
            time_zone = TimeZone_or_missing(_.time_zone)
        )
(name = "Lansdowne Airport", airport_code = "04G", latitude = 41.1304722, longitude = -80.6195833, altitude = 1044, time_zone_offset = -5, daylight_savings = "A", time_zone = America/New_York (UTC-5/UTC-4))

Now that we have a true timezone, we can remove all data that is contingent on timezone.

julia> airport =
        @> airport |>
        remove(_,
            :time_zone_offset,
            :daylight_savings
        )
(name = "Lansdowne Airport", airport_code = "04G", latitude = 41.1304722, longitude = -80.6195833, altitude = 1044, time_zone = America/New_York (UTC-5/UTC-4))

Let's also add proper units to our variables.

julia> airport =
        @> airport |>
        transform(_,
            latitude = _.latitude * °,
            longitude = _.longitude * °,
            altitude = _.altitude * ft
        )
(name = "Lansdowne Airport", airport_code = "04G", latitude = 41.1304722°, longitude = -80.6195833°, altitude = 1044 ft, time_zone = America/New_York (UTC-5/UTC-4))

Let's put it all together.

julia> function process_airport(airport_row)
            @> airport_row |>
            named_tuple |>
            rename(_,
                airport_code = Name(:faa),
                latitude = Name(:lat),
                longitude = Name(:lon),
                altitude = Name(:alt),
                time_zone_offset = Name(:tz),
                daylight_savings = Name(:dst),
                time_zone = Name(:tzone)
            ) |>
            transform(_,
                time_zone = TimeZone_or_missing(_.time_zone),
                latitude = _.latitude * °,
                longitude = _.longitude * °,
                altitude = _.altitude * ft
            ) |>
            remove(_,
                :time_zone_offset,
                :daylight_savings
            )
        end;

I use over to lazily map.

julia> airports =
        @> airports_file |>
        over(_, process_airport);

When it comes time to collect, I'm calling make_columns then rows. It makes sense to store this data column-wise. This is because there are multiple columns that might contain missing data.

julia> airports =
        airports |>
        make_columns |>
        rows;

We can use Peek to get a look at the data.

julia> Peek(airports)
Showing 4 of 1458 rows
|                         :name | :airport_code |   :latitude |   :longitude | :altitude |                     :time_zone |
| -----------------------------:| -------------:| -----------:| ------------:| ---------:| ------------------------------:|
|             Lansdowne Airport |           04G | 41.1304722° | -80.6195833° |   1044 ft | America/New_York (UTC-5/UTC-4) |
| Moton Field Municipal Airport |           06A | 32.4605722° | -85.6800278° |    264 ft |  America/Chicago (UTC-6/UTC-5) |
|           Schaumburg Regional |           06C | 41.9893408° | -88.1012428° |    801 ft |  America/Chicago (UTC-6/UTC-5) |
|               Randall Airport |           06N |  41.431912° | -74.3915611° |    523 ft | America/New_York (UTC-5/UTC-4) |

I'll also make sure the airports are indexed by their code so we can access them quickly.

julia> airports =
        @> airports |>
        indexed(_, Name(:airport_code));

julia> airports["JFK"]
(name = "John F Kennedy Intl", airport_code = "JFK", latitude = 40.639751°, longitude = -73.778925°, altitude = 13 ft, time_zone = America/New_York (UTC-5/UTC-4))

That was just the warm-up. Now let's get started working on the flights data.

julia> flights_file = CSV.File("flights.csv", allowmissing = :auto)
CSV.File("flights.csv", rows=336776):
Tables.Schema:
 :year            Int64
 :month           Int64
 :day             Int64
 :dep_time        Union{Missing, Int64}
 :sched_dep_time  Int64
 :dep_delay       Union{Missing, Int64}
 :arr_time        Union{Missing, Int64}
 :sched_arr_time  Int64
 :arr_delay       Union{Missing, Int64}
 :carrier         String
 :flight          Int64
 :tailnum         Union{Missing, String}
 :origin          String
 :dest            String
 :air_time        Union{Missing, Int64}
 :distance        Int64
 :hour            Int64
 :minute          Int64
 :time_hour       String

julia> flight =
        @> flights_file |>
        first |>
        named_tuple |>
        rename(_,
            departure_time = Name(:dep_time),
            scheduled_departure_time = Name(:sched_dep_time),
            departure_delay = Name(:dep_delay),
            arrival_time = Name(:arr_time),
            scheduled_arrival_time = Name(:sched_arr_time),
            arrival_delay = Name(:arr_delay),
            tail_number = Name(:tailnum),
            destination = Name(:dest)
        )
(year = 2013, month = 1, day = 1, carrier = "UA", flight = 1545, origin = "EWR", air_time = 227, distance = 1400, hour = 5, minute = 15, time_hour = "2013-01-01 05:00:00", departure_time = 517, scheduled_departure_time = 515, departure_delay = 2, arrival_time = 830, scheduled_arrival_time = 819, arrival_delay = 11, tail_number = "N14228", destination = "IAH")

We can use our airports data to make datetimes with timezones.

julia> scheduled_departure_time = ZonedDateTime(
            DateTime(flight.year, flight.month, flight.day, flight.hour, flight.minute),
            airports[flight.origin].time_zone
        )
2013-01-01T05:15:00-05:00

Note the scheduled arrival time is 818. This means 8:18. We can use divrem(_, 100) to split it up. Note I'm accessing the time_zone with Name (which will default to missing). This is because some of the destinations are not in the flights dataset.

julia> scheduled_arrival_time = ZonedDateTime(
            DateTime(flight.year, flight.month, flight.day, divrem(flight.scheduled_arrival_time, 100)...),
            Name(:time_zone)(airports[flight.destination]))
2013-01-01T08:19:00-06:00

What if it was an over-night flight? We can add a day to the arrival time if it wasn't later than the departure time.

julia> if !(scheduled_arrival_time > scheduled_departure_time)
            scheduled_arrival_time = scheduled_arrival_time + Day(1)
        end

Let's put it all together:

julia> function process_flight(row)
            flight =
                @> named_tuple(row) |>
                rename(_,
                    departure_time = Name(:dep_time),
                    scheduled_departure_time = Name(:sched_dep_time),
                    departure_delay = Name(:dep_delay),
                    arrival_time = Name(:arr_time),
                    scheduled_arrival_time = Name(:sched_arr_time),
                    arrival_delay = Name(:arr_delay),
                    tail_number = Name(:tailnum),
                    destination = Name(:dest)
                )
            scheduled_departure_time = ZonedDateTime(
                DateTime(flight.year, flight.month, flight.day, flight.hour, flight.minute),
                airports[flight.origin].time_zone
            )
            scheduled_arrival_time = ZonedDateTime(
                DateTime(flight.year, flight.month, flight.day, divrem(flight.scheduled_arrival_time, 100)...),
                Name(:time_zone)(airports[flight.destination])
            )
            if scheduled_arrival_time !== missing &&
                !(scheduled_arrival_time > scheduled_departure_time)
                scheduled_arrival_time = scheduled_arrival_time + Day(1)
            end
            @> flight |>
                transform(_,
                    scheduled_departure_time = scheduled_departure_time,
                    scheduled_arrival_time = scheduled_arrival_time,
                    air_time = Minute(_.air_time),
                    distance = _.distance * mi,
                    departure_delay = Minute(_.departure_delay),
                    arrival_delay = Minute(_.arrival_delay)
                ) |>
                remove(_, :year, :month, :day, :hour, :minute, :time_hour,
                    :departure_time, :arrival_time)
        end;

julia> flights =
        @> flights_file |>
        over(_, process_flight) |>
        make_columns |>
        rows;

julia> Peek(flights)
Showing 4 of 336776 rows
| :carrier | :flight | :origin |   :air_time | :distance | :scheduled_departure_time | :departure_delay |   :scheduled_arrival_time | :arrival_delay | :tail_number | :destination |
| --------:| -------:| -------:| -----------:| ---------:| -------------------------:| ----------------:| -------------------------:| --------------:| ------------:| ------------:|
|       UA |    1545 |     EWR | 227 minutes |   1400 mi | 2013-01-01T05:15:00-05:00 |        2 minutes | 2013-01-01T08:19:00-06:00 |     11 minutes |       N14228 |          IAH |
|       UA |    1714 |     LGA | 227 minutes |   1416 mi | 2013-01-01T05:29:00-05:00 |        4 minutes | 2013-01-01T08:30:00-06:00 |     20 minutes |       N24211 |          IAH |
|       AA |    1141 |     JFK | 160 minutes |   1089 mi | 2013-01-01T05:40:00-05:00 |        2 minutes | 2013-01-01T08:50:00-05:00 |     33 minutes |       N619AA |          MIA |
|       B6 |     725 |     JFK | 183 minutes |   1576 mi | 2013-01-01T05:45:00-05:00 |        -1 minute |                   missing |    -18 minutes |       N804JB |          BQN |

Theortically, notice the distances between two airports is always the same. Let's make sure this is also the case in our data. First, order by origin, destination, and distance. Then Group By the same variables.

julia> paths_grouped =
        @> flights |>
        order(_, Names(:origin, :destination, :distance)) |>
        Group(By(_, Names(:origin, :destination, :distance)));

Each Group will contain a key and value

julia> path = first(paths_grouped);

julia> key(path)
(origin = "EWR", destination = "ALB", distance = 143 mi)

julia> value(path) |> Peek
Showing 4 of 439 rows
| :carrier | :flight | :origin |  :air_time | :distance | :scheduled_departure_time | :departure_delay |   :scheduled_arrival_time | :arrival_delay | :tail_number | :destination |
| --------:| -------:| -------:| ----------:| ---------:| -------------------------:| ----------------:| -------------------------:| --------------:| ------------:| ------------:|
|       EV |    4112 |     EWR | 33 minutes |    143 mi | 2013-01-01T13:17:00-05:00 |       -2 minutes | 2013-01-01T14:23:00-05:00 |    -10 minutes |       N13538 |          ALB |
|       EV |    3260 |     EWR | 36 minutes |    143 mi | 2013-01-01T16:21:00-05:00 |       34 minutes | 2013-01-01T17:24:00-05:00 |     40 minutes |       N19554 |          ALB |
|       EV |    4170 |     EWR | 31 minutes |    143 mi | 2013-01-01T20:04:00-05:00 |       52 minutes | 2013-01-01T21:12:00-05:00 |     44 minutes |       N12540 |          ALB |
|       EV |    4316 |     EWR | 33 minutes |    143 mi | 2013-01-02T13:27:00-05:00 |        5 minutes | 2013-01-02T14:33:00-05:00 |    -14 minutes |       N14153 |          ALB |

At this point, we don't need any of the value data. All we need is the key.

julia> paths =
        @> paths_grouped |>
        over(_, key) |>
        make_columns |>
        rows;

julia> Peek(paths)
Showing 4 of 226 rows
| :origin | :destination | :distance |
| -------:| ------------:| ---------:|
|     EWR |          ALB |    143 mi |
|     EWR |          ANC |   3370 mi |
|     EWR |          ATL |    746 mi |
|     EWR |          AUS |   1504 mi |

Notice the data is already sorted by origin and destination, so that for our second Group, we don't need to order first.

julia> distinct_distances =
        @> paths |>
        Group(By(_, Names(:origin, :destination))) |>
        over(_, @_ transform(key(_),
            number = length(value(_))
        ));

julia> Peek(distinct_distances)
Showing at most 4 rows
| :origin | :destination | :number |
| -------:| ------------:| -------:|
|     EWR |          ALB |       1 |
|     EWR |          ANC |       1 |
|     EWR |          ATL |       1 |
|     EWR |          AUS |       1 |

Let's see when there are two different distances for the same path:

julia> @> distinct_distances |>
        when(_, @_ _.number != 1) |>
        Peek
Showing at most 4 rows
| :origin | :destination | :number |
| -------:| ------------:| -------:|
|     EWR |          EGE |       2 |
|     JFK |          EGE |       2 |

That's strange. What's up with the EGE airport? Let's take a Peek.

julia> @> flights |>
        when(_, @_ _.destination == "EGE") |>
        Peek
Showing at most 4 rows
| :carrier | :flight | :origin |   :air_time | :distance | :scheduled_departure_time | :departure_delay |   :scheduled_arrival_time | :arrival_delay | :tail_number | :destination |
| --------:| -------:| -------:| -----------:| ---------:| -------------------------:| ----------------:| -------------------------:| --------------:| ------------:| ------------:|
|       UA |    1597 |     EWR | 287 minutes |   1726 mi | 2013-01-01T09:28:00-05:00 |       -2 minutes | 2013-01-01T12:20:00-07:00 |     13 minutes |       N27733 |          EGE |
|       AA |     575 |     JFK | 280 minutes |   1747 mi | 2013-01-01T17:00:00-05:00 |       -5 minutes | 2013-01-01T19:50:00-07:00 |      3 minutes |       N5DRAA |          EGE |
|       UA |    1597 |     EWR | 261 minutes |   1726 mi | 2013-01-02T09:28:00-05:00 |         1 minute | 2013-01-02T12:20:00-07:00 |      3 minutes |       N24702 |          EGE |
|       AA |     575 |     JFK | 260 minutes |   1747 mi | 2013-01-02T17:00:00-05:00 |        5 minutes | 2013-01-02T19:50:00-07:00 |     16 minutes |       N631AA |          EGE |

Looks (to me) like two different sources are reporting different info about the same flight.

Interface

Macros

LightQuery.@_ — Macro.

macro _(body)

Terser function syntax. The arguments are inside the body; the first argument is _, the second argument is __, etc. Will always @inline.

julia> using LightQuery

julia> (@_ _ + 1)(1)
2

julia> map((@_ __ - _), (1, 2), (2, 1))
(1, -1)

source

LightQuery.@> — Macro.

macro >(body)

If body is in the form body_ |> tail_, call @_ on tail, and recur on body.

julia> using LightQuery

julia> @> 0 |> _ - 1 |> abs
1

source

Columns

LightQuery.named_tuple — Function.

named_tuple(it)

Coerce to a named_tuple. For performance with working with arbitrary structs, define and @inline propertynames.

julia> using LightQuery

julia> @inline Base.propertynames(p::Pair) = (:first, :second);

julia> named_tuple(:a => 1)
(first = :a, second = 1)

source

LightQuery.Name — Type.

Name(name)

Create a typed name. Can be used as a function to getproperty, with a default to missing. For multiple names, see Names.

julia> using LightQuery

julia> (a = 1,) |>
        Name(:a)
1

julia> (a = 1,) |>
        Name(:b)
missing

julia> missing |>
        Name(:a)
missing

source

LightQuery.Names — Type.

Names(the_names...)

Create typed names. Can be used to as a function to assign or select names, with a default to missing. For just one name, see Name.

julia> using LightQuery

julia> (1, 1.0) |>
        Names(:a, :b)
(a = 1, b = 1.0)

julia> (a = 1, b = 1.0) |>
        Names(:a)
(a = 1,)

julia> (a = 1,) |>
        Names(:a, :b)
(a = 1, b = missing)

julia> missing |>
        Names(:a)
(a = missing,)

source

LightQuery.rename — Function.

rename(it; renames...)

Rename it. Because constants do not constant propagate through key-word arguments, wrap with Name.

julia> using LightQuery

julia> @> (a = 1, b = 1.0) |>
        rename(_, c = Name(:a))
(b = 1.0, c = 1)

source

LightQuery.transform — Function.

transform(it; assignments...)

Merge assignments into it. Inverse of remove.

julia> using LightQuery

julia> @> (a = 1,) |>
        transform(_, b = 1.0)
(a = 1, b = 1.0)

source

LightQuery.remove — Function.

remove(it, the_names...)

Remove the_names. Inverse of transform.

julia> using LightQuery

julia> @> (a = 1, b = 1.0) |>
        remove(_, :b)
(a = 1,)

source

LightQuery.gather — Function.

gather(it; assignments...)

For each key => value pair in assignments, gather the Names in value into a single key. Inverse of spread.

julia> using LightQuery

julia> @> (a = 1, b = 1.0, c = 1//1) |>
        gather(_, d = Names(:a, :c))
(b = 1.0, d = (a = 1, c = 1//1))

source

LightQuery.spread — Function.

spread(it::NamedTuple, the_names...)

Unnest nested it in name. Inverse of gather.

julia> using LightQuery

julia> @> (b = 1.0, d = (a = 1, c = 1//1)) |>
        spread(_, :d)
(b = 1.0, a = 1, c = 1//1)

source

Rows

LightQuery.unzip — Function.

unzip(it, n)

Unzip an iterator it which returns tuples of length n. Use Val(n) to guarantee type stability.

julia> using LightQuery

julia> unzip([(1, 1.0), (2, 2.0)], 2)
([1, 2], [1.0, 2.0])

julia> unzip([(1, 1.0), (2, 2.0)], Val(2))
([1, 2], [1.0, 2.0])

source

LightQuery.Enumerated — Type.

Enumerated{It}

Relies on the fact that iteration states can be converted to indices; thus, you might have to define LightQuery.state_to_index for unrecognized types. Ignores some iterators like Filter.

julia> using LightQuery

julia> when([4, 3, 2, 1], iseven) |> Enumerated |> collect
2-element Array{Tuple{Int64,Int64},1}:
 (1, 4)
 (3, 2)

source

LightQuery.over — Function.

over(it, call)

Lazy map with argument order reversed.

source

LightQuery.indexed — Function.

indexed(it, call)

Index it by the results of call, with a default to missing. Relies on Enumerated.

```jldoctest julia> using LightQuery

julia> result = indexed( [ (item = "b", index = 2), (item = "a", index = 1) ], Name(:index) );

julia> result1

julia> result[3] missing

source

LightQuery.when — Function.

when(it, call)

Lazy filter with argument order reversed.

source

LightQuery.order — Function.

order(it, call; keywords...)

Generalized sort. keywords will be passed to sort!; see the documentation there for options. See By for a way to explicitly mark that an object has been sorted. Relies on Enumerated.

julia> using LightQuery

julia> order([
            (item = "b", index = 2),
            (item = "a", index = 1)
        ], Names(:index))
2-element view(::Array{NamedTuple{(:item, :index),Tuple{String,Int64}},1}, [2, 1]) with eltype NamedTuple{(:item, :index),Tuple{String,Int64}}:
 (item = "a", index = 1)
 (item = "b", index = 2)

source

LightQuery.By — Type.

By(it, call)

Mark that it has been pre-sorted by call. For use with Group or Join.

julia> using LightQuery

julia> By([
            (item = "a", index = 1),
            (item = "b", index = 2)
        ], Names(:index));

source

LightQuery.Group — Type.

Group(it::By)

Group consecutive keys in it. Requires a presorted object (see By). Relies on Enumerated.

julia> using LightQuery

julia> Group(By(
            [
                (item = "a", group = 1),
                (item = "b", group = 1),
                (item = "c", group = 2),
                (item = "d", group = 2)
            ],
            Names(:group)
        )) |>
        collect
2-element Array{Pair{NamedTuple{(:group,),Tuple{Int64}},SubArray{NamedTuple{(:item, :group),Tuple{String,Int64}},1,Array{NamedTuple{(:item, :group),Tuple{String,Int64}},1},Tuple{UnitRange{Int64}},true}},1}:
 (group = 1,) => [(item = "a", group = 1), (item = "b", group = 1)]
 (group = 2,) => [(item = "c", group = 2), (item = "d", group = 2)]

source

LightQuery.key — Function.

key(it)

The key in a key => value pair.

source

LightQuery.value — Function.

value(it)

The value in a key => value pair.

source

LightQuery.Join — Type.

Join(left::By, right::By)

Find all pairs where isequal(left.call(left.it), right.call(right.it)).

julia> using LightQuery

julia> Join(
            By(
                [
                    (left = "a", index = 1),
                    (left = "b", index = 2),
                    (left = "e", index = 5),
                    (left = "f", index = 6)
                ],
                Names(:index)
            ),
            By(
                [
                    (right = "a", index = 1),
                    (right = "c", index = 3),
                    (right = "d", index = 4),
                    (right = "e", index = 6)
                ],
                Names(:index)
            )
        ) |>
        collect
6-element Array{Pair{Union{Missing, NamedTuple{(:left, :index),Tuple{String,Int64}}},Union{Missing, NamedTuple{(:right, :index),Tuple{String,Int64}}}},1}:
 (left = "a", index = 1) => (right = "a", index = 1)
 (left = "b", index = 2) => missing
                 missing => (right = "c", index = 3)
                 missing => (right = "d", index = 4)
 (left = "e", index = 5) => missing
 (left = "f", index = 6) => (right = "e", index = 6)

Assumes left and right are both strictly sorted (no repeats). If there are repeats, Group first. For other join flavors, combine with when. Make sure to annotate with Length if you know it.

source

LightQuery.Length — Type.

Length(it, length)

Allow optimizations based on length. Especially useful after Join and before make_columns.

julia> using LightQuery

julia> @> Filter(iseven, 1:4) |>
        Length(_, 2) |>
        collect
2-element Array{Int64,1}:
 2
 4

source

Pivot

LightQuery.item_names — Function.

item_names(it)

Find names of items in it. Used in Peek and make_columns.

julia> using LightQuery

julia> [(a = 1, b = 1.0), (a = 2, b = 2.0)] |>
        item_names
(:a, :b)

If inference cannot detect names, it will use propertynames of the first item. Map Names over it to override this behavior.

julia> [(a = 1,), (a = 2, b = 2.0)] |>
        Peek
|  :a |
| ---:|
|   1 |
|   2 |

julia> @> [(a = 1,), (a = 2, b = 2.0)] |>
        over(_, Names(:a, :b)) |>
        Peek
|  :a |      :b |
| ---:| -------:|
|   1 | missing |
|   2 |     2.0 |

source

LightQuery.rows — Function.

rows(it)

Iterator over rows of a NamedTuple of arrays. Always lazy. Inverse of columns. See Peek for a way to view.

julia> using LightQuery

julia> (a = [1, 2], b = [1.0, 2.0]) |>
        rows |>
        collect
2-element Array{NamedTuple{(:a, :b),Tuple{Int64,Float64}},1}:
 (a = 1, b = 1.0)
 (a = 2, b = 2.0)

source

LightQuery.Peek — Type.

Peek(it; max_rows = 4)

Get a peek of an iterator which returns items with propertynames. Will show no more than max_rows. Relies on item_names.

julia> using LightQuery

julia> (a = 1:5, b = 5:-1:1) |>
        rows |>
        Peek
Showing 4 of 5 rows
|  :a |  :b |
| ---:| ---:|
|   1 |   5 |
|   2 |   4 |
|   3 |   3 |
|   4 |   2 |

source

LightQuery.columns — Function.

columns(it)

Inverse of rows. Always lazy, see make_columns for eager version.

julia> using LightQuery

julia> (a = [1], b = [1.0]) |>
        rows |>
        columns
(a = [1], b = [1.0])

source

LightQuery.make_columns — Function.

make_columns(it)

Collect into columns. Always eager, see columns for lazy version. Relies on item_names.

julia> using LightQuery

julia> [(a = 1, b = 1.0), (a = 2, b = 2.0)] |>
        make_columns
(a = [1, 2], b = [1.0, 2.0])

source