LightQuery.jl

LightQuery.jl

For performance with working with arbitrary structs, explicitly define public propertynames.

LightQuery.By

LightQuery.Group

LightQuery.LeftJoin

LightQuery.Name

LightQuery.columns

LightQuery.gather

LightQuery.named_tuple

LightQuery.order_by

LightQuery.over

LightQuery.remove

LightQuery.rename

LightQuery.rows

LightQuery.select

LightQuery.spread

LightQuery.transform

LightQuery.unzip

LightQuery.when

LightQuery.@>

LightQuery.@_

LightQuery.By — Type.

By(it, f)

Marks that it has been pre-sorted by the key f. For use with Group or LeftJoin.

source

LightQuery.Group — Method.

group(b::By)

Group consecutive keys in b. Requires a presorted object (see By).

julia> using LightQuery

julia> Group(By([1, 3, 2, 4], iseven)) |> collect
2-element Array{Pair{Bool,SubArray{Int64,1,Array{Int64,1},Tuple{UnitRange{Int64}},true}},1}:
 0 => [1, 3]
 1 => [2, 4]

source

LightQuery.LeftJoin — Type.

LeftJoin(left::By, right::By)

For each value in left, look for a value with the same key in right. Requires both to be presorted (see By).

julia> using LightQuery

julia> LeftJoin(
            By([1, 2, 5, 6], identity),
            By([1, 3, 4, 6], identity)
       ) |> collect
4-element Array{Pair{Int64,Union{Missing, Int64}},1}:
 1 => 1
 2 => missing
 5 => missing
 6 => 6

source

LightQuery.Name — Type.

Name(x)

Force into the type domain.

julia> using LightQuery

julia> Name(:a)
Name{:a}()

source

LightQuery.columns — Method.

columns(it, names...)

Collect into columns. Inverse of rows.

julia> using LightQuery

julia> using Test: @inferred

julia> test(x) = columns(x, :a, :b);

julia> @inferred test([(a = 1, b = 1.0)])
(a = [1], b = [1.0])

source

LightQuery.gather — Method.

gather(data, name, names...)

Gather all the data in names into a single name. Inverse of spread.

julia> using LightQuery

julia> using Test: @inferred

julia> test(x) = gather(x, :d, :a, :c);

julia> @inferred test((a = 1, b = 2.0, c = "c"))
(b = 2.0, d = (a = 1, c = "c"))

source

LightQuery.named_tuple — Method.

named_tuple(x)

Coerce to a named_tuple.

julia> using LightQuery

julia> named_tuple(:a => 1)
(first = :a, second = 1)

source

LightQuery.order_by — Method.

order_by(it, f)

Generalized sort. Will return a By object.

julia> using LightQuery

julia> order_by(["b", "a"], identity).it
2-element view(::Array{String,1}, [2, 1]) with eltype String:
 "a"
 "b"

source

LightQuery.over — Method.

over(it, f)

Hackable reversed version of Base.Generator.

julia> using LightQuery

julia> over([1, 2], x -> x + 1) |> collect
2-element Array{Int64,1}:
 2
 3

source

LightQuery.remove — Method.

remove(data, names...)

Remove names. Inverse of transform.

julia> using LightQuery

julia> using Test: @inferred

julia> test(x) = remove(x, :b);

julia> @inferred test((a = 1, b = 2.0))
(a = 1,)

source

LightQuery.rename — Method.

rename(data; renames...)

Rename data. Currently unstable.

julia> using LightQuery

julia> using Test: @inferred

julia> test(x) = rename(x, c = :a);

julia> test((a = 1, b = 2.0))
(b = 2.0, c = 1)

source

LightQuery.rows — Method.

rows(n::NamedTuple)

Iterator over rows of a NamedTuple of names. Inverse of columns.

julia> using LightQuery

julia> using Test: @inferred

julia> @inferred first(rows((a = [1, 2], b = [2, 1])))
(a = 1, b = 2)

source

LightQuery.select — Method.

select(data::NamedTuple, names...)

Select names.

select(ss::Symbol...)

Curried form.

julia> using LightQuery

julia> using Test: @inferred

julia> test(x) = select(x, :a);

julia> @inferred test((a = 1, b = 2.0))
(a = 1,)

julia> test(x) = select(:a)(x);

julia> @inferred test((a = 1, b = 2.0))
(a = 1,)

source

LightQuery.spread — Method.

spread(data::NamedTuple, name)

Unnest nested data in name. Inverse of gather.

julia> using LightQuery

julia> using Test: @inferred

julia> test(x) = spread(x, :d);

julia> @inferred test((b = 2.0, d = (a = 1, c = "c")))
(b = 2.0, a = 1, c = "c")

source

LightQuery.transform — Method.

transform(data; assignments...)

Apply the functions in assignments to data, assign to the corresponding keys, and merge back in the original.

julia> using LightQuery

julia> using Test: @inferred

julia> @inferred transform((a = 1, b = 2.0), c = @_ _.a + _.b)
(a = 1, b = 2.0, c = 3.0)

source

LightQuery.unzip — Method.

unzip(it, n)

Unzip an iterator it which returns tuples of length n.

julia> using LightQuery

julia> f(x) = (x, x + 1.0);

julia> unzip(over([1], f), 2)
([1], [2.0])

julia> unzip(over([1, missing], f), 2);

julia> unzip(zip([1], [1.0]), 2)
([1], [1.0])

julia> unzip([(1, 1.0)], 2)
([1], [1.0])

source

LightQuery.when — Method.

when(it, f)

Hackable reversed version of Base.Iterators.Filter.

julia> using LightQuery

julia> when([1, 2], x -> x > 1) |> collect
1-element Array{Int64,1}:
 2

source

LightQuery.@> — Macro.

macro >(body)

If body is in the form body_ |> tail_, call @_ on tail, and recur on body.

julia> using LightQuery

julia> @> 0 |> _ - 1 |> abs
1

source

LightQuery.@_ — Macro.

macro _(body)

Terser function syntax. The arguments are inside the body; the first argument is _, the second argument is __, etc.

julia> using LightQuery

julia> 1 |> @_(_ + 1)
2

julia> map(@_(__ - _), (1, 2), (2, 1))
(1, -1)

source