LightQuery.jl

LightQuery.jl

LightQuery.By

LightQuery.Group

LightQuery.LeftJoin

LightQuery.Name

LightQuery.Names

LightQuery.Peek

LightQuery.columns

LightQuery.gather

LightQuery.in_common

LightQuery.make_columns

LightQuery.named_tuple

LightQuery.order

LightQuery.over

LightQuery.remove

LightQuery.rename

LightQuery.rows

LightQuery.spread

LightQuery.transform

LightQuery.unzip

LightQuery.when

LightQuery.@>

LightQuery.@_

LightQuery.By — Type.

By(it, f)

Marks that it has been pre-sorted by the key f. For use with Group or LeftJoin.

source

LightQuery.Group — Method.

Group(b::By)

Group consecutive keys in b. Requires a presorted object (see By). Relies on the fact that iteration states can be converted to indices; thus, you might have to define LightQuery.state_to_index for unrecognized types (PR's welcome.)

julia> using LightQuery

julia> Group(By([1, 3, 2, 4], iseven)) |> first
false => [1, 3]

source

LightQuery.LeftJoin — Type.

LeftJoin(left::By, right::By)

For each value in left, look for a value with the same key in right. Requires both to be presorted (see By).

julia> using LightQuery

julia> LeftJoin(
            By([1, 2, 5, 6], identity),
            By([1, 3, 4, 6], identity)
       ) |> collect
4-element Array{Pair{Int64,Union{Missing, Int64}},1}:
 1 => 1
 2 => missing
 5 => missing
 6 => 6

source

LightQuery.Name — Type.

Name(x)

Force into the type domain. Can also be used as a function.

julia> using LightQuery

julia> Name(:a)((a = 1,))
1

source

LightQuery.Names — Method.

Names(names...)

Names in the type domain. Can be used to select columns.

julia> using LightQuery

julia> using Test: @inferred

julia> test(x) = Names(:a)(x);

julia> @inferred test((a = 1, b = 2.0))
(a = 1,)

source

LightQuery.Peek — Method.

Peek(it)

If it is over](@ref) Names, LightQuery will automatically show you a peek of the data in table form. However, not all iterators which yield NamedTuples will print this way; in order to get a peek of them, you need to explicitly use Peek. In same cases, will error if inference cannot detect the names. In this case, map a Names object over it first.

julia> using LightQuery

julia> rows((a = [1, 1.0], b = [2, 2.0])) |> Peek
|  :a |  :b |
| ---:| ---:|
| 1.0 | 2.0 |
| 1.0 | 2.0 |

source

LightQuery.columns — Method.

columns(it)

Inverse of rows.

julia> using LightQuery

julia> columns(rows((a = [1, 1.0], b = [2, 2.0])))
(a = [1.0, 1.0], b = [2.0, 2.0])

source

LightQuery.gather — Method.

gather(data, name, names...)

Gather all the data in names into a single name. Inverse of spread.

julia> using LightQuery

julia> using Test: @inferred

julia> test(x) = gather(x, :d, :a, :c);

julia> @inferred test((a = 1, b = 2.0, c = "c"))
(b = 2.0, d = (a = 1, c = "c"))

source

LightQuery.in_common — Method.

in_common(data1, data2)

Find the names in common between data1 and data2.

julia> using LightQuery

julia> in_common((a = 1, b = 2.0), (a = 1, c = 3.0))
(:a,)

source

LightQuery.make_columns — Method.

make_columns(it)

Collect into columns. See also columns. In same cases, will error if inference cannot detect the names. In this case, map a Names object over it first.

julia> using LightQuery

julia> using Test: @inferred

julia> (a = [1, 2], b = [1.0, 2.0]) |>
        rows |>
        make_columns
(a = [1, 2], b = [1.0, 2.0])

source

LightQuery.named_tuple — Method.

named_tuple(x)

Coerce to a named_tuple. For performance with working with arbitrary structs, explicitly define inlined propertynames.

julia> using LightQuery

julia> using Test: @inferred

julia> @inline Base.propertynames(p::Pair) = (:first, :second);

julia> @inferred named_tuple(:a => 1)
(first = :a, second = 1)

source

LightQuery.order — Method.

order(it, f; kwargs...)
order(it, f, condition; kwargs...)

Generalized sort. kwargs will be passed to sort!; see the documentation there for options. See By for a way to explicitly mark that an object has been sorted. Most performant it f is type stable, if not, consider using a condition to filter.

julia> using LightQuery

julia> order(["b", "a"], identity)
2-element view(::Array{String,1}, [2, 1]) with eltype String:
 "a"
 "b"

julia> order([missing, "a"], identity, !ismissing)
1-element view(::Array{Union{Missing, String},1}, [2]) with eltype Union{Missing, String}:
 "a"

source

LightQuery.over — Method.

over(it, f)

Lazy version of map, with the reverse argument order.

julia> using LightQuery

julia> over([1, 2], x -> x + 1) |> collect
2-element Array{Int64,1}:
 2
 3

source

LightQuery.remove — Method.

remove(data, names...)

Remove names. Inverse of transform.

julia> using LightQuery

julia> using Test: @inferred

julia> test(x) = remove(x, :b);

julia> @inferred test((a = 1, b = 2.0))
(a = 1,)

source

LightQuery.rename — Method.

rename(data; renames...)

Rename data. Use Name for type stability; constants don't propagate through keyword arguments.

julia> using LightQuery

julia> using Test: @inferred

julia> test(x) = rename(x, c = Name(:a));

julia> @inferred test((a = 1, b = 2.0))
(b = 2.0, c = 1)

julia> rename((a = 1, b = 2.0), c = :a)
(b = 2.0, c = 1)

source

LightQuery.rows — Method.

rows(n::NamedTuple)

Iterator over rows of a NamedTuple of names. Inverse of columns.

julia> using LightQuery

julia> using Test: @inferred

julia> rows((a = [1, 2], b = [2, 1]))
|  :a |  :b |
| ---:| ---:|
|   1 |   2 |
|   2 |   1 |

julia> rows((a = 1:6,))
Showing 4 of 6 rows
|  :a |
| ---:|
|   1 |
|   2 |
|   3 |
|   4 |

julia> rows((a = [1], b = [2], c = [3], d = [4], e = [5], f = [6], g = [7], h = [8]))
Showing 7 of 8 columns
|  :a |  :b |  :c |  :d |  :e |  :f |  :g |
| ---:| ---:| ---:| ---:| ---:| ---:| ---:|
|   1 |   2 |   3 |   4 |   5 |   6 |   7 |

source

LightQuery.spread — Method.

spread(data::NamedTuple, name)

Unnest nested data in name. Inverse of gather.

julia> using LightQuery

julia> using Test: @inferred

julia> test(x) = spread(x, :d);

julia> @inferred test((b = 2.0, d = (a = 1, c = "c")))
(b = 2.0, a = 1, c = "c")

source

LightQuery.transform — Method.

transform(data; assignments...)

Merge assignments into data.

julia> using LightQuery

julia> using Test: @inferred

julia> @inferred transform((a = 1, b = 2.0), c = "3")
(a = 1, b = 2.0, c = "3")

source

LightQuery.unzip — Method.

unzip(it, n)

Unzip an iterator it which returns tuples of length n.

julia> using LightQuery

julia> using Test: @inferred

julia> f(x) = (x, x + 1.0);

julia> unzip(over([1], f), 2)
([1], [2.0])

julia> unzip(over([1, missing], f), 2)
(Union{Missing, Int64}[1, missing], Union{Missing, Float64}[2.0, missing])

julia> unzip(zip([1], [1.0]), 2)
([1], [1.0])

julia> unzip([(1, 1.0)], 2)
([1], [1.0])

julia> unzip(over(when([1, missing, 2], x -> ismissing(x) || x > 1), f), 2)
(Union{Missing, Int64}[missing, 2], Union{Missing, Float64}[missing, 3.0])

source

LightQuery.when — Method.

when(it, f)

Lazy version of filter, with the reverse argument order.

julia> using LightQuery

julia> when([1, 2], x -> x > 1) |> collect
1-element Array{Int64,1}:
 2

source

LightQuery.@> — Macro.

macro >(body)

If body is in the form body_ |> tail_, call @_ on tail, and recur on body.

julia> using LightQuery

julia> @> 0 |> _ - 1 |> abs
1

source

LightQuery.@_ — Macro.

macro _(body)

Terser function syntax. The arguments are inside the body; the first argument is _, the second argument is __, etc.

julia> using LightQuery

julia> 1 |> @_(_ + 1)
2

julia> map(@_(__ - _), (1, 2), (2, 1))
(1, -1)

source

Tutorial

For an example of how to use this package, see the demo below, which follows the example here.

A copy of the flights data is included in the test folder of this package. I've reexported CSV from the CSV package for convenient IO. The data comes in as a data frame, but you can easily convert most objects to named tuples using named_tuple. I recollect to remove extra missing annotations unhelpfully provided by CSV.

julia> using LightQuery

julia> flight_columns =
        @> CSV.read("flights.csv", missingstring = "NA") |>
        named_tuple |>
        map(x -> collect(over(x, identity)), _);

As a named tuple, the data will be in a column-wise form; lazily convert it to rows.

julia> flights = rows(flight_columns)
Showing 7 of 19 columns
Showing 4 of 336776 rows
| :year | :month | :day | :dep_time | :sched_dep_time | :dep_delay | :arr_time |
| -----:| ------:| ----:| ---------:| ---------------:| ----------:| ---------:|
|  2013 |      1 |    1 |       517 |             515 |          2 |       830 |
|  2013 |      1 |    1 |       533 |             529 |          4 |       850 |
|  2013 |      1 |    1 |       542 |             540 |          2 |       923 |
|  2013 |      1 |    1 |       544 |             545 |         -1 |      1004 |

You can lazily filter data with when.

julia> using LightQuery

julia> @> flights |>
        when(_, @_ _.month == 1 && _.day == 1)
Showing 7 of 19 columns
Showing at most 4 rows
| :year | :month | :day | :dep_time | :sched_dep_time | :dep_delay | :arr_time |
| -----:| ------:| ----:| ---------:| ---------------:| ----------:| ---------:|
|  2013 |      1 |    1 |       517 |             515 |          2 |       830 |
|  2013 |      1 |    1 |       533 |             529 |          4 |       850 |
|  2013 |      1 |    1 |       542 |             540 |          2 |       923 |
|  2013 |      1 |    1 |       544 |             545 |         -1 |      1004 |

You might find it more efficient to do this columns-wise:

julia> using LightQuery

julia> @> flight_columns |>
        (_.month .== 1) .& (_.day .== 1) |>
        view(flights, _)
Showing 7 of 19 columns
Showing 4 of 842 rows
| :year | :month | :day | :dep_time | :sched_dep_time | :dep_delay | :arr_time |
| -----:| ------:| ----:| ---------:| ---------------:| ----------:| ---------:|
|  2013 |      1 |    1 |       517 |             515 |          2 |       830 |
|  2013 |      1 |    1 |       533 |             529 |          4 |       850 |
|  2013 |      1 |    1 |       542 |             540 |          2 |       923 |
|  2013 |      1 |    1 |       544 |             545 |         -1 |      1004 |

You can order the flights, using Names to select columns.

julia> using MappedArrays: mappedarray

julia> get_date = Names(:year, :month, :day)
Names{(:year, :month, :day)}()

julia> by_date =
        @> flights |>
        order(_, get_date)
Showing 7 of 19 columns
Showing 4 of 336776 rows
| :year | :month | :day | :dep_time | :sched_dep_time | :dep_delay | :arr_time |
| -----:| ------:| ----:| ---------:| ---------------:| ----------:| ---------:|
|  2013 |      1 |    1 |       517 |             515 |          2 |       830 |
|  2013 |      1 |    1 |       533 |             529 |          4 |       850 |
|  2013 |      1 |    1 |       542 |             540 |          2 |       923 |
|  2013 |      1 |    1 |       544 |             545 |         -1 |      1004 |

You can also pass in keyword arguments to sort! via order, like rev = true. Note also that arr_delay includes missing data. For performance, I'm adding a condition to remove the missing data. This will be a common pattern.

julia> @> flights |>
        order(_, Names(:arr_delay), (@_ !ismissing(_.arr_delay)), rev = true)
Showing 7 of 19 columns
Showing 4 of 327346 rows
| :year | :month | :day | :dep_time | :sched_dep_time | :dep_delay | :arr_time |
| -----:| ------:| ----:| ---------:| ---------------:| ----------:| ---------:|
|  2013 |      1 |    9 |       641 |             900 |       1301 |      1242 |
|  2013 |      6 |   15 |      1432 |            1935 |       1137 |      1607 |
|  2013 |      1 |   10 |      1121 |            1635 |       1126 |      1239 |
|  2013 |      9 |   20 |      1139 |            1845 |       1014 |      1457 |

In the original column-wise form, you can select with Names. Then, use rows again to print.

julia> get_date(flight_columns) |>
          rows
Showing 4 of 336776 rows
| :year | :month | :day |
| -----:| ------:| ----:|
|  2013 |      1 |    1 |
|  2013 |      1 |    1 |
|  2013 |      1 |    1 |
|  2013 |      1 |    1 |

You can also remove columns in the column-wise form.

julia> @> flight_columns |>
        remove(_, :year, :month, :day) |>
        rows
Showing 7 of 16 columns
Showing 4 of 336776 rows
| :dep_time | :sched_dep_time | :dep_delay | :arr_time | :sched_arr_time | :arr_delay | :carrier |
| ---------:| ---------------:| ----------:| ---------:| ---------------:| ----------:| --------:|
|       517 |             515 |          2 |       830 |             819 |         11 |       UA |
|       533 |             529 |          4 |       850 |             830 |         20 |       UA |
|       542 |             540 |          2 |       923 |             850 |         33 |       AA |
|       544 |             545 |         -1 |      1004 |            1022 |        -18 |       B6 |

You can also rename columns. Because constants (currently) do not propagate through keyword arguments in Julia, it's smart to wrap column names with Name.

julia> @> flight_columns |>
        rename(_, tail_num = Name(:tailnum)) |>
        rows
Showing 7 of 19 columns
Showing 4 of 336776 rows
| :year | :month | :day | :dep_time | :sched_dep_time | :dep_delay | :arr_time |
| -----:| ------:| ----:| ---------:| ---------------:| ----------:| ---------:|
|  2013 |      1 |    1 |       517 |             515 |          2 |       830 |
|  2013 |      1 |    1 |       533 |             529 |          4 |       850 |
|  2013 |      1 |    1 |       542 |             540 |          2 |       923 |
|  2013 |      1 |    1 |       544 |             545 |         -1 |      1004 |

You can add new columns with transform. If you want to refer to previous columns, you'll have to transform twice. You can do this row-wise. Note that I've added a second over simply to specify the names of the items. This will be a common pattern for when inference can't keep track of names.

julia> @> flights |>
        over(_, @_ @> transform(_,
            gain = _.arr_delay - _.dep_delay,
            speed = _.distance / _.air_time * 60
        ) |> transform(_,
            gain_per_hour = _.gain ./ (_.air_time / 60)
        )) |>
        over(_, Names(propertynames(flight_columns)..., :gain, :speed,
            :gain_per_hour))
Showing 7 of 22 columns
Showing 4 of 336776 rows
| :year | :month | :day | :dep_time | :sched_dep_time | :dep_delay | :arr_time |
| -----:| ------:| ----:| ---------:| ---------------:| ----------:| ---------:|
|  2013 |      1 |    1 |       517 |             515 |          2 |       830 |
|  2013 |      1 |    1 |       533 |             529 |          4 |       850 |
|  2013 |      1 |    1 |       542 |             540 |          2 |       923 |
|  2013 |      1 |    1 |       544 |             545 |         -1 |      1004 |

You can also do the same thing column-wise:

julia> @> flight_columns |>
        transform(_,
            gain = _.arr_delay .- _.dep_delay,
            speed = _.distance ./ _.air_time .* 60
        ) |>
        transform(_,
            gain_per_hour = _.gain ./ (_.air_time / 60)
        ) |>
        rows
Showing 7 of 22 columns
Showing 4 of 336776 rows
| :year | :month | :day | :dep_time | :sched_dep_time | :dep_delay | :arr_time |
| -----:| ------:| ----:| ---------:| ---------------:| ----------:| ---------:|
|  2013 |      1 |    1 |       517 |             515 |          2 |       830 |
|  2013 |      1 |    1 |       533 |             529 |          4 |       850 |
|  2013 |      1 |    1 |       542 |             540 |          2 |       923 |
|  2013 |      1 |    1 |       544 |             545 |         -1 |      1004 |

You can't summarize ungrouped data, but you can just directly access columns:

julia> using Statistics: mean;

julia> mean(skipmissing(flight_columns.dep_delay))
12.639070257304708

I don't provide a export a sample function here, but StatsBase does.

Grouping here works differently than in dplyr:

You can only Group sorted data. To let Julia know that the data has been sorted, you need to explicitly wrap the data with By.

Groups return a pair, key => sub-data-frame. So:

julia> by_tailnum =
        @> flights |>
        order(_, Names(:tailnum), (@_ !ismissing(_.tailnum))) |>
        Group(By(_, Names(:tailnum)));

julia> first(by_tailnum)
(tailnum = "D942DN",) => Showing 7 of 19 columns
| :year | :month | :day | :dep_time | :sched_dep_time | :dep_delay | :arr_time |
| -----:| ------:| ----:| ---------:| ---------------:| ----------:| ---------:|
|  2013 |      2 |   11 |      1508 |            1400 |         68 |      1807 |
|  2013 |      3 |   23 |      1340 |            1300 |         40 |      1638 |
|  2013 |      3 |   24 |       859 |             835 |         24 |      1142 |
|  2013 |      7 |    5 |      1253 |            1259 |         -6 |      1518 |

You can play around with the pair structure of groups to coerce it to the shape you want. Notice that I'm collecting after the Group (for performance). I'm also explicitly calling Peek here; Julia can only guess that you want a peek if over was used with a Names object.

julia> @> by_tailnum |>
        collect |>
        over(_, @_ transform(_.first,
            count = length(_.second),
            distance = columns(_.second).distance |> mean,
            delay = columns(_.second).arr_delay |> skipmissing |> mean
        )) |>
        Peek
Showing 4 of 4043 rows
| :tailnum | :count |         :distance |             :delay |
| --------:| ------:| -----------------:| ------------------:|
|   D942DN |      4 |             854.5 |               31.5 |
|   N0EGMQ |    371 |  676.188679245283 |  9.982954545454545 |
|   N10156 |    153 | 757.9477124183006 | 12.717241379310344 |
|   N102UW |     48 |           535.875 |             2.9375 |

For the n-distinct example, I've switched things around to be just a smidge more efficient. This example shows how calling make_columns and then rows is sometimes necessary to trigger eager evaluation. I've also defined a count_flights function because we'll be using it again.

julia> count_flights(x) = over(x, @_ transform(_.first,
            flights = length(_.second)
        ));

julia> @> flights |>
        order(_, Names(:dest, :tailnum), (@_ !ismissing(_.tailnum))) |>
        Group(By(_, Names(:dest, :tailnum))) |>
        collect |>
        count_flights |>
        make_columns |>
        rows |>
        Group(By(_, Names(:dest))) |>
        over(_, @_ transform(_.first,
            planes = length(_.second),
            flights = sum(columns(_.second).flights)
        )) |>
        Peek
Showing at most 4 rows
| :dest | :planes | :flights |
| -----:| -------:| --------:|
|   ABQ |     108 |      254 |
|   ACK |      58 |      265 |
|   ALB |     172 |      439 |
|   ANC |       6 |        8 |

Of course, you can group repeatedly. You don't have to reorder each time if you do this.

julia> grouped_by_date =
        @> by_date |>
        Group(By(_, get_date)) |>
        collect;

julia> per_day =
        @> grouped_by_date |>
        count_flights |>
        make_columns |>
        rows
Showing 4 of 365 rows
| :year | :month | :day | :flights |
| -----:| ------:| ----:| --------:|
|  2013 |      1 |    1 |      842 |
|  2013 |      1 |    2 |      943 |
|  2013 |      1 |    3 |      914 |
|  2013 |      1 |    4 |      915 |

julia> sum_flights(x) = over(x, @_ transform(_.first,
            flights = sum(columns(_.second).flights)
        ));

julia> per_month =
        @> per_day |>
        Group(By(_, Names(:year, :month))) |>
        collect |>
        sum_flights |>
        make_columns |>
        rows
Showing 4 of 12 rows
| :year | :month | :flights |
| -----:| ------:| --------:|
|  2013 |      1 |    27004 |
|  2013 |      2 |    24951 |
|  2013 |      3 |    28834 |
|  2013 |      4 |    28330 |

julia> @> per_month |>
          Group(By(_, Names(:year))) |>
          sum_flights |>
          Peek
Showing at most 4 rows
| :year | :flights |
| -----:| --------:|
|  2013 |   336776 |

Here's the example in the dplyr docs for piping:

julia> @> grouped_by_date |>
        over(_, @_ transform(_.first,
            arr = columns(_.second).arr_delay |> skipmissing |> mean,
            dep = columns(_.second).dep_delay |> skipmissing |> mean
        )) |>
        when(_, @_ _.arr > 30 || _.dep > 30) |>
        over(_, Names(:year, :month, :day, :arr, :dep))
Showing at most 4 rows
| :year | :month | :day |               :arr |               :dep |
| -----:| ------:| ----:| ------------------:| ------------------:|
|  2013 |      1 |   16 |  34.24736225087925 | 24.612865497076022 |
|  2013 |      1 |   31 | 32.602853745541026 | 28.658362989323845 |
|  2013 |      2 |   11 |  36.29009433962264 |  39.07359813084112 |
|  2013 |      2 |   27 |  31.25249169435216 |  37.76327433628319 |

Two table verbs

I'm following the example here.

Again, for inference reasons, natural joins won't work. I only provide one join at the moment, but it's super efficient. Let's start by reading in airlines and letting julia know that it's already sorted by :carrier.

julia> airlines =
        @> CSV.read("airlines.csv", missingstring = "NA") |>
        named_tuple |>
        map(x -> collect(over(x, identity)), _) |>
        rows |>
        By(_, Names(:carrier));

If we want to join this data into the flights data, here's what we do. LeftJoin requires not only presorted but unique keys. Of course, there are multiple flights from the same airline, so we need to group first. Then, we tell Julia that the groups are themselves sorted (by the first item, the key). Finally we can join in the airline data. But the results are a bit tricky. Let's take a look at the first item. Just like the dplyr manual, I'm only using a few of the columns from flights for demonstration.

julia> flight2_columns =
        @> flight_columns |>
        Names(:year, :month, :day, :hour, :origin, :dest, :tailnum, :carrier)(_);

julia> flights2 = rows(flight2_columns)
Showing 7 of 8 columns
Showing 4 of 336776 rows
| :year | :month | :day | :hour | :origin | :dest | :tailnum |
| -----:| ------:| ----:| -----:| -------:| -----:| --------:|
|  2013 |      1 |    1 |     5 |     EWR |   IAH |   N14228 |
|  2013 |      1 |    1 |     5 |     LGA |   IAH |   N24211 |
|  2013 |      1 |    1 |     5 |     JFK |   MIA |   N619AA |
|  2013 |      1 |    1 |     5 |     JFK |   BQN |   N804JB |

julia> airline_join =
        @> flights2 |>
        order(_, Names(:carrier)) |>
        Group(By(_, Names(:carrier))) |>
        By(_, first) |>
        LeftJoin(_, airlines);

julia> first(airline_join)
((carrier = "9E",)=>Showing 7 of 8 columns
Showing 4 of 18460 rows
| :year | :month | :day | :hour | :origin | :dest | :tailnum |
| -----:| ------:| ----:| -----:| -------:| -----:| --------:|
|  2013 |      1 |    1 |     8 |     JFK |   MSP |   N915XJ |
|  2013 |      1 |    1 |    15 |     JFK |   IAD |   N8444F |
|  2013 |      1 |    1 |    14 |     JFK |   BUF |   N920XJ |
|  2013 |      1 |    1 |    15 |     JFK |   SYR |   N8409N |
) => (carrier = "9E", name = "Endeavor Air Inc.")

We end up getting a group and subframe on the left, and a row on the right.

If you want to collect your results into a flat new dataframe, you need to do a bit of surgery, including making use of flatten (which I reexport from Base). We also need to make a fake row to insert on the right in case we can't find a match.

julia> @> airline_join |>
        over(_, @_ over(_.first.second, x -> merge(x, _.second))) |>
        flatten |>
        over(_, Names(:year, :month, :day, :hour, :origin, :dest, :tailnum,
            :carrier, :name))
Showing 7 of 9 columns
Showing at most 4 rows
| :year | :month | :day | :hour | :origin | :dest | :tailnum |
| -----:| ------:| ----:| -----:| -------:| -----:| --------:|
|  2013 |      1 |    1 |     8 |     JFK |   MSP |   N915XJ |
|  2013 |      1 |    1 |    15 |     JFK |   IAD |   N8444F |
|  2013 |      1 |    1 |    14 |     JFK |   BUF |   N920XJ |
|  2013 |      1 |    1 |    15 |     JFK |   SYR |   N8409N |

Let's keep going in the examples. I'm going to read in the weather data, and let Julia know that it has already been sorted.

julia> weather_columns =
        @> CSV.read( "weather.csv", missingstring = "NA") |>
        named_tuple |>
        map(x -> collect(over(x, identity)), _);

julia> weather =
        @> weather_columns |>
        rows |>
        By(_, Names(:origin, :year, :month, :day, :hour));

Unfortunately, we have to deal with another problem: there's gaps in the weather data. We need to make a missing row of data. I'm also going to use union to get all the names together.

julia> const missing_weather =
        @> weather_columns |>
        remove(_, :origin, :year, :month, :day, :hour) |>
        map(x -> missing, _);

julia> weather_join =
        @> flights2 |>
        order(_, Names(:origin, :year, :month, :day, :hour)) |>
        Group(By(_, Names(:origin, :year, :month, :day, :hour))) |>
        By(_, first) |>
        LeftJoin(_, weather);

julia> flight2_and_weather_names = Names(union(
            propertynames(flight2_columns),
            propertynames(weather_columns)
        )...);

julia> @> weather_join |>
        over(_, @_ over(_.first.second, x -> merge(x, coalesce(_.second, missing_weather)))) |>
        flatten |>
        over(_, flight2_and_weather_names)
Showing 7 of 18 columns
Showing at most 4 rows
| :year | :month | :day | :hour | :origin | :dest | :tailnum |
| -----:| ------:| ----:| -----:| -------:| -----:| --------:|
|  2013 |      1 |    1 |     5 |     EWR |   IAH |   N14228 |
|  2013 |      1 |    1 |     5 |     EWR |   ORD |   N39463 |
|  2013 |      1 |    1 |     6 |     EWR |   FLL |   N516JB |
|  2013 |      1 |    1 |     6 |     EWR |   SFO |   N53441 |

Of course, if you wanted, you could just remove the rows with missing weather data, essentially doing an inner join:

julia> @> weather_join |>
        when(_, @_ _.second !== missing) |>
        over(_, @_ over(_.first.second, x -> merge(x, missing_weather))) |>
        flatten |>
        over(_, flight2_and_weather_names)
Showing 7 of 18 columns
Showing at most 4 rows
| :year | :month | :day | :hour | :origin | :dest | :tailnum |
| -----:| ------:| ----:| -----:| -------:| -----:| --------:|
|  2013 |      1 |    1 |     5 |     EWR |   IAH |   N14228 |
|  2013 |      1 |    1 |     5 |     EWR |   ORD |   N39463 |
|  2013 |      1 |    1 |     6 |     EWR |   FLL |   N516JB |
|  2013 |      1 |    1 |     6 |     EWR |   SFO |   N53441 |