Tutorial

Tutorial

I'm going to use the flights data from the dplyr tutorial. This data is in the test folder of this package; I created it with the following R code:

library(nycflights13)
setwd("C:/Users/hp/.julia/dev/LightQuery/test")
write.csv(airports, "airports.csv", na = "", row.names = FALSE)
write.csv(flights, "flights.csv", na = "", row.names = FALSE)

Let's import the tools we need. I'm pulling in from tools from Dates, TimeZones, and Unitful.

julia> using LightQuery

julia> using Dates: DateTime, Day

julia> import Dates: Minute

julia> Minute(::Missing) = missing;

julia> using Unitful: mi, °, ft

julia> using TimeZones: ZonedDateTime, Class, VariableTimeZone

julia> import TimeZones: TimeZone

julia> TimeZone(::Missing, ::Class) = missing;

I re-export CSV for input-output. See the documentation there for information about CSV.File.

julia> airports_file = CSV.File("airports.csv",
            allowmissing = :auto,
            missingstrings = ["", "\\N"]
        )
CSV.File("airports.csv", rows=1458):
Tables.Schema:
 :faa    String
 :name   String
 :lat    Float64
 :lon    Float64
 :alt    Int64
 :tz     Int64
 :dst    String
 :tzone  Union{Missing, String}

We use row_type to see what type rows this will return:

isconcretetype

julia> const Airport = row_type(airports_file)
Tuple{Tuple{LightQuery.Name{:faa},String},Tuple{LightQuery.Name{:name},String},Tuple{LightQuery.Name{:lat},Float64},Tuple{LightQuery.Name{:lon},Float64},Tuple{LightQuery.Name{:alt},Int64},Tuple{LightQuery.Name{:tz},Int64},Tuple{LightQuery.Name{:dst},String},Tuple{LightQuery.Name{:tzone},T} where T<:Union{Missing, String}}

Let's take a look at the first row. Use named_tuple to coerce a CSV.Row to a named tuple. I use the chaining macro @> to chain calls together.

julia> airport =
        @> airports_file |>
        first |>
        named_tuple(_)::Airport
((faa, "04G"), (name, "Lansdowne Airport"), (lat, 41.1304722), (lon, -80.6195833), (alt, 1044), (tz, -5), (dst, "A"), (tzone, "America/New_York"))

You'll probably notice this doesn't look like the NamedTuples you're probably familiar with. I've created a homemade version of NamedTuples.

As a start, I want to rename so that I understand what the columns mean. I use the [@name] macro to switch to my version of named tuples.

julia> airport =
        @name @> airport |>
        rename(_,
            airport_code = :faa,
            latitude = :lat,
            longitude = :lon,
            altitude = :alt,
            time_zone_offset = :tz,
            daylight_savings = :dst,
            time_zone = :tzone
        )
((name, "Lansdowne Airport"), (airport_code, "04G"), (latitude, 41.1304722), (longitude, -80.6195833), (altitude, 1044), (time_zone_offset, -5), (daylight_savings, "A"), (time_zone, "America/New_York"))

Let's create a proper TimeZone. Note the data contains some LEGACY timezones. Note the type annotation: TimeZone is unstable without it.

julia> const time_zone_classes = Class(:STANDARD) | Class(:LEGACY);

julia> airport =
        @name @> airport |>
        transform(_,
            time_zone = TimeZone(_.time_zone, time_zone_classes)::Union{VariableTimeZone, Missing}
        )
((name, "Lansdowne Airport"), (airport_code, "04G"), (latitude, 41.1304722), (longitude, -80.6195833), (altitude, 1044), (time_zone_offset, -5), (daylight_savings, "A"), (time_zone, tz"America/New_York"))

Now that we have a true timezone, we can remove all data that is contingent on timezone.

julia> airport =
        @name @> airport |>
        remove(_,
            :time_zone_offset,
            :daylight_savings
        )
((name, "Lansdowne Airport"), (airport_code, "04G"), (latitude, 41.1304722), (longitude, -80.6195833), (altitude, 1044), (time_zone, tz"America/New_York"))

Let's also add proper units to our variables.

julia> airport =
        @name @> airport |>
        transform(_,
            latitude = _.latitude * °,
            longitude = _.longitude * °,
            altitude = _.altitude * ft
        )
((name, "Lansdowne Airport"), (airport_code, "04G"), (time_zone, tz"America/New_York"), (latitude, 41.1304722°), (longitude, -80.6195833°), (altitude, 1044 ft))

Let's put it all together.

julia> function process_airport(row)
            @name @> row |>
            named_tuple(_)::Airport |>
            rename(_,
                airport_code = :faa,
                latitude = :lat,
                longitude = :lon,
                altitude = :alt,
                time_zone_offset = :tz,
                daylight_savings = :dst,
                time_zone = :tzone
            ) |>
            transform(_,
                time_zone = TimeZone(_.time_zone, time_zone_classes)::Union{VariableTimeZone, Missing},
                latitude = _.latitude * °,
                longitude = _.longitude * °,
                altitude = _.altitude * ft
            ) |>
            remove(_,
                :time_zone_offset,
                :daylight_savings
            )
        end;

I use over to lazily map.

julia> airports =
        @> airports_file |>
        over(_, process_airport);

When it comes time to collect, I'm calling make_columns then rows. It makes sense to store this data column-wise. This is because there are multiple columns that might contain missing data.

julia> airports =
        airports |>
        make_columns |>
        rows;

We can use Peek to get a look at the data.

julia> Peek(airports)
Showing 4 of 1458 rows
|                          name | airport_code |                      time_zone |    latitude |    longitude | altitude |
| -----------------------------:| ------------:| ------------------------------:| -----------:| ------------:| --------:|
|             Lansdowne Airport |          04G | America/New_York (UTC-5/UTC-4) | 41.1304722° | -80.6195833° |  1044 ft |
| Moton Field Municipal Airport |          06A |  America/Chicago (UTC-6/UTC-5) | 32.4605722° | -85.6800278° |   264 ft |
|           Schaumburg Regional |          06C |  America/Chicago (UTC-6/UTC-5) | 41.9893408° | -88.1012428° |   801 ft |
|               Randall Airport |          06N | America/New_York (UTC-5/UTC-4) |  41.431912° | -74.3915611° |   523 ft |

I'll also make sure the airports are indexed by their code so we can access them quickly.

julia> const indexed_airports =
        @name @> airports |>
        indexed(_, :airport_code);

julia> indexed_airports["JFK"]
((name, "John F Kennedy Intl"), (airport_code, "JFK"), (time_zone, tz"America/New_York"), (latitude, 40.639751°), (longitude, -73.778925°), (altitude, 13 ft))

That was just the warm-up. Now let's get started working on the flights data.

julia> flights_file = CSV.File("flights.csv", allowmissing = :auto)
CSV.File("flights.csv", rows=336776):
Tables.Schema:
 :year            Int64
 :month           Int64
 :day             Int64
 :dep_time        Union{Missing, Int64}
 :sched_dep_time  Int64
 :dep_delay       Union{Missing, Int64}
 :arr_time        Union{Missing, Int64}
 :sched_arr_time  Int64
 :arr_delay       Union{Missing, Int64}
 :carrier         String
 :flight          Int64
 :tailnum         Union{Missing, String}
 :origin          String
 :dest            String
 :air_time        Union{Missing, Int64}
 :distance        Int64
 :hour            Int64
 :minute          Int64
 :time_hour       String

julia> const Flight = row_type(flights_file);

julia> flight =
        @name @> flights_file |>
        first |>
        named_tuple(_)::Flight |>
        rename(_,
            departure_time = :dep_time,
            scheduled_departure_time = :sched_dep_time,
            departure_delay = :dep_delay,
            arrival_time = :arr_time,
            scheduled_arrival_time = :sched_arr_time,
            arrival_delay = :arr_delay,
            tail_number = :tailnum,
            destination = :dest
        )
((year, 2013), (month, 1), (day, 1), (carrier, "UA"), (flight, 1545), (origin, "EWR"), (air_time, 227), (distance, 1400), (hour, 5), (minute, 15), (time_hour, "2013-01-01 05:00:00"), (departure_time, 517), (scheduled_departure_time, 515), (departure_delay, 2), (arrival_time, 830), (scheduled_arrival_time, 819), (arrival_delay, 11), (tail_number, "N14228"), (destination, "IAH"))

We can use our airports data to make datetimes with timezones.

julia> scheduled_departure_time = ZonedDateTime(
            DateTime(flight.year, flight.month, flight.day, flight.hour, flight.minute),
            indexed_airports[flight.origin].time_zone
        )
2013-01-01T05:15:00-05:00

Note the scheduled arrival time is 818. This means 8:18. We can use divrem(_, 100) to split it up. Not all destinations are not in the flights dataset. If it was an overnight flight, add a day to the arrival time.

julia> if haskey(indexed_airports, flight.destination)
            maybe_arrival_time = ZonedDateTime(
                DateTime(flight.year, flight.month, flight.day, divrem(flight.scheduled_arrival_time, 100)...),
                indexed_airports[flight.destination].time_zone
            )
            scheduled_arrival_time =
                if maybe_arrival_time < scheduled_departure_time
                    maybe_arrival_time + Day(1)
                else
                    maybe_arrival_time
                end
        else
            scheduled_arrival_time = missing
        end
2013-01-01T08:19:00-06:00

Let's put it all together.

julia> function process_flight(row)
            flight =
                @name @> row |>
                named_tuple(_)::Flight |>
                rename(_,
                    departure_time = :dep_time,
                    scheduled_departure_time = :sched_dep_time,
                    departure_delay = :dep_delay,
                    arrival_time = :arr_time,
                    scheduled_arrival_time = :sched_arr_time,
                    arrival_delay = :arr_delay,
                    tail_number = :tailnum,
                    destination = :dest
                )
            scheduled_departure_time = ZonedDateTime(
                DateTime(flight.year, flight.month, flight.day, flight.hour, flight.minute),
                indexed_airports[flight.origin].time_zone
            )
            if haskey(indexed_airports, flight.destination)
                maybe_arrival_time = ZonedDateTime(
                    DateTime(flight.year, flight.month, flight.day, divrem(flight.scheduled_arrival_time, 100)...),
                    indexed_airports[flight.destination].time_zone
                )
                scheduled_arrival_time =
                    if maybe_arrival_time < scheduled_departure_time
                        maybe_arrival_time + Day(1)
                    else
                        maybe_arrival_time
                    end
            else
                scheduled_arrival_time = missing
            end
            @name @> flight |>
            remove(_,
                :year,
                :month,
                :day,
                :hour,
                :minute,
                :time_hour,
                :departure_time,
                :arrival_time
            ) |>
            transform(_,
                scheduled_departure_time = scheduled_departure_time,
                scheduled_arrival_time = scheduled_arrival_time,
                air_time = Minute(_.air_time),
                distance = _.distance * mi,
                departure_delay = Minute(_.departure_delay),
                arrival_delay = Minute(_.arrival_delay)
            )
        end;

julia> flights =
        @> flights_file |>
        over(_, process_flight) |>
        make_columns |>
        rows;

julia> Peek(flights)
Showing 4 of 336776 rows
| carrier | flight | origin | tail_number | destination |  scheduled_departure_time |    scheduled_arrival_time |    air_time | distance | departure_delay | arrival_delay |
| -------:| ------:| ------:| -----------:| -----------:| -------------------------:| -------------------------:| -----------:| --------:| ---------------:| -------------:|
|      UA |   1545 |    EWR |      N14228 |         IAH | 2013-01-01T05:15:00-05:00 | 2013-01-01T08:19:00-06:00 | 227 minutes |  1400 mi |       2 minutes |    11 minutes |
|      UA |   1714 |    LGA |      N24211 |         IAH | 2013-01-01T05:29:00-05:00 | 2013-01-01T08:30:00-06:00 | 227 minutes |  1416 mi |       4 minutes |    20 minutes |
|      AA |   1141 |    JFK |      N619AA |         MIA | 2013-01-01T05:40:00-05:00 | 2013-01-01T08:50:00-05:00 | 160 minutes |  1089 mi |       2 minutes |    33 minutes |
|      B6 |    725 |    JFK |      N804JB |         BQN | 2013-01-01T05:45:00-05:00 |                   missing | 183 minutes |  1576 mi |       -1 minute |   -18 minutes |

Theoretically, the distances between two airports is always the same. Let's make sure this is also the case in our data. First, order by origin, destination, and distance. Then Group By the same variables.

julia> paths_grouped =
        @name @> flights |>
        order(_, (:origin, :destination, :distance)) |>
        Group(By(_, (:origin, :destination, :distance)));

Each Group will contain a key and value

julia> path = first(paths_grouped);

julia> key(path)
((origin, "EWR"), (destination, "ALB"), (distance, 143 mi))

julia> value(path) |> Peek
Showing 4 of 439 rows
| carrier | flight | origin | tail_number | destination |  scheduled_departure_time |    scheduled_arrival_time |   air_time | distance | departure_delay | arrival_delay |
| -------:| ------:| ------:| -----------:| -----------:| -------------------------:| -------------------------:| ----------:| --------:| ---------------:| -------------:|
|      EV |   4112 |    EWR |      N13538 |         ALB | 2013-01-01T13:17:00-05:00 | 2013-01-01T14:23:00-05:00 | 33 minutes |   143 mi |      -2 minutes |   -10 minutes |
|      EV |   3260 |    EWR |      N19554 |         ALB | 2013-01-01T16:21:00-05:00 | 2013-01-01T17:24:00-05:00 | 36 minutes |   143 mi |      34 minutes |    40 minutes |
|      EV |   4170 |    EWR |      N12540 |         ALB | 2013-01-01T20:04:00-05:00 | 2013-01-01T21:12:00-05:00 | 31 minutes |   143 mi |      52 minutes |    44 minutes |
|      EV |   4316 |    EWR |      N14153 |         ALB | 2013-01-02T13:27:00-05:00 | 2013-01-02T14:33:00-05:00 | 33 minutes |   143 mi |       5 minutes |   -14 minutes |

At this point, we don't need any of the value data. All we need is the key.

julia> paths =
        @> paths_grouped |>
        over(_, key) |>
        make_columns |>
        rows;

julia> Peek(paths)
Showing 4 of 226 rows
| origin | destination | distance |
| ------:| -----------:| --------:|
|    EWR |         ALB |   143 mi |
|    EWR |         ANC |  3370 mi |
|    EWR |         ATL |   746 mi |
|    EWR |         AUS |  1504 mi |

Notice the data is already sorted by origin and destination, so that for our second Group, we don't need to order first.

julia> distinct_distances =
        @name @> paths |>
        Group(By(_, (:origin, :destination))) |>
        over(_, @_ transform(key(_),
            number = length(value(_))
        ));

julia> Peek(distinct_distances)
Showing at most 4 rows
| origin | destination | number |
| ------:| -----------:| ------:|
|    EWR |         ALB |      1 |
|    EWR |         ANC |      1 |
|    EWR |         ATL |      1 |
|    EWR |         AUS |      1 |

Let's see when there are multiple distances for the same path:

julia> @> distinct_distances |>
        when(_, @_ _.number != 1) |>
        Peek
Showing at most 4 rows
| origin | destination | number |
| ------:| -----------:| ------:|
|    EWR |         EGE |      2 |
|    JFK |         EGE |      2 |

That's strange. What's up with the EGE airport? Let's take a Peek.

julia> @> flights |>
        when(_, @_ _.destination == "EGE") |>
        Peek
Showing at most 4 rows
| carrier | flight | origin | tail_number | destination |  scheduled_departure_time |    scheduled_arrival_time |    air_time | distance | departure_delay | arrival_delay |
| -------:| ------:| ------:| -----------:| -----------:| -------------------------:| -------------------------:| -----------:| --------:| ---------------:| -------------:|
|      UA |   1597 |    EWR |      N27733 |         EGE | 2013-01-01T09:28:00-05:00 | 2013-01-01T12:20:00-07:00 | 287 minutes |  1726 mi |      -2 minutes |    13 minutes |
|      AA |    575 |    JFK |      N5DRAA |         EGE | 2013-01-01T17:00:00-05:00 | 2013-01-01T19:50:00-07:00 | 280 minutes |  1747 mi |      -5 minutes |     3 minutes |
|      UA |   1597 |    EWR |      N24702 |         EGE | 2013-01-02T09:28:00-05:00 | 2013-01-02T12:20:00-07:00 | 261 minutes |  1726 mi |        1 minute |     3 minutes |
|      AA |    575 |    JFK |      N631AA |         EGE | 2013-01-02T17:00:00-05:00 | 2013-01-02T19:50:00-07:00 | 260 minutes |  1747 mi |       5 minutes |    16 minutes |

Looks (to me) like two different sources are reporting different info about the same flight.

Interface

Macros

LightQuery.@_ — Macro.

macro _(body)

Terser function syntax. The arguments are inside the body; the first argument is _, the second argument is __, etc.

julia> using LightQuery

julia> (@_ _ + 1)(1)
2

julia> map((@_ __ - _), (1, 2), (2, 1))
(1, -1)

source

LightQuery.@> — Macro.

macro >(body)

If body is in the form body_ |> tail_, call @_ on tail, and recur on body.

julia> using LightQuery

julia> @> 0 |> _ - 1 |> abs
1

source

LightQuery.@name — Macro.

macro name(something)

Replace raw symbols with typed names, and named tuples with typed named tuples. Typed names can be used with getindex (and will keep names) or as a function.

julia> using LightQuery

julia> @name :a
a

julia> data = @name (a = 1, b = 2, c = 3)
((a, 1), (b, 2), (c, 3))

julia> @name data[:a]
1

julia> @name (:a)(data)
1

julia> data.a
1

julia> @name data[(:a, :b)]
((a, 1), (b, 2))

julia> @name (1, 2)[(:a, :b)]
((a, 1), (b, 2))

julia> @name (:a, :b)(data)
((a, 1), (b, 2))

julia> @name (:a, :b)((1, 2))
((a, 1), (b, 2))

source

Columns

LightQuery.named_tuple — Function.

named_tuple(anything)

Coerce anything to a typed named tuple. For performance with structs, define and @inline propertynames.

julia> using LightQuery

julia> data = @name ((a = 1, b = 2))
((a, 1), (b, 2))

julia> struct MyType
            a::Int
            b::Int
        end

julia> @inline Base.propertynames(::MyType) = (:a, :b);

julia> named_tuple(MyType(1, 2))
((a, 1), (b, 2))

source

LightQuery.rename — Function.

rename(data, assignments...)

Rename data.

julia> using LightQuery

julia> @name rename((a = 1, b = 2), c = :a)
((b, 2), (c, 1))

source

LightQuery.transform — Function.

transform(data, assignments...)

Merge assignments into data, overwriting old values.

julia> using LightQuery

julia> @name transform((a = 1, b = 2), a = 3)
((b, 2), (a, 3))

source

LightQuery.remove — Function.

remove(data, names...)

Remove names from data.

julia> using LightQuery

julia> @name remove((a = 1, b = 2, c = 3), :b)
((a, 1), (c, 3))

source

LightQuery.gather — Function.

gather(data; assignments...)

For each key => value pair in assignments, gather the names in value into a single key. Inverse of spread.

julia> using LightQuery

julia> @name gather((a = 1, b = 2, c = 3), d = (:a, :c))
((b, 2), (d, ((a, 1), (c, 3))))

source

LightQuery.spread — Function.

spread(data, names...)

Unnest nested name in names. Inverse of gather.

julia> using LightQuery

julia> @name spread((b = 2, d = (a = 1, c = 3)), :d)
((b, 2), (a, 1), (c, 3))

source

LightQuery.row_type — Function.

row_type(file::CSV.File)

Get the row_type of a CSV file.

julia> using LightQuery

julia> row_type(CSV.File("test.csv"))
Tuple{Tuple{LightQuery.Name{:a},T} where T<:Union{Missing, Int64},Tuple{LightQuery.Name{:b},T} where T<:Union{Missing, Float64}}

source

Rows

LightQuery.unzip — Function.

unzip(it, n)

Unzip an iterator it which returns tuples of length n. Use Val(n) to guarantee type stability.

julia> using LightQuery

julia> unzip([(1, 1.0), (2, 2.0)], 2)
([1, 2], [1.0, 2.0])

julia> unzip([(1, 1.0), (2, 2.0)], Val(2))
([1, 2], [1.0, 2.0])

source

LightQuery.Enumerated — Type.

Enumerated{It}

Relies on the fact that iteration states can be converted to indices; thus, you might have to define LightQuery.state_to_index for unrecognized types. Ignores some iterators like Filter.

julia> using LightQuery

julia> collect(Enumerated(when([4, 3, 2, 1], iseven)))
2-element Array{Tuple{Int64,Int64},1}:
 (1, 4)
 (3, 2)

source

LightQuery.over — Function.

over(it, call)

Lazy map with argument order reversed.

source

LightQuery.indexed — Function.

indexed(it, call)

Index it by the results of call, with a default to missing. Relies on Enumerated.

julia> using LightQuery

julia> result = @name indexed(
            [
                (item = "b", index = 2),
                (item = "a", index = 1)
            ],
            :index
        );

julia> result[1]
((item, "a"), (index, 1))

source

LightQuery.when — Function.

when(it, call)

Lazy filter with argument order reversed.

source

LightQuery.order — Function.

order(it, call; keywords...)

Generalized sort. keywords will be passed to sort!; see the documentation there for options. See By for a way to explicitly mark that an object has been sorted. Relies on Enumerated.

julia> using LightQuery

julia> @name order([
            (item = "b", index = 2),
            (item = "a", index = 1)
        ], :index)
2-element view(::Array{Tuple{Tuple{LightQuery.Name{:item},String},Tuple{LightQuery.Name{:index},Int64}},1}, [2, 1]) with eltype Tuple{Tuple{LightQuery.Name{:item},String},Tuple{LightQuery.Name{:index},Int64}}:
 ((item, "a"), (index, 1))
 ((item, "b"), (index, 2))

source

LightQuery.By — Type.

By(it, call)

Mark that it has been pre-sorted by call. For use with Group or Join.

julia> using LightQuery

julia> @name By([
            (item = "a", index = 1),
            (item = "b", index = 2)
        ], :index);

source

LightQuery.Group — Type.

Group(it::By)

Group consecutive keys in it. Requires a presorted object (see By). Relies on Enumerated.

julia> using LightQuery

julia> @name Group(By(
            [
                (item = "a", group = 1),
                (item = "b", group = 1),
                (item = "c", group = 2),
                (item = "d", group = 2)
            ],
            :group
        )) |>
        collect
2-element Array{Pair{Int64,SubArray{Tuple{Tuple{LightQuery.Name{:item},String},Tuple{LightQuery.Name{:group},Int64}},1,Array{Tuple{Tuple{LightQuery.Name{:item},String},Tuple{LightQuery.Name{:group},Int64}},1},Tuple{UnitRange{Int64}},true}},1}:
 1 => [((item, "a"), (group, 1)), ((item, "b"), (group, 1))]
 2 => [((item, "c"), (group, 2)), ((item, "d"), (group, 2))]

source

LightQuery.key — Function.

key(it)

The key in a key => value pair.

source

LightQuery.value — Function.

value(it)

The value in a key => value pair.

source

value(named::Tuple{LightQuery.Name,Any})

named[2]

source

LightQuery.Join — Type.

Join(left::By, right::By)

Find all pairs where isequal(left.call(left.it), right.call(right.it)).

julia> using LightQuery

julia> @name Join(
            By(
                [
                    (left = "a", index = 1),
                    (left = "b", index = 2),
                    (left = "e", index = 5),
                    (left = "f", index = 6)
                ],
                :index
            ),
            By(
                [
                    (right = "a", index = 1),
                    (right = "c", index = 3),
                    (right = "d", index = 4),
                    (right = "e", index = 6)
                ],
                :index
            )
        ) |>
        collect
6-element Array{Pair{Union{Missing, Tuple{Tuple{Name{:left},String},Tuple{Name{:index},Int64}}},Union{Missing, Tuple{Tuple{Name{:right},String},Tuple{Name{:index},Int64}}}},1}:
 ((left, "a"), (index, 1)) => ((right, "a"), (index, 1))
 ((left, "b"), (index, 2)) => missing
                   missing => ((right, "c"), (index, 3))
                   missing => ((right, "d"), (index, 4))
 ((left, "e"), (index, 5)) => missing
 ((left, "f"), (index, 6)) => ((right, "e"), (index, 6))

Assumes left and right are both strictly sorted (no repeats). If there are repeats, Group first. For other join flavors, combine with when. Make sure to annotate with Length if you know it.

source

LightQuery.Length — Type.

Length(it, length)

Allow optimizations based on length. Especially useful after Join and before make_columns.

julia> using LightQuery

julia> @> Filter(iseven, 1:4) |>
        Length(_, 2) |>
        collect
2-element Array{Int64,1}:
 2
 4

source

Pivot

LightQuery.item_names — Function.

item_names(it)

Find names of items in it. Used in Peek and make_columns.

julia> using LightQuery

julia> @name item_names([(a = 1, b = 1.0), (a = 2, b = 2.0)])
(a, b)

source

LightQuery.rows — Function.

rows(it)

Iterator over rows of a NamedTuple of arrays. Always lazy. Inverse of columns. See Peek for a way to view.

julia> using LightQuery

julia> @name collect(rows((a = [1, 2], b = [1.0, 2.0])))
2-element Array{Tuple{Tuple{LightQuery.Name{:a},Int64},Tuple{LightQuery.Name{:b},Float64}},1}:
 ((a, 1), (b, 1.0))
 ((a, 2), (b, 2.0))

source

LightQuery.Peek — Type.

Peek(it; max_rows = 4)

Get a peek of an iterator which returns items with propertynames. Will show no more than max_rows. Relies on item_names.

julia> using LightQuery

julia> @name Peek(rows((a = 1:5, b = 5:-1:1)))
Showing 4 of 5 rows
|   a |   b |
| ---:| ---:|
|   1 |   5 |
|   2 |   4 |
|   3 |   3 |
|   4 |   2 |

source

LightQuery.columns — Function.

columns(it)

Inverse of rows. Always lazy, see make_columns for eager version.

julia> using LightQuery

julia> @name columns(rows((a = [1, 2], b = [1.0, 2.0])))
((a, [1, 2]), (b, [1.0, 2.0]))

source

LightQuery.make_columns — Function.

make_columns(it)

Collect into columns. Always eager, see columns for lazy version. Relies on item_names.

julia> using LightQuery

julia> @name make_columns([(a = 1, b = 1.0), (a = 2, b = 2.0)])
((a, [1, 2]), (b, [1.0, 2.0]))

source