LightQuery.jl

LightQuery.jl

For performance with working with arbitrary structs, explicitly define public propertynames.

LightQuery.By

LightQuery.LeftJoin

LightQuery.Name

LightQuery.Names

LightQuery.columns

LightQuery.gather

LightQuery.group

LightQuery.in_common

LightQuery.key

LightQuery.name

LightQuery.named

LightQuery.order_by

LightQuery.over

LightQuery.remove

LightQuery.rename

LightQuery.rows

LightQuery.select

LightQuery.spread

LightQuery.transform

LightQuery.unname

LightQuery.unzip

LightQuery.value

LightQuery.when

LightQuery.@>

LightQuery.@_

LightQuery.By — Type.

By(it, f)

Marks that it has been pre-sorted by the key f. If f is a symbol, interpret is as a Name. Returned by order_by. For use with group or LeftJoin.

julia> using LightQuery

julia> By([(a = 1,), (a = 2,)], :a).f
Name{:a}()

source

LightQuery.LeftJoin — Type.

LeftJoin(left::By, right::By)

For each value in left, look for a value with the same key in right. Requires both to be presorted (see By).

julia> using LightQuery

julia> LeftJoin(
            By([1, 2, 5, 6], identity),
            By([1, 3, 4, 6], identity)
       ) |> collect
4-element Array{Pair{Int64,Union{Missing, Int64}},1}:
 1 => 1
 2 => missing
 5 => missing
 6 => 6

source

LightQuery.Name — Type.

Name(name)

Container for a name. Can use to select.

julia> using LightQuery

julia> Name(:a)((a = 1, b = 2.0,))
1

source

LightQuery.Names — Type.

struct Names{T} end

Container for names. Can use to select.

julia> using LightQuery

julia> Names(:a, :b)((a = 1, b = 2.0, c = 3//1))
(a = 1, b = 2.0)

source

LightQuery.columns — Method.

columns(it, into_names...)

Collect into columns. Inverse of rows.

julia> using LightQuery

julia> it = [(a = 1, b = 1.0), (a = 2, b = 2.0)];

julia> columns(it, :a, :b)
(a = [1, 2], b = [1.0, 2.0])

source

LightQuery.gather — Method.

gather(data, new_column, columns...)

Gather all the data in columns into a single new_column. Inverse of spread.

julia> using LightQuery

julia> gather((a = 1, b = 2.0, c = "c"), :d, :a, :c)
(b = 2.0, d = (a = 1, c = "c"))

source

LightQuery.group — Method.

group(b::By)

Group consecutive keys in b. Requires a presorted and indexible object (see By).

julia> using LightQuery

julia> group(By([1, 3, 2, 4], iseven)) |> first
false => [1, 3]

source

LightQuery.in_common — Method.

in_common(data1, data2)

Find Names in common.

julia> using LightQuery

julia> in_common((a = 1, b = 2.0), (a = 1, c = "3"))
Names{(:a,)}()

source

LightQuery.key — Method.

key(pair)

The first item

source

LightQuery.name — Method.

name(data, names...)

Wholesale rename data. Inverse of unname.

julia> using LightQuery

julia> name((a = 1, b = 2.0), :c, :d)
(c = 1, d = 2.0)

source

LightQuery.named — Method.

named(data)

Convert to a named tuple.

julia> using LightQuery

julia> named((a = 1, b = 2.0))
(a = 1, b = 2.0)

julia> struct Triple{T1, T2, T3}
            first::T1
            second::T2
            third::T3
        end;

julia> Base.propertynames(t::Triple) = (:first, :second, :third);

julia> named(Triple(1, 1.0, "a"))
(first = 1, second = 1.0, third = "a")

source

LightQuery.order_by — Method.

order_by(it, f)

Generalized sort. If f is a symbol, interpret is as a Name. Will return a By object.

julia> using LightQuery

julia> order_by(["b", "a"], identity).it
2-element view(::Array{String,1}, [2, 1]) with eltype String:
 "a"
 "b"

julia> order_by([(a = 2,), (a = 1,)], :a).it
2-element view(::Array{NamedTuple{(:a,),Tuple{Int64}},1}, [2, 1]) with eltype NamedTuple{(:a,),Tuple{Int64}}:
 (a = 1,)
 (a = 2,)

source

LightQuery.over — Method.

over(it, f)

Hackable version of Generator. If f is a symbol, interpret is as a Name.

julia> using LightQuery

julia> over([1, 2], x -> x + 1) |> collect
2-element Array{Int64,1}:
 2
 3

julia> over([(a = 1,), (a = 2,)], :a) |> collect
2-element Array{Int64,1}:
 1
 2

source

LightQuery.remove — Method.

remove(data, columns...)

Remove columns. Inverse of transform.

julia> using LightQuery

julia> remove((a = 1, b = 2.0), :b)
(a = 1,)

source

LightQuery.rename — Method.

rename(data; renames...)

Rename data. Until we get constant propagation through keyword arguments, use Name for stability.

julia> using LightQuery

julia> rename((a = 1, b = 2.0), c = Name(:a))
(b = 2.0, c = 1)

source

LightQuery.rows — Method.

rows(n::NamedTuple)

Iterator over rows of a NamedTuple of columns. Inverse of columns.

julia> using LightQuery

julia> rows((a = [1, 2], b = [2, 1])) |> first
(a = 1, b = 2)

source

LightQuery.select — Method.

select(data, columns...)

Select columns

julia> using LightQuery

julia> select((a = 1, b = 2.0), :a)
(a = 1,)

source

LightQuery.spread — Method.

spread(data, column::Name)

Unnest nested data in column. Inverse of gather.

julia> using LightQuery

julia> spread((b = 2.0, d = (a = 1, c = "c")), :d)
(b = 2.0, a = 1, c = "c")

source

LightQuery.transform — Method.

transform(data; assignments...)

Apply the functions in assignments to data, assign to the corresponding keys, and merge back in the original.

julia> using LightQuery

julia> transform((a = 1, b = 2.0), c = @_ _.a + _.b)
(a = 1, b = 2.0, c = 3.0)

source

LightQuery.unname — Method.

unname

Remove names. Inverse of name.

julia> using LightQuery

julia> unname((a = 1, b = 2.0))
(1, 2.0)

julia> unname((1, 2.0))
(1, 2.0)

julia> struct Triple{T1, T2, T3}
            first::T1
            second::T2
            third::T3
        end;

julia> Base.propertynames(t::Triple) = (:first, :second, :third);

julia> unname(Triple(1, 1.0, "a"))
(1, 1.0, "a")

source

LightQuery.unzip — Method.

unzip(it, n)

Unzip an iterator it which returns tuples of length n.

julia> using LightQuery

julia> f(x) = (x, x + 1.0);

julia> unzip(over([1], f), 2)
([1], [2.0])

julia> unzip(over([1, missing], f), 2);

julia> unzip(zip([1], [1.0]), 2)
([1], [1.0])

julia> unzip([(1, 1.0)], 2)
([1], [1.0])

source

LightQuery.value — Method.

value(pair)

The second item

source

LightQuery.when — Method.

when(it, f)

Hackable version of Base.Iterators.Filter. If f is a symbol, interpret is as 4 a Name.

julia> using LightQuery

julia> when([1, 2], x -> x > 1) |> collect
1-element Array{Int64,1}:
 2

source

LightQuery.@> — Macro.

macro >(body)

If body is in the form body_ |> tail_, call @_ on tail, and recur on body.

julia> using LightQuery

julia> @> 0 |> _ - 1 |> abs
1

source

LightQuery.@_ — Macro.

macro _(body)

Terser function syntax. The arguments are inside the body; the first argument is _, the second argument is __, etc.

julia> using LightQuery

julia> 1 |> @_(_ + 1)
2

julia> map(@_(__ - _), (1, 2), (2, 1))
(1, -1)

source