Tutorial

Tutorial

I'm using the data from the dplyr tutorial. The data is in the test folder of this package.

I created it with the following R code:

library(nycflights13)
setwd("C:/Users/hp/.julia/dev/LightQuery/test")
write.csv(airports, "airports.csv", na = "", row.names = FALSE)
write.csv(flights, "flights.csv", na = "", row.names = FALSE)
write.csv(weathers, "weather.csv", na = "", row.names = FALSE)

Import tools from Dates, TimeZones, and Unitful.

julia> using LightQuery

julia> using Dates: Date, DateTime, Hour

julia> using TimeZones: Class, Local, TimeZone, VariableTimeZone, ZonedDateTime

julia> using Unitful: °, °F, ft, hr, inch, mbar, mi, minute

Airports cleaning

Use CSV.File to import the airports data.

julia> import CSV

julia> airports_file = CSV.File("airports.csv", missingstrings = ["", "\\N"])
CSV.File("airports.csv"):
Size: 1458 x 8
Tables.Schema:
 :faa    String
 :name   String
 :lat    Float64
 :lon    Float64
 :alt    Int64
 :tz     Int64
 :dst    String
 :tzone  Union{Missing, String}

For this package, I made named_tuples to replace NamedTuples. Use @name to work with them.

Convert the schema to_Columns.

julia> using Tables: schema

julia> const Airport = to_Columns(schema(airports_file))
(Column{`faa`,String,1}(), Column{`name`,String,2}(), Column{`lat`,Float64,3}(), Column{`lon`,Float64,4}(), Column{`alt`,Int64,5}(), Column{`tz`,Int64,6}(), Column{`dst`,String,7}(), Column{`tzone`,Union{Missing, String},8}())

Read the first row.

julia> airport = Airport(first(airports_file))
((`faa`, "04G"), (`name`, "Lansdowne Airport"), (`lat`, 41.1304722), (`lon`, -80.6195833), (`alt`, 1044), (`tz`, -5), (`dst`, "A"), (`tzone`, "America/New_York"))

rename.

julia> airport = @name rename(airport,
            airport_code = :faa,
            altitude = :alt,
            daylight_savings = :dst,
            latitude = :lat,
            longitude = :lon,
            time_zone = :tzone,
            time_zone_offset = :tz
        )
((`name`, "Lansdowne Airport"), (`airport_code`, "04G"), (`altitude`, 1044), (`daylight_savings`, "A"), (`latitude`, 41.1304722), (`longitude`, -80.6195833), (`time_zone`, "America/New_York"), (`time_zone_offset`, -5))

remove redundant data.

julia> airport = @name remove(airport,
            :daylight_savings,
            :time_zone_offset
        )
((`name`, "Lansdowne Airport"), (`airport_code`, "04G"), (`altitude`, 1044), (`latitude`, 41.1304722), (`longitude`, -80.6195833), (`time_zone`, "America/New_York"))

Add units with transform.

julia> airport = @name transform(airport,
            altitude = airport.altitude * ft,
            latitude = airport.latitude * °,
            longitude = airport.longitude * °
        )
((`name`, "Lansdowne Airport"), (`airport_code`, "04G"), (`time_zone`, "America/New_York"), (`altitude`, 1044 ft), (`latitude`, 41.1304722°), (`longitude`, -80.6195833°))

Get a true TimeZone. Use @if_known to handle missing data. Note the data contains some LEGACY timezones. Use a type annotation: TimeZone is unstable without it.

julia> get_time_zone(time_zone) = TimeZone(
            (@if_known time_zone),
            Class(:STANDARD) | Class(:LEGACY)
        )::VariableTimeZone;

julia> @name get_time_zone(airport.time_zone)
America/New_York (UTC-5/UTC-4)

Use the chaining macro @> to chain calls together.

julia> function get_airport(row)
            @name @> Airport(row) |>
            rename(_,
                airport_code = :faa,
                altitude = :alt,
                daylight_savings = :dst,
                latitude = :lat,
                longitude = :lon,
                time_zone = :tzone,
                time_zone_offset = :tz
            ) |>
            remove(_,
                :daylight_savings,
                :time_zone_offset,
            ) |>
            transform(_,
                altitude = _.altitude * ft,
                latitude = _.latitude * °,
                longitude = _.longitude * °,
                time_zone = get_time_zone(_.time_zone)
            )
        end;

julia> get_airport(first(airports_file))
((`name`, "Lansdowne Airport"), (`airport_code`, "04G"), (`altitude`, 1044 ft), (`latitude`, 41.1304722°), (`longitude`, -80.6195833°), (`time_zone`, tz"America/New_York"))

over each row.

julia> airports = over(airports_file, get_airport);

julia> first(airports)
((`name`, "Lansdowne Airport"), (`airport_code`, "04G"), (`altitude`, 1044 ft), (`latitude`, 41.1304722°), (`longitude`, -80.6195833°), (`time_zone`, tz"America/New_York"))

Call make_columns then to_rows to store the data column-wise but view it row-wise. Peek to view.

julia> airports = to_rows(make_columns(airports));

julia> Peek(airports)
Showing 4 of 1458 rows
|                        `name` | `airport_code` | `altitude` |  `latitude` |  `longitude` |                    `time_zone` |
| -----------------------------:| --------------:| ----------:| -----------:| ------------:| ------------------------------:|
|             Lansdowne Airport |            04G |    1044 ft | 41.1304722° | -80.6195833° | America/New_York (UTC-5/UTC-4) |
| Moton Field Municipal Airport |            06A |     264 ft | 32.4605722° | -85.6800278° |  America/Chicago (UTC-6/UTC-5) |
|           Schaumburg Regional |            06C |     801 ft | 41.9893408° | -88.1012428° |  America/Chicago (UTC-6/UTC-5) |
|               Randall Airport |            06N |     523 ft |  41.431912° | -74.3915611° | America/New_York (UTC-5/UTC-4) |

index airports by code.

julia> const indexed_airports = @name index(airports, :airport_code);

julia> indexed_airports["JFK"]
((`name`, "John F Kennedy Intl"), (`airport_code`, "JFK"), (`altitude`, 13 ft), (`latitude`, 40.639751°), (`longitude`, -73.778925°), (`time_zone`, tz"America/New_York"))

Flights cleaning

Use CSV.File to import the flights data.

julia> flights_file = CSV.File("flights.csv")
CSV.File("flights.csv"):
Size: 336776 x 19
Tables.Schema:
 :year            Int64
 :month           Int64
 :day             Int64
 :dep_time        Union{Missing, Int64}
 :sched_dep_time  Int64
 :dep_delay       Union{Missing, Int64}
 :arr_time        Union{Missing, Int64}
 :sched_arr_time  Int64
 :arr_delay       Union{Missing, Int64}
 :carrier         String
 :flight          Int64
 :tailnum         Union{Missing, String}
 :origin          String
 :dest            String
 :air_time        Union{Missing, Int64}
 :distance        Int64
 :hour            Int64
 :minute          Int64
 :time_hour       String

Get the first flight, rename, remove, and transform to add units.

julia> const Flight = to_Columns(schema(flights_file))
(Column{`year`,Int64,1}(), Column{`month`,Int64,2}(), Column{`day`,Int64,3}(), Column{`dep_time`,Union{Missing, Int64},4}(), Column{`sched_dep_time`,Int64,5}(), Column{`dep_delay`,Union{Missing, Int64},6}(), Column{`arr_time`,Union{Missing, Int64},7}(), Column{`sched_arr_time`,Int64,8}(), Column{`arr_delay`,Union{Missing, Int64},9}(), Column{`carrier`,String,10}(), Column{`flight`,Int64,11}(), Column{`tailnum`,Union{Missing, String},12}(), Column{`origin`,String,13}(), Column{`dest`,String,14}(), Column{`air_time`,Union{Missing, Int64},15}(), Column{`distance`,Int64,16}(), Column{`hour`,Int64,17}(), Column{`minute`,Int64,18}(), Column{`time_hour`,String,19}())

julia> flight =
        @name @> flights_file |>
        first |>
        Flight |>
        rename(_,
            arrival_delay = :arr_delay,
            arrival_time = :arr_time,
            departure_delay = :dep_delay,
            departure_time = :dep_time,
            destination = :dest,
            scheduled_arrival_time = :sched_arr_time,
            scheduled_departure_time = :sched_dep_time,
            tail_number = :tailnum
        ) |>
        remove(_,
            :arrival_time,
            :departure_time,
            :hour,
            :minute,
            :time_hour
        ) |>
        transform(_,
            air_time = _.air_time * minute,
            arrival_delay = _.arrival_delay * minute,
            departure_delay = _.departure_delay * minute,
            distance = _.distance * mi
        )
((`year`, 2013), (`month`, 1), (`day`, 1), (`carrier`, "UA"), (`flight`, 1545), (`origin`, "EWR"), (`destination`, "IAH"), (`scheduled_arrival_time`, 819), (`scheduled_departure_time`, 515), (`tail_number`, "N14228"), (`air_time`, 227 minute), (`arrival_delay`, 11 minute), (`departure_delay`, 2 minute), (`distance`, 1400 mi))

Find the time_zone of the airport the flight departed from. Use @if_known to handle missing data.

julia> airport = @if_known @name get(indexed_airports, flight.origin, missing)
((`name`, "Newark Liberty Intl"), (`airport_code`, "EWR"), (`altitude`, 18 ft), (`latitude`, 40.6925°), (`longitude`, -74.168667°), (`time_zone`, tz"America/New_York"))

julia> time_zone = @if_known @name airport.time_zone
America/New_York (UTC-5/UTC-4)

Use divrem(_, 100) to split the scheduled_departure_time.

julia> @name divrem(flight.scheduled_departure_time, 100)
(5, 15)

Create a ZonedDateTime.

julia> @name ZonedDateTime(
            flight.year,
            flight.month,
            flight.day,
            divrem(flight.scheduled_departure_time, 100)...,
            time_zone
        )
2013-01-01T05:15:00-05:00

All together:

julia> get_time(indexed_airports, flight, airport, time) =
            @name ZonedDateTime(
                flight.year,
                flight.month,
                flight.day,
                divrem(time, 100)...,
                @if_known (
                    @if_known get(indexed_airports, airport, missing)
                ).time_zone
            );

julia> @name get_time(
            indexed_airports,
            flight,
            flight.origin,
            flight.scheduled_departure_time
        )
2013-01-01T05:15:00-05:00

We can do the same for the scheduled_arrival_time.

julia> arrival = @name get_time(
            indexed_airports,
            flight,
            flight.destination,
            flight.scheduled_arrival_time
        )
2013-01-01T08:19:00-06:00

All together.

julia> function get_flight(indexed_airports, row)
            @name @> row |>
            Flight |>
            rename(_,
                arrival_delay = :arr_delay,
                arrival_time = :arr_time,
                departure_delay = :dep_delay,
                departure_time = :dep_time,
                destination = :dest,
                scheduled_arrival_time = :sched_arr_time,
                scheduled_departure_time = :sched_dep_time,
                tail_number = :tailnum
            ) |>
            remove(_,
                :arrival_time,
                :departure_time,
                :hour,
                :minute,
                :time_hour
            ) |>
            transform(_,
                air_time = _.air_time * minute,
                distance = _.distance * mi,
                departure_delay = _.departure_delay * minute,
                arrival_delay = _.arrival_delay * minute,
                scheduled_departure_time =
                    get_time(indexed_airports, _, _.origin, _.scheduled_departure_time),
                scheduled_arrival_time =
                    get_time(indexed_airports, _, _.destination, _.scheduled_arrival_time)
            ) |>
            remove(_,
                :year,
                :month,
                :day
            )
        end;

julia> get_flight(indexed_airports, first(flights_file))
((`carrier`, "UA"), (`flight`, 1545), (`origin`, "EWR"), (`destination`, "IAH"), (`tail_number`, "N14228"), (`air_time`, 227 minute), (`distance`, 1400 mi), (`departure_delay`, 2 minute), (`arrival_delay`, 11 minute), (`scheduled_departure_time`, ZonedDateTime(2013, 1, 1, 5, 15, tz"America/New_York")), (`scheduled_arrival_time`, ZonedDateTime(2013, 1, 1, 8, 19, tz"America/Chicago")))

over each row. Use @_ to create an anonymous function.

julia> flights =
        @> flights_file |>
        over(_, @_ get_flight(indexed_airports, _));

julia> flights =
        flights |>
        make_columns |>
        to_rows;

julia> Peek(flights)
Showing 4 of 336776 rows
| `carrier` | `flight` | `origin` | `destination` | `tail_number` | `air_time` | `distance` | `departure_delay` | `arrival_delay` | `scheduled_departure_time` |  `scheduled_arrival_time` |
| ---------:| --------:| --------:| -------------:| -------------:| ----------:| ----------:| -----------------:| ---------------:| --------------------------:| -------------------------:|
|        UA |     1545 |      EWR |           IAH |        N14228 | 227 minute |    1400 mi |          2 minute |       11 minute |  2013-01-01T05:15:00-05:00 | 2013-01-01T08:19:00-06:00 |
|        UA |     1714 |      LGA |           IAH |        N24211 | 227 minute |    1416 mi |          4 minute |       20 minute |  2013-01-01T05:29:00-05:00 | 2013-01-01T08:30:00-06:00 |
|        AA |     1141 |      JFK |           MIA |        N619AA | 160 minute |    1089 mi |          2 minute |       33 minute |  2013-01-01T05:40:00-05:00 | 2013-01-01T08:50:00-05:00 |
|        B6 |      725 |      JFK |           BQN |        N804JB | 183 minute |    1576 mi |         -1 minute |      -18 minute |  2013-01-01T05:45:00-05:00 |                   missing |

Flights validation

Theoretically, the distances between two airports is always the same. Make sure this is the case in our data. First, order by origin, destination, and distance. Then Group By the same variables.

julia> paths_grouped =
        @name @> flights |>
        order(_, (:origin, :destination, :distance)) |>
        Group(By(_, (:origin, :destination, :distance)));

Each Group contains a key and value

julia> path = first(paths_grouped);

julia> key(path)
((`origin`, "EWR"), (`destination`, "ALB"), (`distance`, 143 mi))

julia> value(path) |> Peek
Showing 4 of 439 rows
| `carrier` | `flight` | `origin` | `destination` | `tail_number` | `air_time` | `distance` | `departure_delay` | `arrival_delay` | `scheduled_departure_time` |  `scheduled_arrival_time` |
| ---------:| --------:| --------:| -------------:| -------------:| ----------:| ----------:| -----------------:| ---------------:| --------------------------:| -------------------------:|
|        EV |     4112 |      EWR |           ALB |        N13538 |  33 minute |     143 mi |         -2 minute |      -10 minute |  2013-01-01T13:17:00-05:00 | 2013-01-01T14:23:00-05:00 |
|        EV |     3260 |      EWR |           ALB |        N19554 |  36 minute |     143 mi |         34 minute |       40 minute |  2013-01-01T16:21:00-05:00 | 2013-01-01T17:24:00-05:00 |
|        EV |     4170 |      EWR |           ALB |        N12540 |  31 minute |     143 mi |         52 minute |       44 minute |  2013-01-01T20:04:00-05:00 | 2013-01-01T21:12:00-05:00 |
|        EV |     4316 |      EWR |           ALB |        N14153 |  33 minute |     143 mi |          5 minute |      -14 minute |  2013-01-02T13:27:00-05:00 | 2013-01-02T14:33:00-05:00 |

All we need is the key.

julia> paths =
        @> paths_grouped |>
        over(_, key) |>
        make_columns |>
        to_rows;

julia> Peek(paths)
Showing 4 of 226 rows
| `origin` | `destination` | `distance` |
| --------:| -------------:| ----------:|
|      EWR |           ALB |     143 mi |
|      EWR |           ANC |    3370 mi |
|      EWR |           ATL |     746 mi |
|      EWR |           AUS |    1504 mi |

Find the number of distinct distances.

julia> distinct_distances =
        @name @> paths |>
        Group(By(_, (:origin, :destination))) |>
        over(_, @_ transform(key(_),
            number = length(value(_))
        ));

julia> Peek(distinct_distances)
Showing at most 4 rows
| `origin` | `destination` | `number` |
| --------:| -------------:| --------:|
|      EWR |           ALB |        1 |
|      EWR |           ANC |        1 |
|      EWR |           ATL |        1 |
|      EWR |           AUS |        1 |

See when there are multiple distances for the same path:

julia> @name @> distinct_distances |>
        when(_, @_ _.number != 1) |>
        Peek
Showing at most 4 rows
| `origin` | `destination` | `number` |
| --------:| -------------:| --------:|
|      EWR |           EGE |        2 |
|      JFK |           EGE |        2 |

Peek at flights going to "EGE" using when.

julia> @name @> flights |>
        when(_, @_ _.destination == "EGE") |>
        Peek
Showing at most 4 rows
| `carrier` | `flight` | `origin` | `destination` | `tail_number` | `air_time` | `distance` | `departure_delay` | `arrival_delay` | `scheduled_departure_time` |  `scheduled_arrival_time` |
| ---------:| --------:| --------:| -------------:| -------------:| ----------:| ----------:| -----------------:| ---------------:| --------------------------:| -------------------------:|
|        UA |     1597 |      EWR |           EGE |        N27733 | 287 minute |    1726 mi |         -2 minute |       13 minute |  2013-01-01T09:28:00-05:00 | 2013-01-01T12:20:00-07:00 |
|        AA |      575 |      JFK |           EGE |        N5DRAA | 280 minute |    1747 mi |         -5 minute |        3 minute |  2013-01-01T17:00:00-05:00 | 2013-01-01T19:50:00-07:00 |
|        UA |     1597 |      EWR |           EGE |        N24702 | 261 minute |    1726 mi |          1 minute |        3 minute |  2013-01-02T09:28:00-05:00 | 2013-01-02T12:20:00-07:00 |
|        AA |      575 |      JFK |           EGE |        N631AA | 260 minute |    1747 mi |          5 minute |       16 minute |  2013-01-02T17:00:00-05:00 | 2013-01-02T19:50:00-07:00 |

Weather cleaning

Import weather data. Get the first row, rename, remove, and transform to add units.

julia> weathers_file = CSV.File("weather.csv")
CSV.File("weather.csv"):
Size: 26115 x 15
Tables.Schema:
 :origin      String
 :year        Int64
 :month       Int64
 :day         Int64
 :hour        Int64
 :temp        Union{Missing, Float64}
 :dewp        Union{Missing, Float64}
 :humid       Union{Missing, Float64}
 :wind_dir    Union{Missing, Int64}
 :wind_speed  Union{Missing, Float64}
 :wind_gust   Union{Missing, Float64}
 :precip      Float64
 :pressure    Union{Missing, Float64}
 :visib       Float64
 :time_hour   String

julia> const Weather = to_Columns(schema(weathers_file));

julia> function get_weather(indexed_airports, row)
            @name @> row |>
            Weather |>
            rename(_,
                airport_code = :origin,
                dew_point = :dewp,
                humidity = :humid,
                precipitation = :precip,
                temperature = :temp,
                visibility = :visib,
                wind_direction = :wind_dir
            ) |>
            transform(_,
                dew_point = _.dew_point * °F,
                humidity = _.humidity / 100,
                precipitation = _.precipitation * inch,
                pressure = _.pressure * mbar,
                temperature = _.temperature * °F,
                visibility = _.visibility * mi,
                wind_direction = _.wind_direction * °,
                wind_gust = _.wind_gust * mi / hr,
                wind_speed = _.wind_speed * mi / hr,
                date_time = ZonedDateTime(
                    _.year,
                    _.month,
                    _.day,
                    _.hour,
                    indexed_airports[_.airport_code].time_zone,
                    1
                )
            ) |>
            remove(_,
                :year,
                :month,
                :day,
                :hour
            )
        end;

julia> weathers =
        @> weathers_file |>
        over(_, @_ get_weather(indexed_airports, _));

julia> Peek(weathers)
Showing 4 of 26115 rows
|         `time_hour` | `airport_code` | `dew_point` |         `humidity` | `precipitation` |  `pressure` | `temperature` | `visibility` | `wind_direction` | `wind_gust` |      `wind_speed` |               `date_time` |
| -------------------:| --------------:| -----------:| ------------------:| ---------------:| -----------:| -------------:| ------------:| ----------------:| -----------:| -----------------:| -------------------------:|
| 2013-01-01 01:00:00 |            EWR |    26.06 °F |             0.5937 |        0.0 inch | 1012.0 mbar |      39.02 °F |      10.0 mi |             270° |     missing | 10.35702 mi hr^-1 | 2013-01-01T01:00:00-05:00 |
| 2013-01-01 02:00:00 |            EWR |    26.96 °F | 0.6163000000000001 |        0.0 inch | 1012.3 mbar |      39.02 °F |      10.0 mi |             250° |     missing |  8.05546 mi hr^-1 | 2013-01-01T02:00:00-05:00 |
| 2013-01-01 03:00:00 |            EWR |    28.04 °F | 0.6443000000000001 |        0.0 inch | 1012.5 mbar |      39.02 °F |      10.0 mi |             240° |     missing |  11.5078 mi hr^-1 | 2013-01-01T03:00:00-05:00 |
| 2013-01-01 04:00:00 |            EWR |    28.04 °F |             0.6221 |        0.0 inch | 1012.2 mbar |      39.92 °F |      10.0 mi |             250° |     missing | 12.65858 mi hr^-1 | 2013-01-01T04:00:00-05:00 |

Joining flights and weather

I know that the weather data is already sorted by airport_code and hour.

To InnerJoin it with flights, order and Group flights By matching variables. Only use data when the departure_delay is present.

julia> grouped_flights =
        @name @> flights |>
        when(_, @_ _.departure_delay !== missing) |>
        order(_, (:origin, :scheduled_departure_time)) |>
        Group(By(_, @_ (_.origin, floor(_.scheduled_departure_time, Hour))));

julia> key(first(grouped_flights))
("EWR", ZonedDateTime(2013, 1, 1, 5, tz"America/New_York"))

Now InnerJoin.

julia> weathers_to_flights = @name @> InnerJoin(
            By(weathers, @_ (_.airport_code, _.date_time)),
            By(grouped_flights, key)
        );

Look at the first match.

julia> a_match = first(weathers_to_flights);

julia> weather, (flights_key, flights_value) = a_match;

julia> weather
((`time_hour`, "2013-01-01 05:00:00"), (`airport_code`, "EWR"), (`dew_point`, 28.04 °F), (`humidity`, 0.6443000000000001), (`precipitation`, 0.0 inch), (`pressure`, 1011.9 mbar), (`temperature`, 39.02 °F), (`visibility`, 10.0 mi), (`wind_direction`, 260°), (`wind_gust`, missing), (`wind_speed`, 12.65858 mi hr^-1), (`date_time`, ZonedDateTime(2013, 1, 1, 5, tz"America/New_York")))

julia> flights_key
("EWR", ZonedDateTime(2013, 1, 1, 5, tz"America/New_York"))

julia> Peek(flights_value)
| `carrier` | `flight` | `origin` | `destination` | `tail_number` | `air_time` | `distance` | `departure_delay` | `arrival_delay` | `scheduled_departure_time` |  `scheduled_arrival_time` |
| ---------:| --------:| --------:| -------------:| -------------:| ----------:| ----------:| -----------------:| ---------------:| --------------------------:| -------------------------:|
|        UA |     1545 |      EWR |           IAH |        N14228 | 227 minute |    1400 mi |          2 minute |       11 minute |  2013-01-01T05:15:00-05:00 | 2013-01-01T08:19:00-06:00 |
|        UA |     1696 |      EWR |           ORD |        N39463 | 150 minute |     719 mi |         -4 minute |       12 minute |  2013-01-01T05:58:00-05:00 | 2013-01-01T07:28:00-06:00 |

We're interested in visibility and departure_delay.

julia> visibility = @name weather.visibility
10.0 mi

julia> @name over(flights_value, @_ (
            visibility = visibility,
            departure_delay = _.departure_delay
        )) |>
        Peek
| `visibility` | `departure_delay` |
| ------------:| -----------------:|
|      10.0 mi |          2 minute |
|      10.0 mi |         -4 minute |

All together.

julia> function interested_in(a_match)
            weather, (flights_key, flights_value) = a_match
            visibility = @name weather.visibility
            @name over(flights_value, @_ (
                visibility = visibility,
                departure_delay = _.departure_delay
            ))
        end;

julia> Peek(interested_in(a_match))
| `visibility` | `departure_delay` |
| ------------:| -----------------:|
|      10.0 mi |          2 minute |
|      10.0 mi |         -4 minute |

Use flatten to unnest data (exported from Base).

julia> data =
        @name @> weathers_to_flights |>
        over(_, interested_in) |>
        flatten |>
        make_columns |>
        to_rows;

julia> Peek(data)
Showing 4 of 326993 rows
| `visibility` | `departure_delay` |
| ------------:| -----------------:|
|      10.0 mi |          2 minute |
|      10.0 mi |         -4 minute |
|      10.0 mi |         -5 minute |
|      10.0 mi |         -2 minute |

Visibility vs. departure delay

How does visibility affect departure_delay? Group by visibility.

julia> by_visibility =
        @name @> data |>
        order(_, :visibility) |>
        Group(By(_, :visibility));

julia> visibility_group = first(by_visibility);

julia> key(visibility_group)
0.0 mi

julia> value(visibility_group) |> Peek
Showing 4 of 87 rows
| `visibility` | `departure_delay` |
| ------------:| -----------------:|
|       0.0 mi |         -5 minute |
|       0.0 mi |         -1 minute |
|       0.0 mi |         -8 minute |
|       0.0 mi |         -7 minute |

Calculate the mean departure_delay. Use to_columns to lazily view columns.

julia> using Statistics: mean

julia> @name @> visibility_group |>
        value |>
        to_columns |>
        _.departure_delay |>
        mean
32.252873563218394 minute

For each group.

julia> get_mean_departure_delay(visibility_group) = @name (
            visibility = key(visibility_group),
            mean_departure_delay =
                (@> visibility_group |>
                value |>
                to_columns |>
                _.departure_delay |>
                mean),
            count = length(value(visibility_group))
        );

julia> @> by_visibility |>
            over(_, get_mean_departure_delay) |>
            Peek(_, 20)
Showing at most 20 rows
| `visibility` |    `mean_departure_delay` | `count` |
| ------------:| -------------------------:| -------:|
|       0.0 mi | 32.252873563218394 minute |      87 |
|      0.06 mi |               22.2 minute |      85 |
|      0.12 mi |  50.69975186104218 minute |     403 |
|      0.25 mi | 20.481110254433307 minute |    1297 |
|       0.5 mi |   32.5890826383624 minute |    1319 |
|      0.75 mi |  30.06759906759907 minute |     429 |
|       1.0 mi |  32.24348473566642 minute |    1343 |
|      1.25 mi | 53.187845303867405 minute |     181 |
|       1.5 mi |  25.90661478599222 minute |    1542 |
|      1.75 mi | 43.333333333333336 minute |     132 |
|       2.0 mi | 22.701923076923077 minute |    2912 |
|       2.5 mi |  21.18074398249453 minute |    2285 |
|       3.0 mi |   21.2113218731476 minute |    3374 |
|       4.0 mi |  19.48311444652908 minute |    2132 |
|       5.0 mi |  21.10387902695595 minute |    4563 |
|       6.0 mi | 19.807032301480483 minute |    5944 |
|       7.0 mi | 19.208963745361118 minute |    7006 |
|       8.0 mi |  19.98660103910309 minute |    7314 |
|       9.0 mi | 18.762949476558944 minute |   10985 |
|      10.0 mi | 10.951549367828692 minute |  273660 |

This data suggests that low visibility levels lead to larger departure delays, on average.

Reshaping

For this section, I will use data from the Global Historical Climatology Network. I got this idea from the tidyr tutorial.

julia> file = open("climate.txt");

julia> line = readline(file)
"ACW00011604194901TMAX  289  X  289  X  283  X  283  X  289  X  289  X  278  X  267  X  272  X  278  X  267  X  278  X  267  X  267  X  278  X  267  X  267  X  272  X  272  X  272  X  278  X  272  X  267  X  267  X  267  X  278  X  272  X  272  X  272  X  272  X  272  X"

julia> close(file)

Parse the first chunk.

julia> month_variable = @name (
            year = parse(Int, SubString(line, 12, 15)),
            month = parse(Int, SubString(line, 16, 17)),
            variable = Symbol(SubString(line, 18, 21))
        )
((`year`, 1949), (`month`, 1), (`variable`, :TMAX))

Parse a day. missing is represented by -9999.

julia> function get_day(line, day)
            start = 14 + 8 * day
            value = parse(Int, line[start:start + 4])
            @name (day = day, value =
                if value == -9999
                    missing
                else
                    value
                end
            )
        end;

julia> get_day(line, 1)
((`day`, 1), (`value`, 289))

over each day.

julia> days = @> over(1:31, @_ (
            month_variable...,
            get_day(line, _)...
        ));

julia> first(days)
((`year`, 1949), (`month`, 1), (`variable`, :TMAX), (`day`, 1), (`value`, 289))

Use when to remove missing data;

julia> days = @name when(days, @_ _.value !== missing);

julia> first(days)
((`year`, 1949), (`month`, 1), (`variable`, :TMAX), (`day`, 1), (`value`, 289))

Use transform and remove to create a Date.

julia> get_date(day) =
        @name @> day |>
        transform(_, date = Date(_.year, _.month, _.day)) |>
        remove(_, :year, :month, :day);

julia> get_date(first(days))
((`variable`, :TMAX), (`value`, 289), (`date`, 1949-01-01))

All together.

julia> function get_month_variable(line)
            month_variable = @name (
                year = parse(Int, SubString(line, 12, 15)),
                month = parse(Int, SubString(line, 16, 17)),
                variable = Symbol(SubString(line, 18, 21))
            )
            @name @> over(1:31, @_ (
                month_variable...,
                get_day(line, _)...
            )) |>
            when(_, @_ _.value !== missing) |>
            over(_, get_date)
        end;

julia> first(get_month_variable(line))
((`variable`, :TMAX), (`value`, 289), (`date`, 1949-01-01))

over each line. Use flatten to unnest data.

julia> climate_data =
        @> eachline("climate.txt") |>
        over(_, get_month_variable) |>
        flatten |>
        make_columns |>
        to_rows;

julia> Peek(climate_data)
Showing 4 of 1231 rows
| `variable` | `value` |     `date` |
| ----------:| -------:| ----------:|
|      :TMAX |     289 | 1949-01-01 |
|      :TMAX |     289 | 1949-01-02 |
|      :TMAX |     283 | 1949-01-03 |
|      :TMAX |     283 | 1949-01-04 |

Sort and group by date.

julia> by_date =
        @name @> climate_data |>
        order(_, :date) |>
        Group(By(_, :date));

julia> day_variables = first(by_date);

julia> key(day_variables)
1949-01-01

julia> value(day_variables) |> Peek
Showing 4 of 5 rows
| `variable` | `value` |     `date` |
| ----------:| -------:| ----------:|
|      :TMAX |     289 | 1949-01-01 |
|      :TMIN |     217 | 1949-01-01 |
|      :PRCP |       0 | 1949-01-01 |
|      :SNOW |       0 | 1949-01-01 |

over each variable. Use Name to make a Name. This is unavoidably type-unstable.

julia> spread_variables(day_variables) = @name (
            date = key(day_variables),
            over(
                value(day_variables),
                @_ (Name(_.variable), _.value)
            )...
        );

julia> spread_variables(day_variables)
((`date`, 1949-01-01), (`TMAX`, 289), (`TMIN`, 217), (`PRCP`, 0), (`SNOW`, 0), (`SNWD`, 0))

over each day.

julia> @> by_date |>
        over(_, spread_variables) |>
        Peek
Showing at most 4 rows
|  `WT16` |     `date` | `TMAX` | `TMIN` | `PRCP` | `SNOW` | `SNWD` |
| -------:| ----------:| ------:| ------:| ------:| ------:| ------:|
| missing | 1949-01-01 |    289 |    217 |      0 |      0 |      0 |
|       1 | 1949-01-02 |    289 |    228 |     30 |      0 |      0 |
| missing | 1949-01-03 |    283 |    222 |      0 |      0 |      0 |
|       1 | 1949-01-04 |    283 |    233 |      0 |      0 |      0 |

Interface

Macros

LightQuery.@_ — Macro.

macro _(body)

Terser function syntax. The arguments are inside the body; the first argument is _, the second argument is __, etc.

julia> using LightQuery

julia> using Test: @inferred

julia> @inferred (@_ _ + 1)(1)
2

julia> @inferred map((@_ __ - _), (1, 2), (2, 1))
(1, -1)

source

LightQuery.@> — Macro.

macro >(body)

If body is in the form object_ |> call_, call @_ on call, and recur on object.

julia> using LightQuery

julia> @> 0 |> _ - 1 |> abs
1

You can nest chains:

julia> @> 1 |> (@> _ + 1 |> _ + 1)
3

source

LightQuery.@if_known — Macro.

if_known(something)

If something is missing, return missing, otherwise, something.

julia> using LightQuery

julia> using Test: @inferred

julia> function test(x)
            first(@if_known(x))
        end;

julia> @inferred test((1, 2))
1

julia> @inferred test(missing)
missing

source

LightQuery.@name — Macro.

macro name(code)

Switch to named_tuples.

julia> using LightQuery

julia> using Test: @inferred

julia> data = @name (a = 1, b = 1.0, c = 1, d = 1.0, e = 1, f = 1.0)
((`a`, 1), (`b`, 1.0), (`c`, 1), (`d`, 1.0), (`e`, 1), (`f`, 1.0))

based on typed Names.

julia> @name :a
`a`

Names can be used as properties

julia> @name @inferred data.c
1

julia> @name data.g
ERROR: BoundsError: attempt to access ((`a`, 1), (`b`, 1.0), (`c`, 1), (`d`, 1.0), (`e`, 1), (`f`, 1.0))
[...]

and selector functions.

julia> @name @inferred (:c)(data)
1

Multiple names can be used as selector functions

julia> @name @inferred (:c, :f)(data)
((`c`, 1), (`f`, 1.0))

You can also convert back to NamedTuples.

julia> @inferred NamedTuple(data)
(a = 1, b = 1.0, c = 1, d = 1.0, e = 1, f = 1.0)

source

Values

LightQuery.if_else — Function.

if_else(switch, yes, no)

ifelse that you can add methods to.

julia> using LightQuery

julia> if_else(true, 1, 0)
1

julia> if_else(false, 1, 0)
0

source

LightQuery.key — Function.

key(pair)

The key in a key => value pair.

julia> using LightQuery

julia> using Test: @inferred

julia> @inferred key(:a => 1)
:a

julia> @inferred key((:a, 1))
:a

source

LightQuery.value — Function.

value(pair)

The value in a key => value pair.

julia> using LightQuery

julia> using Test: @inferred

julia> @inferred value(:a => 1)
1

julia> @inferred value((:a, 1))
1

source

LightQuery.Name — Type.

Name(name)

Create a typed Name. Inverse of unname

julia> using LightQuery

julia> Name(:a)
`a`

source

LightQuery.unname — Function.

unname(::Name{name}) where name

Inverse of Name.

julia> using LightQuery

julia> using Test: @inferred

julia> @inferred unname(Name(:a))
:a

source

LightQuery.Column — Type.

struct Column{name, type, position}

Contains the name, type, and position of each column. Useful for type stable iteration of rows. See example in to_Columns.

source

LightQuery.Descending — Type.

struct Descending{Increasing}

Reverse sorting order.

julia> using LightQuery

julia> using Test: @inferred

julia> @inferred order([1, -2], x -> Descending(abs(x)))
2-element view(::Array{Int64,1}, [2, 1]) with eltype Int64:
 -2
  1

source

Columns

LightQuery.named_tuple — Function.

named_tuple(data)

Convert data to a named_tuple (see @name).

julia> using LightQuery

julia> using Test: @inferred

julia> @inferred named_tuple((a = 1, b = 1.0, c = 1, d = 1.0, e = 1, f = 1.0))
((`a`, 1), (`b`, 1.0), (`c`, 1), (`d`, 1.0), (`e`, 1), (`f`, 1.0))

For stability working with arbitrary structs, propertynames must constant propagate.

julia> struct MyType
            a::Int
            b::Float64
            c::Int
            d::Float64
            e::Int
            f::Float64
        end

julia> import Base: propertynames

julia> @inline propertynames(::MyType) = (:a, :b, :c, :d, :e, :f);

julia> @inferred named_tuple(MyType(1, 1.0, 1, 1.0, 1, 1.0))
((`a`, 1), (`b`, 1.0), (`c`, 1), (`d`, 1.0), (`e`, 1), (`f`, 1.0))

source

LightQuery.rename — Function.

rename(data, new_name_old_names...)

Rename data.

julia> using LightQuery

julia> using Test: @inferred

julia> data = @name (a = 1, b = 1.0, c = 1, d = 1.0, e = 1, f = 1.0)
((`a`, 1), (`b`, 1.0), (`c`, 1), (`d`, 1.0), (`e`, 1), (`f`, 1.0))

julia> @name @inferred rename(data, c2 = :c, f2 = :f)
((`a`, 1), (`b`, 1.0), (`d`, 1.0), (`e`, 1), (`c2`, 1), (`f2`, 1.0))

source

LightQuery.transform — Function.

transform(data, assignments...)

Merge assignments into data, overwriting old values.

julia> using LightQuery

julia> using Test: @inferred

julia> data = @name (a = 1, b = 1.0, c = 1, d = 1.0, e = 1, f = 1.0)
((`a`, 1), (`b`, 1.0), (`c`, 1), (`d`, 1.0), (`e`, 1), (`f`, 1.0))

julia> @name @inferred transform(data, c = 2.0, f = 2, g = 1, h = 1.0)
((`a`, 1), (`b`, 1.0), (`d`, 1.0), (`e`, 1), (`c`, 2.0), (`f`, 2), (`g`, 1), (`h`, 1.0))

source

LightQuery.remove — Function.

remove(data, old_names...)

Remove old_names from data.

julia> using LightQuery

julia> using Test: @inferred

julia> data = @name (a = 1, b = 1.0, c = 1, d = 1.0, e = 1, f = 1.0)
((`a`, 1), (`b`, 1.0), (`c`, 1), (`d`, 1.0), (`e`, 1), (`f`, 1.0))

julia> @name @inferred remove(data, :c, :f)
((`a`, 1), (`b`, 1.0), (`d`, 1.0), (`e`, 1))

source

LightQuery.gather — Function.

gather(data, new_name_old_names...)

For each new_name, old_names pair in new_name_old_names, gather the old_names into a single new_name. Inverse of spread.

julia> using LightQuery

julia> using Test: @inferred

julia> data = @name (a = 1, b = 1.0, c = 1, d = 1.0, e = 1, f = 1.0)
((`a`, 1), (`b`, 1.0), (`c`, 1), (`d`, 1.0), (`e`, 1), (`f`, 1.0))

julia> @name @inferred gather(data, g = (:b, :e), h = (:c, :f))
((`a`, 1), (`d`, 1.0), (`g`, ((`b`, 1.0), (`e`, 1))), (`h`, ((`c`, 1), (`f`, 1.0))))

source

LightQuery.spread — Function.

spread(data, some_names...)

Unnest nested named tuples. Inverse of gather.

julia> using LightQuery

julia> using Test: @inferred

julia> gathered = @name (a = 1, d = 1.0, g = (b = 1.0, e = 1), h = (c = 1, f = 1.0))
((`a`, 1), (`d`, 1.0), (`g`, ((`b`, 1.0), (`e`, 1))), (`h`, ((`c`, 1), (`f`, 1.0))))

julia> @name @inferred spread(gathered, :g, :h)
((`a`, 1), (`d`, 1.0), (`b`, 1.0), (`e`, 1), (`c`, 1), (`f`, 1.0))

source

LightQuery.Apply — Type.

struct Apply{Names}

Apply Names to unnamed values.

julia> using LightQuery

julia> using Test: @inferred

julia> @name @inferred Apply((:a, :b, :c, :d, :e, :f))((1, 1.0, 1, 1.0, 1, 1.0))
((`a`, 1), (`b`, 1.0), (`c`, 1), (`d`, 1.0), (`e`, 1), (`f`, 1.0))

source

LightQuery.to_Columns — Function.

to_Columns(::Tables.Schema)

Get the Columns in a schema. Columns can be used as a type stable selector function.

julia> using LightQuery

julia> using Test: @inferred

julia> using CSV: File

julia> using Tables: schema

julia> test = File("test.csv")
CSV.File("test.csv"):
Size: 1 x 6
Tables.Schema:
 :a  Int64
 :b  Float64
 :c  Int64
 :d  Float64
 :e  Int64
 :f  Float64

julia> template = @inferred to_Columns(schema(test))
(Column{`a`,Int64,1}(), Column{`b`,Float64,2}(), Column{`c`,Int64,3}(), Column{`d`,Float64,4}(), Column{`e`,Int64,5}(), Column{`f`,Float64,6}())

julia> @inferred template(first(test))
((`a`, 1), (`b`, 1.0), (`c`, 1), (`d`, 1.0), (`e`, 1), (`f`, 1.0))

source

Rows

LightQuery.Enumerate — Type.

Enumerate{Unenumerated}

Relies on the fact that iteration states can be converted to indices; thus, you might have to define LightQuery.state_to_index for unrecognized types. "Sees through" some iterators like when.

julia> using LightQuery

julia> using Test: @inferred

julia> @inferred collect(Enumerate(when([4, 3, 2, 1], iseven)))
2-element Array{Tuple{Int64,Int64},1}:
 (1, 4)
 (3, 2)

source

LightQuery.over — Function.

over(iterator, call)

Lazy map with the reverse argument order.

julia> using LightQuery

julia> using Test: @inferred

julia> @inferred collect(over([1, -2, -3, 4], abs))
4-element Array{Int64,1}:
 1
 2
 3
 4

source

LightQuery.index — Function.

index(unindexed, key_function)

Index unindexed by the results of key_function. Relies on Enumerate. Results of key_function must be unique.

julia> using LightQuery

julia> using Test: @inferred

julia> @inferred index([-2, 1], abs)
LightQuery.Indexed{Int64,Int64,Array{Int64,1},Dict{Int64,Int64}} with 2 entries:
  2 => -2
  1 => 1

source

LightQuery.when — Function.

when(iterator, call)

Lazy filter with the reverse argument order.

julia> using LightQuery

julia> using Test: @inferred

julia> @inferred collect(when(1:4, iseven))
2-element Array{Int64,1}:
 2
 4

source

LightQuery.order — Function.

order(unordered, key_function; keywords...)

Generalized sort. keywords will be passed to sort!; see the documentation there for options. Use By to mark that an object has been sorted. Relies on Enumerate. If the results of key_function are type unstable, consider using hash ∘ key_function instead.

julia> using LightQuery

julia> using Test: @inferred

julia> @inferred order([-2, 1], abs)
2-element view(::Array{Int64,1}, [2, 1]) with eltype Int64:
  1
 -2

source

LightQuery.By — Type.

By(sorted, key_function)

Mark that sorted has been pre-sorted by key_function. Use with Group, InnerJoin, LeftJoin, RightJoin, and OuterJoin.

julia> using LightQuery

julia> using Test: @inferred

julia> @inferred By([1, -2], abs)
By{Array{Int64,1},typeof(abs)}([1, -2], abs)

source

LightQuery.Group — Type.

Group(ungrouped::By)

Group consecutive keys in ungrouped. Requires a presorted object (see By). Relies on Enumerate.

julia> using LightQuery

julia> using Test: @inferred

julia> @inferred collect(Group(By([1, -1, -2, 2, 3, -3], abs)))
3-element Array{Tuple{Int64,SubArray{Int64,1,Array{Int64,1},Tuple{UnitRange{Int64}},true}},1}:
 (1, [1, -1])
 (2, [-2, 2])
 (3, [3, -3])

julia> @inferred collect(Group(By(Int[], abs)))
0-element Array{Tuple{Int64,SubArray{Int64,1,Array{Int64,1},Tuple{UnitRange{Int64}},true}},1}

source

LightQuery.InnerJoin — Type.

InnerJoin(left::By, right::By) <: Join{Left, Right}

Find all pairs where isequal(left.key_function(left.sorted), right.key_function(right.sorted)).

julia> using LightQuery

julia> using Test: @inferred

julia> @inferred collect(InnerJoin(By([1, -2, 5, -6], abs), By([-1, 3, -4, 6], abs)))
2-element Array{Tuple{Int64,Int64},1}:
 (1, -1)
 (-6, 6)

julia> @inferred collect(InnerJoin(By(Int[], abs), By(Int[], abs)))
0-element Array{Tuple{Int64,Int64},1}

julia> @inferred collect(InnerJoin(By([1], abs), By(Int[], abs)))
0-element Array{Tuple{Int64,Int64},1}

julia> @inferred collect(InnerJoin(By(Int[], abs), By([1], abs)))
0-element Array{Tuple{Int64,Int64},1}

Assumes left and right are both strictly sorted (no repeats). If there are repeats, Group first.

source

LightQuery.LeftJoin — Type.

LeftJoin(left::By, right::By) <: Join{Left, Right}

Find all pairs where isequal(left.key_function(left.sorted), right.key_function(right.sorted)), using missing when there is no right match.

julia> using LightQuery

julia> collect(LeftJoin(By([1, -2, 5, -6], abs), By([-1, 3, -4, 6], abs)))
4-element Array{Tuple{Int64,Union{Missing, Int64}},1}:
 (1, -1)
 (-2, missing)
 (5, missing)
 (-6, 6)

julia> collect(LeftJoin(By(Int[], abs), By(Int[], abs)))
0-element Array{Tuple{Int64,Union{Missing, Int64}},1}

julia> collect(LeftJoin(By([1], abs), By(Int[], abs)))
1-element Array{Tuple{Int64,Union{Missing, Int64}},1}:
 (1, missing)

julia> collect(LeftJoin(By(Int[], abs), By([1], abs)))
0-element Array{Tuple{Int64,Union{Missing, Int64}},1}

Assumes left and right are both strictly sorted (no repeats). If there are repeats, Group first.

source

LightQuery.RightJoin — Type.

RightJoin(left::By, right::By) <: Join{Left, Right}

Find all pairs where isequal(left.key_function(left.sorted), right.key_function(right.sorted)), using missing when there is no left match.

julia> using LightQuery

julia> collect(RightJoin(By([1, -2, 5, -6], abs), By([-1, 3, -4, 6], abs)))
4-element Array{Tuple{Union{Missing, Int64},Int64},1}:
 (1, -1)
 (missing, 3)
 (missing, -4)
 (-6, 6)

julia> collect(RightJoin(By(Int[], abs), By(Int[], abs)))
0-element Array{Tuple{Union{Missing, Int64},Int64},1}

julia> collect(RightJoin(By([1], abs), By(Int[], abs)))
0-element Array{Tuple{Union{Missing, Int64},Int64},1}

julia> collect(RightJoin(By(Int[], abs), By([1], abs)))
1-element Array{Tuple{Union{Missing, Int64},Int64},1}:
 (missing, 1)

Assumes left and right are both strictly sorted (no repeats). If there are repeats, Group first.

source

LightQuery.OuterJoin — Type.

OuterJoin(left::By, right::By) <: Join{Left, Right}

Find all pairs where isequal(left.key_function(left.sorted), right.key_function(right.sorted)), using missing when there is no left or right match.

julia> using LightQuery

julia> collect(OuterJoin(By([1, -2, 5, -6], abs), By([-1, 3, -4, 6], abs)))
6-element Array{Tuple{Union{Missing, Int64},Union{Missing, Int64}},1}:
 (1, -1)
 (-2, missing)
 (missing, 3)
 (missing, -4)
 (5, missing)
 (-6, 6)

julia> collect(OuterJoin(By(Int[], abs), By(Int[], abs)))
0-element Array{Tuple{Union{Missing, Int64},Union{Missing, Int64}},1}

julia> collect(OuterJoin(By([1], abs), By(Int[], abs)))
1-element Array{Tuple{Union{Missing, Int64},Union{Missing, Int64}},1}:
 (1, missing)

julia> collect(OuterJoin(By(Int[], abs), By([1], abs)))
1-element Array{Tuple{Union{Missing, Int64},Union{Missing, Int64}},1}:
 (missing, 1)

Assumes left and right are both strictly sorted (no repeats). If there are repeats, Group first.

source

LightQuery.distinct — Function.

distinct(repeated, key_function = identity)

Generalized version of unique.

julia> using LightQuery

julia> using Test: @inferred

julia> @inferred distinct([1, 2, missing, missing, 2, 1])
3-element view(::Array{Union{Missing, Int64},1}, [1, 2, 3]) with eltype Union{Missing, Int64}:
 1
 2
  missing

source

Tables

LightQuery.to_rows — Function.

to_rows(columns)

Iterator over rows of a table. Always lazy. Inverse of to_columns. Use Peek to view.

julia> using LightQuery

julia> using Test: @inferred

julia> @name @inferred collect(to_rows((a = [1, 2], b = [1.0, 2.0])))
2-element Array{Tuple{Tuple{Name{:a},Int64},Tuple{Name{:b},Float64}},1}:
 ((`a`, 1), (`b`, 1.0))
 ((`a`, 2), (`b`, 2.0))

source

LightQuery.Peek — Type.

Peek(rows, maximum_length = 4)

Peek an iterator which returns named tuples. Will show no more than maximum_length rows.

julia> using LightQuery

julia> using Test: @inferred

julia> @name @inferred Peek(to_rows((a = 1:5, b = 5:-1:1)))
Showing 4 of 5 rows
| `a` | `b` |
| ---:| ---:|
|   1 |   5 |
|   2 |   4 |
|   3 |   3 |
|   4 |   2 |

source

LightQuery.to_columns — Function.

to_columns(rows)

Inverse of to_rows. Always lazy, see make_columns for an eager version.

julia> using LightQuery

julia> using Test: @inferred

julia> @name @inferred to_columns(to_rows((a = [1, 2], b = [1.0, 2.0])))
((`a`, [1, 2]), (`b`, [1.0, 2.0]))

source

LightQuery.make_columns — Function.

make_columns(rows)

Collect into columns. Always eager, see to_columns for a lazy version.

julia> using LightQuery

julia> using Test: @inferred

julia> stable(x) = @name (a = x, b = x + 0.0, c = x, d = x + 0.0, e = x, f = x + 0.0);

julia> @inferred make_columns(over(1:4, stable))
((`a`, [1, 2, 3, 4]), (`b`, [1.0, 2.0, 3.0, 4.0]), (`c`, [1, 2, 3, 4]), (`d`, [1.0, 2.0, 3.0, 4.0]), (`e`, [1, 2, 3, 4]), (`f`, [1.0, 2.0, 3.0, 4.0]))

julia> unstable(x) =
            @name if x <= 2
                (a = missing, b = string(x), d = string(x))
            else
                (a = x, b = missing, c = x)
            end;

julia> make_columns(over(1:4, unstable))
((`d`, Union{Missing, String}["1", "2", missing, missing]), (`a`, Union{Missing, Int64}[missing, missing, 3, 4]), (`b`, Union{Missing, String}["1", "2", missing, missing]), (`c`, Union{Missing, Int64}[missing, missing, 3, 4]))

julia> make_columns(when(over(1:4, unstable), row -> true))
((`d`, Union{Missing, String}["1", "2", missing, missing]), (`a`, Union{Missing, Int64}[missing, missing, 3, 4]), (`b`, Union{Missing, String}["1", "2", missing, missing]), (`c`, Union{Missing, Int64}[missing, missing, 3, 4]))

source